### 2024-11-13 #7478 @editorialbot
[REVIEW]: MQT Core: The Backbone of the Munich Quantum Toolkit (MQT)
**Submitting author:** @burgholzer (<a href="http://orcid.org/0000-0003-4699-1316">Lukas Burgholzer</a>)
**Repository:** https://github.com/munich-quantum-toolkit/core
**Branch with paper.md** (empty if default branch): main
**Version:** v3.0.0
**Archive:** 10.6084/m9.figshare.28740254.v1
