### 2024-11-21 747f4e0
🏗️ WIP structure
