### 2024-11-21 a6f99ae
🎨 Tackle clang-tidy warnings
