+++ include/qir/qir.h
@@ -5,9 +5,16 @@
 
 // Instructions to wrap a C++ class with a C interface are taken from
 // [https://stackoverflow.com/a/11971205](https://stackoverflow.com/a/11971205)
-#include <stdbool.h>
 
 #pragma once
+
+// NOLINTBEGIN(bugprone-reserved-identifier)
+// NOLINTBEGIN(modernize-use-using)
+// NOLINTBEGIN(modernize-deprecated-headers)
+// NOLINTBEGIN(readability-identifier-naming)
+
+#include <stdint.h>
+#include <stdnoreturn.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -16,14 +23,14 @@ extern "C" {
 // *** SIMPLE TYPES ***
 // cf.
 // https://github.com/qir-alliance/qir-spec/blob/main/specification/v0.1/1_Data_Types.md#simple-types
-typedef signed long long int Int;
+typedef int64_t Int;
 typedef double Double;
 typedef bool Bool;
-typedef enum {
-  PauliId_I = 0,
-  PauliId_X = 1,
-  PauliId_Z = 2,
-  PauliId_Y = 3,
+typedef enum : uint8_t {
+  PauliI = 0,
+  PauliX = 1,
+  PauliZ = 2,
+  PauliY = 3,
 } Pauli;
 typedef struct {
   Int start;
@@ -53,7 +60,7 @@ Bool __quantum__rt__result_equal(Result*, Result*);
 
 // Adds the given integer value to the reference count for the result.
 // Deallocates the result if the reference count becomes 0.
-void __quantum__rt__result_update_reference_count(Result*, long int);
+void __quantum__rt__result_update_reference_count(Result*, int32_t);
 
 // *** QUBITS ***
 // cf.
@@ -73,11 +80,11 @@ String* __quantum__rt__string_create(const char*);
 const char* __quantum__rt__string_get_data(String*);
 
 // Returns the length of the byte array that contains the string data.
-long int __quantum__rt__string_get_length(String*);
+int32_t __quantum__rt__string_get_length(String*);
 
 // Adds the given integer value to the reference count for the string.
 // Deallocates the string if the reference count becomes 0.
-void __quantum__rt__string_update_reference_count(String*, long int);
+void __quantum__rt__string_update_reference_count(String*, int32_t);
 
 // Creates a new string that is the concatenation of the two argument strings.
 String* __quantum__rt__string_concatenate(String*, String*);
@@ -119,18 +126,18 @@ BigInt* __quantum__rt__bigint_create_i64(Int);
 // Creates a big integer with the initial value specified by the i8 array. The
 // 0-th element of the array is the highest-order byte, followed by the first
 // element, etc.
-BigInt* __quantum__rt__bigint_create_array(long int, char*);
+BigInt* __quantum__rt__bigint_create_array(int32_t, char*);
 
 // Returns a pointer to the i8 array containing the value of the big integer.
 char* __quantum__rt__bigint_get_data(BigInt*);
 
 // Returns the length of the i8 array that represents the big integer value.
-long int __quantum__rt__bigint_get_length(BigInt*);
+int32_t __quantum__rt__bigint_get_length(BigInt*);
 
 // Adds the given integer value to the reference count for the big integer.
 // Deallocates the big integer if the reference count becomes 0. The behavior is
 // undefined if the reference count becomes negative.
-void __quantum__rt__bigint_update_reference_count(BigInt*, long int);
+void __quantum__rt__bigint_update_reference_count(BigInt*, int32_t);
 
 // Returns the negative of the big integer.
[…]
