### 2024-11-21 ac6561c
🎨 Make Backend singleton and initilaize
