### 2024-11-21 96d162f
✨ Finish apply function
