+++ include/qir/qir_dd_backend.hpp
@@ -15,9 +15,12 @@ private:
 
   AddressMode addressMode;
   std::unordered_map<qc::Qubit, qc::Qubit> qRegister;
+  dd::Package<> dd;
   dd::vEdge qState{};
 
   QIR_DD_Backend();
+
+  auto determineAddressMode() -> void;
 
 public:
   static QIR_DD_Backend& getInstance() {
@@ -27,5 +30,7 @@ public:
 
   QIR_DD_Backend(const QIR_DD_Backend&) = delete;
   QIR_DD_Backend& operator=(const QIR_DD_Backend&) = delete;
+
+  template <typename... Args> auto apply(qc::OpType op, Args... qubits) -> void;
 };
 } // namespace mqt
