### 2024-11-22 54c594a
🎨 Enlarge qState if necessary
