+++ src/qir/qir_dd_backend.cpp
@@ -5,13 +5,15 @@
 
 #include <cstdio>
 #include <cstdlib>
+#include <exception>
 #include <stdexcept>
 
 namespace mqt {
 
-QIR_DD_Backend::QIR_DD_Backend() : addressMode(AddressMode::UNKNOWN) {
+QIR_DD_Backend::QIR_DD_Backend()
+    : addressMode(AddressMode::UNKNOWN), numQubits(0),
+      qState(dd::vEdge::one()) {
   qRegister = std::unordered_map<qc::Qubit, qc::Qubit>();
-  qState = dd::vEdge::one();
 }
 
 auto QIR_DD_Backend::determineAddressMode() -> void {
@@ -26,13 +28,22 @@ auto QIR_DD_Backend::determineAddressMode() -> void {
 
 template <typename... Args>
 auto QIR_DD_Backend::apply(const qc::OpType op, Args... qubits) -> void {
+  // get hardware addresses if necessary
   determineAddressMode();
   std::vector<qc::Qubit> addresses = {qubits...};
   if (addressMode == AddressMode::DYNAMIC) {
     std::transform(addresses.cbegin(), addresses.cend(), addresses.begin(),
-                   [&](const auto q) { return qRegister.at(q); });
+                   [&](const auto q) {
+                     try {
+                       return qRegister.at(q);
+                     } catch (const std::out_of_range&) {
+                       std::throw_with_nested(
+                           std::invalid_argument("Qubit address not found: q"));
+                     }
+                   });
   }
-  int t = 0;
+  // split addresses into control and target
+  uint8_t t = 0;
   if (isSingleQubitGate(op)) {
     t = 1;
   } else if (isTwoQubitGate(op)) {
@@ -42,8 +53,17 @@ auto QIR_DD_Backend::apply(const qc::OpType op, Args... qubits) -> void {
   }
   const auto& controls = qc::Controls(addresses.cbegin(), addresses.cend() - t);
   const auto& targets = qc::Targets(addresses.cbegin() - t, addresses.cend());
+  // retrieve operation matrix
   const auto& operation = qc::StandardOperation(controls, targets, op);
   const auto& matrix = getDD(&operation, dd);
+  // enlarge quantum state if necessary
+  const auto maxTarget = *std::max_element(targets.cbegin(), targets.cend());
+  const auto d = maxTarget - numQubits + 1;
+  if (d > 0) {
+    qState = dd.kronecker(qState, dd.makeZeroState(d), d);
+    numQubits += d;
+  }
+  // apply operation
   const auto& tmp = dd.multiply(matrix, qState);
   dd.incRef(tmp);
   dd.decRef(qState);
@@ -71,7 +91,7 @@ Bool __quantum__rt__result_equal(Result*, Result*) {
   return false;
 }
 
-void __quantum__rt__result_update_reference_count(Result*, long int) {
+void __quantum__rt__result_update_reference_count(Result*, int32_t) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
 }
 
@@ -86,12 +106,12 @@ const char* __quantum__rt__string_get_data(String*) {
   return NULL;
 }
 
-long int __quantum__rt__string_get_length(String*) {
+int32_t __quantum__rt__string_get_length(String*) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return 0;
 }
 
[…]
