### 2024-11-22 958aac7
🎨 Make pointer opaque and define them
