### 2024-11-22 e8efdd6
⏪️ Make opaque pointers to void pointer again since only pointer matters
