### 2024-11-22 778e99c
🎨 Reintroduce opaque pointers, fix alloc function
