### 2024-11-22 fec7e01
✨ Implement qFree
