### 2024-11-22 fde2449
🎨 Implement Result QIR functions
