### 2024-11-22 dfe2225
🎨 Allocate one qubit at a time
