### 2024-11-23 46c520f
⬆️🪝 Update pre-commit hook pre-commit/mirrors-clang-format to v19.1.4
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
