+++ include/qir/qir.h
@@ -23,9 +23,6 @@ extern "C" {
 // *** SIMPLE TYPES ***
 // cf.
 // https://github.com/qir-alliance/qir-spec/blob/main/specification/v0.1/1_Data_Types.md#simple-types
-typedef int64_t Int;
-typedef double Double;
-typedef bool Bool;
 typedef enum : uint8_t {
   PauliI = 0,
   PauliX = 1,
@@ -33,9 +30,9 @@ typedef enum : uint8_t {
   PauliY = 3,
 } Pauli;
 typedef struct {
-  Int start;
-  Int step;
-  Int end;
+  int64_t start;
+  int64_t step;
+  int64_t end;
 } Range;
 
 // moved up, because it is required for Strings already, see BigInt section for
@@ -56,7 +53,7 @@ Result* __quantum__rt__result_get_one();
 
 // Returns true if the two results are the same, and false if they are
 // different.
-Bool __quantum__rt__result_equal(const Result*, const Result*);
+bool __quantum__rt__result_equal(const Result*, const Result*);
 
 // Adds the given integer value to the reference count for the result.
 // Deallocates the result if the reference count becomes 0.
@@ -90,16 +87,16 @@ void __quantum__rt__string_update_reference_count(String*, int32_t);
 String* __quantum__rt__string_concatenate(String*, String*);
 
 // Returns true if the two strings are equal, false otherwise.
-Bool __quantum__rt__string_equal(String*, String*);
+bool __quantum__rt__string_equal(String*, String*);
 
 // Returns a string representation of the integer.
-String* __quantum__rt__int_to_string(Int);
+String* __quantum__rt__int_to_string(int64_t);
 
 // Returns a string representation of the double.
-String* __quantum__rt__double_to_string(Double);
+String* __quantum__rt__double_to_string(double);
 
 // Returns a string representation of the Boolean.
-String* __quantum__rt__bool_to_string(Bool);
+String* __quantum__rt__bool_to_string(bool);
 
 // Returns a string representation of the result.
 String* __quantum__rt__result_to_string(Result*);
@@ -121,7 +118,7 @@ String* __quantum__rt__bigint_to_string(BigInt*);
 // https://github.com/qir-alliance/qir-spec/blob/main/specification/v0.1/1_Data_Types.md#big-integers
 
 // Creates a big integer with the specified initial value.
-BigInt* __quantum__rt__bigint_create_i64(Int);
+BigInt* __quantum__rt__bigint_create_i64(int64_t);
 
 // Creates a big integer with the initial value specified by the i8 array. The
 // 0-th element of the array is the highest-order byte, followed by the first
@@ -174,22 +171,22 @@ BigInt* __quantum__rt__bigint_bitnot(BigInt*);
 
 // Returns the big integer arithmetically shifted left by the (positive) integer
 // amount of bits.
-BigInt* __quantum__rt__bigint_shiftleft(BigInt*, Int);
+BigInt* __quantum__rt__bigint_shiftleft(BigInt*, int64_t);
 
 // Returns the big integer arithmetically shifted right by the (positive)
 // integer amount of bits.
-BigInt* __quantum__rt__bigint_shiftright(BigInt*, Int);
+BigInt* __quantum__rt__bigint_shiftright(BigInt*, int64_t);
 
 // Returns true if the two big integers are equal, false otherwise.
-Bool __quantum__rt__bigint_equal(BigInt*, BigInt*);
+bool __quantum__rt__bigint_equal(BigInt*, BigInt*);
 
 // Returns true if the first big integer is greater than the second, false
 // otherwise.
-Bool __quantum__rt__bigint_greater(BigInt*, BigInt*);
+bool __quantum__rt__bigint_greater(BigInt*, BigInt*);
 
[…]
