+++ src/qir/qir_dd_backend.cpp
@@ -117,7 +117,7 @@ Result* __quantum__rt__result_get_one() {
   return &one;
 }
 
-Bool __quantum__rt__result_equal(const Result* r1, const Result* r2) {
+bool __quantum__rt__result_equal(const Result* r1, const Result* r2) {
   if (r1 == nullptr) {
     throw std::invalid_argument("First argument must not be null.");
   }
@@ -162,22 +162,22 @@ String* __quantum__rt__string_concatenate(String*, String*) {
   return NULL;
 }
 
-Bool __quantum__rt__string_equal(String*, String*) {
+bool __quantum__rt__string_equal(String*, String*) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return false;
 }
 
-String* __quantum__rt__int_to_string(Int) {
+String* __quantum__rt__int_to_string(int64_t) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
 
-String* __quantum__rt__double_to_string(Double) {
+String* __quantum__rt__double_to_string(double) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
 
-String* __quantum__rt__bool_to_string(Bool) {
+String* __quantum__rt__bool_to_string(bool) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
@@ -208,7 +208,7 @@ String* __quantum__rt__bigint_to_string(BigInt*) {
 }
 
 // *** BIG INTEGERS ***
-BigInt* __quantum__rt__bigint_create_i64(Int) {
+BigInt* __quantum__rt__bigint_create_i64(int64_t) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
@@ -287,38 +287,38 @@ BigInt* __quantum__rt__bigint_bitnot(BigInt*) {
   return NULL;
 }
 
-BigInt* __quantum__rt__bigint_shiftleft(BigInt*, Int) {
+BigInt* __quantum__rt__bigint_shiftleft(BigInt*, int64_t) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
 
-BigInt* __quantum__rt__bigint_shiftright(BigInt*, Int) {
+BigInt* __quantum__rt__bigint_shiftright(BigInt*, int64_t) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
 
-Bool __quantum__rt__bigint_equal(BigInt*, BigInt*) {
+bool __quantum__rt__bigint_equal(BigInt*, BigInt*) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
 
-Bool __quantum__rt__bigint_greater(BigInt*, BigInt*) {
+bool __quantum__rt__bigint_greater(BigInt*, BigInt*) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
 
-Bool __quantum__rt__bigint_greater_eq(BigInt*, BigInt*) {
+bool __quantum__rt__bigint_greater_eq(BigInt*, BigInt*) {
   printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
   return NULL;
 }
 
 // *** TUPLES ***
-Tuple* __quantum__rt__tuple_create(Int) {
[…]
