### 2024-11-23 3eaf66c
🎨 Figure out beginning of array impl
