### 2024-11-23 844096a
✨ Implement basic qis functions
