+++ src/qir/qir_dd_backend.cpp
@@ -27,7 +27,7 @@ auto QIR_DD_Backend::determineAddressMode() -> void {
 }
 
 template <typename... Args>
-auto QIR_DD_Backend::apply(const qc::OpType op, Args... qubits) -> void {
+auto QIR_DD_Backend::apply(const qc::OpType op, const Args... qubits) -> void {
   determineAddressMode();
   // extract addresses from opaque qubit pointers
   const std::vector<Qubit*> rawAddresses = {qubits...};
@@ -42,13 +42,13 @@ auto QIR_DD_Backend::apply(const qc::OpType op, Args... qubits) -> void {
         rawAddresses.cbegin(), rawAddresses.cend(),
         std::back_inserter(addresses), [&](const auto a) {
           if (a < MIN_DYN_ADDRESS) {
-            std::invalid_argument("Qubit not allocated (out of range): " +
-                                  std::to_string(a));
+            std::out_of_range("Qubit not allocated (out of range): " +
+                              std::to_string(a));
           }
           try {
             return qRegister.at(a->id);
           } catch (const std::out_of_range&) {
-            std::throw_with_nested(std::invalid_argument(
+            std::throw_with_nested(std::out_of_range(
                 "Qubit not allocated (not found): " + std::to_string(a->id)));
           }
         });
@@ -62,13 +62,24 @@ auto QIR_DD_Backend::apply(const qc::OpType op, Args... qubits) -> void {
   } else {
     throw std::invalid_argument("Operation type is not known: " + toString(op));
   }
-  const auto& controls = qc::Controls(addresses.cbegin(), addresses.cend() - t);
-  const auto& targets = qc::Targets(addresses.cbegin() - t, addresses.cend());
+  qc::StandardOperation operation;
+  if (addresses.size() > t) { // create controlled operation
+    const auto& controls =
+        qc::Controls(addresses.cbegin(), addresses.cend() - t);
+    const auto& targets = qc::Targets(addresses.cbegin() - t, addresses.cend());
+    operation = qc::StandardOperation(controls, targets, op);
+  } else if (addresses.size() == t) { // create uncontrolled operation
+    operation = qc::StandardOperation(addresses, op);
+  } else {
+    throw std::invalid_argument("Operation requires more qubits than given (" +
+                                qc::toString(op) +
+                                "): " + std::to_string(addresses.size()));
+  }
   // retrieve operation matrix
-  const auto& operation = qc::StandardOperation(controls, targets, op);
   const auto& matrix = getDD(&operation, dd);
   // enlarge quantum state if necessary
-  const auto maxTarget = *std::max_element(targets.cbegin(), targets.cend());
+  const auto maxTarget =
+      *std::max_element(addresses.cbegin(), addresses.cend());
   const auto d = maxTarget - numQubits + 1;
   if (d > 0) {
     qState = dd.kronecker(qState, dd.makeZeroState(d), d);
[…]
