+++ src/qir/qir_dd_backend.cpp
@@ -149,197 +149,196 @@ void __quantum__rt__result_update_reference_count(Result* r, const int32_t k) {
 }
 
 // *** STRINGS ***
-String* __quantum__rt__string_create(const int8_t*) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return NULL;
+String* __quantum__rt__string_create(const char* utf8) {
+  // NOLINTNEXTLINE(cppcoreguidelines-owning-memory)
+  auto* string = new String;
+  string->refcount = 1;
+  string->content = std::string(utf8);
+  return string;
 }
 
-const int8_t* __quantum__rt__string_get_data(String*) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return NULL;
+const char* __quantum__rt__string_get_data(const String* string) {
+  return string->content.c_str();
 }
 
-int32_t __quantum__rt__string_get_length(String*) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return 0;
+int32_t __quantum__rt__string_get_length(const String* string) {
+  return static_cast<int32_t>(string->content.size());
 }
 
-void __quantum__rt__string_update_reference_count(String*, int32_t) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
+void __quantum__rt__string_update_reference_count(String* string,
+                                                  const int32_t k) {
+  if (string != nullptr) {
+    string->refcount += k;
+    if (string->refcount == 0) {
+      // NOLINTNEXTLINE(cppcoreguidelines-owning-memory)
+      delete string;
+    }
+  }
 }
 
 String* __quantum__rt__string_concatenate(String*, String*) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return NULL;
+  throw std::bad_function_call();
 }
 
-bool __quantum__rt__string_equal(String*, String*) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return false;
+bool __quantum__rt__string_equal(const String* string1, const String* string2) {
+  return string1->content == string2->content;
 }
 
-String* __quantum__rt__int_to_string(int64_t) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return NULL;
+String* __quantum__rt__int_to_string(const int64_t z) {
+  // NOLINTNEXTLINE(cppcoreguidelines-owning-memory)
+  auto* string = new String;
+  string->refcount = 1;
+  string->content = std::to_string(z);
+  return string;
 }
 
-String* __quantum__rt__double_to_string(double) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return NULL;
+String* __quantum__rt__double_to_string(const double d) {
+  // NOLINTNEXTLINE(cppcoreguidelines-owning-memory)
+  auto* string = new String;
+  string->refcount = 1;
+  string->content = std::to_string(d);
+  return string;
 }
 
-String* __quantum__rt__bool_to_string(bool) {
-  printf("%s:%s:%d \n", __FILE__, __FUNCTION__, __LINE__);
-  return NULL;
[…]
