### 2024-11-23 2b1857e
🐛 Fix and implement more string functions
