### 2024-11-23 63fe362
🎨 Declare unsused parameters and implement reset function
