+++ src/qir/qir_dd_backend.cpp
@@ -3,8 +3,6 @@
 #include "dd/Operations.hpp"
 #include "qir/qir.h"
 
-#include <cstdio>
-#include <cstdlib>
 #include <exception>
 #include <stdexcept>
 
@@ -262,111 +260,119 @@ String* __quantum__rt__qubit_to_string(const Qubit* qubit) {
   return __quantum__rt__int_to_string(static_cast<int32_t>(qubit->id));
 }
 
-String* __quantum__rt__range_to_string(Range) {
+String* __quantum__rt__range_to_string(Range /*unused*/) {
   throw std::bad_function_call();
 }
 
-String* __quantum__rt__bigint_to_string(BigInt*) {
+String* __quantum__rt__bigint_to_string(BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
 // *** BIG INTEGERS ***
-BigInt* __quantum__rt__bigint_create_i64(int64_t) {
+BigInt* __quantum__rt__bigint_create_i64(int64_t /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_create_array(int32_t, int8_t*) {
+BigInt* __quantum__rt__bigint_create_array(int32_t /*unused*/,
+                                           int8_t* /*unused*/) {
   throw std::bad_function_call();
 }
 
-int8_t* __quantum__rt__bigint_get_data(BigInt*) {
+int8_t* __quantum__rt__bigint_get_data(BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-int32_t __quantum__rt__bigint_get_length(BigInt*) {
+int32_t __quantum__rt__bigint_get_length(BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-void __quantum__rt__bigint_update_reference_count(BigInt*, int32_t) {
+void __quantum__rt__bigint_update_reference_count(BigInt* /*unused*/,
+                                                  int32_t /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_negate(BigInt*) {
+BigInt* __quantum__rt__bigint_negate(BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_add(BigInt*, BigInt*) {
+BigInt* __quantum__rt__bigint_add(BigInt* /*unused*/, BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_subtract(BigInt*, BigInt*) {
+BigInt* __quantum__rt__bigint_subtract(BigInt* /*unused*/, BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_multiply(BigInt*, BigInt*) {
+BigInt* __quantum__rt__bigint_multiply(BigInt* /*unused*/, BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_divide(BigInt*, BigInt*) {
+BigInt* __quantum__rt__bigint_divide(BigInt* /*unused*/, BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_modulus(BigInt*, BigInt*) {
+BigInt* __quantum__rt__bigint_modulus(BigInt* /*unused*/, BigInt* /*unused*/) {
   throw std::bad_function_call();
 }
 
-BigInt* __quantum__rt__bigint_power(BigInt*, int32_t) {
[…]
