### 2024-11-23 5bb4d7e
🚧 Add parameters to apply function
