### 2024-11-24 9bba88c
✨ Implement measure, include adding random number generator
