+++ CMakeLists.txt
@@ -3,7 +3,7 @@ cmake_minimum_required(VERSION 3.19...3.30)
 
 project(
   mqt-mlir
-  LANGUAGES CXX
+  LANGUAGES C CXX
   DESCRIPTION "MQT MLIR - Interfacing between the Munich Quantum Toolkit and MLIR")
 
 # Add path for custom modules
@@ -17,6 +17,16 @@ else()
 endif()
 
 option(BUILD_MQT_MLIR_TESTS "Also build tests for the MQT MLIR project" ${MQT_MLIR_MASTER_PROJECT})
+
+# Require C standard
+set(CMAKE_C_STANDARD 11)
+set(CMAKE_C_STANDARD_REQUIRED True)
+set(CMAKE_C_EXTENSIONS False)
+
+# Require C++ standard
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+set(CMAKE_CXX_EXTENSIONS OFF)
 
 include(cmake/ExternalDependencies.cmake)
 
