### 2024-11-24 d97e6bc
🎨 Add missing functions to QIR
