+++ include/qir/qir.h
@@ -45,18 +45,18 @@ typedef struct BigIntImpl BigInt;
 
 typedef struct ResultImpl Result;
 
-// Returns a constant representing a measurement result zero.
+/// Returns a constant representing a measurement result zero.
 Result* __quantum__rt__result_get_zero();
 
-// Returns a constant representing a measurement result one.
+/// Returns a constant representing a measurement result one.
 Result* __quantum__rt__result_get_one();
 
-// Returns true if the two results are the same, and false if they are
-// different.
+/// Returns true if the two results are the same, and false if they are
+/// different.
 bool __quantum__rt__result_equal(const Result*, const Result*);
 
-// Adds the given integer value to the reference count for the result.
-// Deallocates the result if the reference count becomes 0.
+/// Adds the given integer value to the reference count for the result.
+/// Deallocates the result if the reference count becomes 0.
 void __quantum__rt__result_update_reference_count(Result*, int32_t);
 
 // *** QUBITS ***
@@ -70,122 +70,123 @@ typedef struct QubitImpl Qubit;
 
 typedef struct StringImpl String;
 
-// Creates a string from an array of UTF-8 bytes.
+/// Creates a string from an array of UTF-8 bytes.
 String* __quantum__rt__string_create(const char*);
 
-// Returns a pointer to the zero-terminated array of UTF-8 bytes.
+/// Returns a pointer to the zero-terminated array of UTF-8 bytes.
 const char* __quantum__rt__string_get_data(const String*);
 
-// Returns the length of the byte array that contains the string data.
+/// Returns the length of the byte array that contains the string data.
 int32_t __quantum__rt__string_get_length(const String*);
 
-// Adds the given integer value to the reference count for the string.
-// Deallocates the string if the reference count becomes 0.
+/// Adds the given integer value to the reference count for the string.
+/// Deallocates the string if the reference count becomes 0.
 void __quantum__rt__string_update_reference_count(String*, int32_t);
 
-// Creates a new string that is the concatenation of the two argument strings.
+/// Creates a new string that is the concatenation of the two argument strings.
 String* __quantum__rt__string_concatenate(const String*, const String*);
 
-// Returns true if the two strings are equal, false otherwise.
+/// Returns true if the two strings are equal, false otherwise.
 bool __quantum__rt__string_equal(const String*, const String*);
 
-// Returns a string representation of the integer.
+/// Returns a string representation of the integer.
 String* __quantum__rt__int_to_string(int64_t);
 
-// Returns a string representation of the double.
+/// Returns a string representation of the double.
 String* __quantum__rt__double_to_string(double);
 
-// Returns a string representation of the Boolean.
+/// Returns a string representation of the Boolean.
 String* __quantum__rt__bool_to_string(bool);
 
-// Returns a string representation of the result.
+/// Returns a string representation of the result.
 String* __quantum__rt__result_to_string(const Result*);
 
-// Returns a string representation of the Pauli.
+/// Returns a string representation of the Pauli.
 String* __quantum__rt__pauli_to_string(Pauli);
 
-// Returns a string representation of the qubit.
+/// Returns a string representation of the qubit.
 String* __quantum__rt__qubit_to_string(const Qubit*);
 
-// Returns a string representation of the range.
+/// Returns a string representation of the range.
 String* __quantum__rt__range_to_string(Range);
 
-// Returns a string representation of the big integer.
+/// Returns a string representation of the big integer.
 String* __quantum__rt__bigint_to_string(BigInt*);
 
 // *** BIG INTEGERS ***
 // cf.
[…]
