### 2024-11-24 #7478 @danielskatz
👋 @josh146 - thanks for getting started - is there anything blocking your progress (other than time)?
