### 2024-11-24 #7478 @danielskatz
👋 @1ucian0 & @edyounis - Can you also create your checklists (see https://github.com/openjournals/joss-reviews/issues/7478#issuecomment-2474721914 for instructions) and get started with your reviews?  Thanks.
