### 2024-11-25 1e5e6ef
🩹 do not add empty register
Signed-off-by: burgholzer <burgholzer@me.com>
