### 2024-11-25 b11b565
✅ add a regression test for register-less empty quantum circuit
Signed-off-by: burgholzer <burgholzer@me.com>
