### 2024-11-25 4ed96d5
🎨 refine regression test
Signed-off-by: burgholzer <burgholzer@me.com>
