### 2024-11-25 04a1b37
🐛 Fix out-of-bounds error in ZX `EdgeIterator` and needless dummy register in empty circuit construction (#758)
## Description

This PR fixes an out-of-bounds error in the ZX packages `EdgeIterator`
that would happen whenever a ZX diagram would only consist of eliminated
nodes.

In addition, this PR fixes an oversight in the QuantumComputation
constructor where an empty qubit register `"q"` was added even on a
zero-qubit circuit.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only
[…]
