+++ test/ir/test_qfr_functionality.cpp
@@ -1154,3 +1154,12 @@ TEST_F(QFRFunctionality, emptyPermutation) {
   EXPECT_EQ(perm.maxKey(), 0U);
   EXPECT_EQ(perm.maxValue(), 0U);
 }
+
+TEST_F(QFRFunctionality, NoRegisterOnEmptyCircuit) {
+  // This is a regression test. Previously, the following code would throw an
+  // exception because even zero-qubit circuits had an empty register named "q".
+  qc::QuantumComputation qc(0U);
+  qc.addQubitRegister(1U, "p");
+  EXPECT_NO_THROW(qc.addQubitRegister(1U, "q"));
+  EXPECT_EQ(qc.getQregs().size(), 2U);
+}
