### 2024-11-26 d6cfeb4
🎨 Update QIR header
