### 2024-11-26 f7dbbd0
🚨 adjust for new ruff version
Signed-off-by: burgholzer <burgholzer@me.com>
