### 2024-11-26 6162c24
🚨 enable "all" of ruff
Signed-off-by: burgholzer <burgholzer@me.com>
