+++ src/mqt/core/plugins/qiskit.py
@@ -423,7 +423,13 @@ def _import_definition(
         mapped_cargs = [carg_map[carg] for carg in instruction.clbits]
         operation = instruction.operation
         new_params = _emplace_operation(
-            comp_op, operation, mapped_qargs, mapped_cargs, operation.params, qubit_map, clbit_map
+            comp_op,
+            operation,
+            mapped_qargs,
+            mapped_cargs,
+            operation.params,
+            qubit_map,
+            clbit_map,
         )
         params.extend(new_params)
     return params
