### 2024-11-26 520b2d1
⬆️🪝 Update pre-commit hook rbubley/mirrors-prettier to v3.4.0 (#761)
This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
|
[rbubley/mirrors-prettier](https://redirect.github.com/rbubley/mirrors-prettier)
| repository | minor | `v3.3.3` -> `v3.4.0` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
there, instead [create a Discussion in the Renovate
repository](https://redirect.github.com/renovatebot/renovate/discussions/new)
if you have any questions.

---

##
[…]
