### 2024-11-26 48b927f
🔧 replace deprecated option
Signed-off-by: burgholzer <burgholzer@me.com>
