+++ pyproject.toml
@@ -236,7 +236,7 @@ skip = "*-musllinux_*"
 archs = "auto64"
 test-command = "python -c \"from mqt import core\""
 build-frontend = "build[uv]"
-free-threaded-support = true
+enable = ["cpython-freethreading"]
 manylinux-x86_64-image = "manylinux_2_28"
 manylinux-aarch64-image = "manylinux_2_28"
 manylinux-ppc64le-image = "manylinux_2_28"
