### 2024-11-26 10a78d2
⬆️ update schema-store
Signed-off-by: burgholzer <burgholzer@me.com>
