+++ .pre-commit-config.yaml
@@ -139,6 +139,6 @@ repos:
 
   # Check the pyproject.toml file
   - repo: https://github.com/henryiii/validate-pyproject-schema-store
-    rev: 2024.09.23
+    rev: 2024.11.25
     hooks:
       - id: validate-pyproject
