+++ .pre-commit-config.yaml
@@ -109,7 +109,7 @@ repos:
 
   # Check for spelling
   - repo: https://github.com/crate-ci/typos
-    rev: v1.27.3
+    rev: v1.28.0
     hooks:
       - id: typos
 
