### 2024-11-26 be0f044
⬆️👨‍💻 Update cda-tum/mqt-workflows action to v1.5 (#759)
This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
|
[cda-tum/mqt-workflows](https://redirect.github.com/cda-tum/mqt-workflows)
| action | minor | `v1.4` -> `v1.5` |

---

### Release Notes

<details>
<summary>cda-tum/mqt-workflows (cda-tum/mqt-workflows)</summary>

###
[`v1.5`](https://redirect.github.com/cda-tum/mqt-workflows/compare/v1.4...v1.5)

[Compare
Source](https://redirect.github.com/cda-tum/mqt-workflows/compare/v1.4...v1.5)

</details>

---

### Configuration

📅 **Schedule**: Branc
[…]
