+++ .github/workflows/ci.yml
@@ -14,13 +14,13 @@ concurrency:
 jobs:
   change-detection:
     name: 🔍 Change
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-change-detection.yml@v1.4
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-change-detection.yml@v1.5
 
   cpp-tests:
     name: 🇨‌ Test
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cpp-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@v1.4
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@v1.5
     with:
       cmake-args-macos: -DMQT_CORE_WITH_GMP=ON
 
@@ -28,7 +28,7 @@ jobs:
     name: 🇨‌ Lint
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cpp-linter)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-linter.yml@v1.4
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-linter.yml@v1.5
     with:
       cmake-args: -DBUILD_MQT_CORE_BENCHMARKS=ON
 
@@ -36,19 +36,19 @@ jobs:
     name: 🐍 Test
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-python-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.4
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.5
 
   code-ql:
     name: 📝 CodeQL
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-code-ql)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-code-ql.yml@v1.4
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-code-ql.yml@v1.5
 
   cd:
     name: 🚀 CD
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cd)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-packaging.yml@v1.4
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-packaging.yml@v1.5
 
   required-checks-pass: # This job does nothing and is only used for branch protection
     name: 🚦 Check
