### 2024-11-26 74d83c6
🎨 Implement both measure philosophies
