+++ src/qir/qir_dd_backend.cpp
@@ -21,53 +21,47 @@ auto QIR_DD_Backend::generateRandomSeed() -> uint64_t {
 QIR_DD_Backend::QIR_DD_Backend() : QIR_DD_Backend(generateRandomSeed()) {}
 
 QIR_DD_Backend::QIR_DD_Backend(const uint64_t randomSeed)
-    : addressMode(AddressMode::UNKNOWN), currentMaxAddress(0), numQubits(0),
+    : addressMode(AddressMode::UNKNOWN),
+      currentMaxQubitAddress(MIN_DYN_QUBIT_ADDRESS), currentMaxQubitId(0),
+      currentMaxResultAddress(MIN_DYN_RESULT_ADDRESS), numQubitsInQState(0),
       qState(dd::vEdge::one()), mt(randomSeed) {
-  qRegister = std::unordered_map<qc::Qubit, qc::Qubit>();
-}
-
-auto QIR_DD_Backend::determineAddressMode() -> void {
-  if (addressMode == AddressMode::UNKNOWN) {
-    if (qRegister.empty()) {
-      addressMode = AddressMode::STATIC;
-    } else {
-      addressMode = AddressMode::DYNAMIC;
-    }
-  }
+  qRegister = std::unordered_map<Qubit*, qc::Qubit>();
+  rRegister = std::unordered_map<Result*, Result>();
+  rRegister.emplace(RESULT_ZERO_ADDRESS, Result{0, false});
+  rRegister.emplace(RESULT_ONE_ADDRESS, Result{0, true});
 }
 
 template <typename... Args>
-auto QIR_DD_Backend::translateAddresses(const Args... qubits) const
+auto QIR_DD_Backend::translateAddresses(const Args... qubits)
     -> std::vector<qc::Qubit> {
-  if (addressMode == AddressMode::UNKNOWN) {
-    throw std::logic_error(
-        "Address mode must be determined prior to a call of this function.");
-  }
   // extract addresses from opaque qubit pointers
-  const std::vector<Qubit*> rawAddresses = {qubits...};
-  std::vector<qc::Qubit> addresses;
-  addresses.reserve(rawAddresses.size());
-  if (addressMode == AddressMode::STATIC) {
-    std::transform(rawAddresses.cbegin(), rawAddresses.cend(),
-                   std::back_inserter(addresses),
-                   [](const auto a) { return static_cast<qc::Qubit>(a); });
-  } else { // addressMode == AddressMode::DYNAMIC
-    std::transform(
-        rawAddresses.cbegin(), rawAddresses.cend(),
-        std::back_inserter(addresses), [&](const auto a) {
-          if (a < MIN_DYN_ADDRESS) {
-            std::out_of_range("Qubit not allocated (out of range): " +
-                              std::to_string(a));
-          }
-          try {
-            return qRegister.at(a->id);
-          } catch (const std::out_of_range&) {
-            std::throw_with_nested(std::out_of_range(
-                "Qubit not allocated (not found): " + std::to_string(a->id)));
-          }
-        });
+  const std::vector<Qubit*> addresses = {qubits...};
+  std::vector<qc::Qubit> qubitIds;
+  qubitIds.reserve(addresses.size());
+  if (addressMode != AddressMode::STATIC) {
+    // addressMode == AddressMode::DYNAMIC or AddressMode::UNKNOWN
+    for (Qubit* address : addresses) {
+      try {
+        qubitIds.emplace_back(qRegister.at(address));
+      } catch (const std::out_of_range&) {
+        if (addressMode == AddressMode::DYNAMIC) {
+          std::stringstream ss;
[…]
