### 2024-11-26 9f1d5f7
✨ use dependency groups
Signed-off-by: burgholzer <burgholzer@me.com>
