+++ noxfile.py
@@ -14,22 +14,11 @@ if TYPE_CHECKING:
     from collections.abc import Sequence
 
 nox.needs_version = ">=2024.3.2"
-nox.options.default_venv_backend = "uv|virtualenv"
+nox.options.default_venv_backend = "uv"
 
 nox.options.sessions = ["lint", "tests", "minimums"]
 
 PYTHON_ALL_VERSIONS = ["3.9", "3.10", "3.11", "3.12"]
-
-# The following lists all the build requirements for building the package.
-# Note that this includes transitive build dependencies of package dependencies,
-# since we use `--no-build-isolation` to install the package in editable mode
-# and get better caching performance. This only concerns dependencies that are
-# not available via wheels on PyPI (i.e., only as source distributions).
-BUILD_REQUIREMENTS = [
-    "scikit-build-core>=0.10.1",
-    "setuptools_scm>=8.1",
-    "pybind11>=2.13.5",
-]
 
 if os.environ.get("CI", None):
     nox.options.error_on_missing_interpreters = True
@@ -49,10 +38,8 @@ def _run_tests(
     *,
     install_args: Sequence[str] = (),
     run_args: Sequence[str] = (),
-    extras: Sequence[str] = (),
 ) -> None:
-    posargs = list(session.posargs)
-    env = {}
+    env = {"UV_PROJECT_ENVIRONMENT": session.virtualenv.location}
     if os.environ.get("CI", None) and sys.platform == "win32":
         env["CMAKE_GENERATOR"] = "Ninja"
 
@@ -61,15 +48,35 @@ def _run_tests(
     if shutil.which("ninja") is None:
         session.install("ninja")
 
-    _extras = ["test", *extras]
-    if "--cov" in posargs:
-        _extras.append("coverage")
-        posargs.append("--cov-config=pyproject.toml")
-
-    session.install(*BUILD_REQUIREMENTS, *install_args, env=env)
-    install_arg = f"-ve.[{','.join(_extras)}]"
-    session.install("--no-build-isolation", install_arg, *install_args, env=env)
-    session.run("pytest", *run_args, *posargs, env=env)
+    session.run(
+        "uv",
+        "sync",
+        "--frozen",  # do not update lockfile
+        "--no-dev",  # do not auto-install dev dependencies
+        "--group",
+        "build",  # explicitly install build dependencies
+        "--no-install-project",  # do not install the project
+        "--inexact",  # do not wipe existing environment
+        *install_args,
+        env=env,
+    )
+    session.run(
+        "uv",
+        "run",
+        "--frozen",  # do not update lockfile
+        "--no-dev",  # do not auto-install dev dependencies
+        "--group",
+        "test",  # explicitly install test dependencies
+        "--all-extras",  # install all extras
+        "--no-build-isolation-package",
+        "mqt-core",  # build the project without isolation
+        *install_args,
+        "pytest",
+        *run_args,
+        *session.posargs,
+        "--cov-config=pyproject.toml",
+        env=env,
+    )
 
 
 @nox.session(reuse_venv=True, python=PYTHON_ALL_VERSIONS)
@@ -85,7 +92,6 @@ def minimums(session: nox.Session) -> None:
         session,
         install_args=["--resolution=lowest-direct"],
         run_args=["-Wdefault"],
-        extras=["qiskit", "evaluation"],
     )
     session.run("uv", "pip", "list")
 
@@ -98,24 +104,42 @@ def docs(session: nox.Session) -> None:
     args, posargs = parser.parse_known_args(session.posargs)
 
     serve = args.builder == "html" and session.interactive
-    extra_installs = ["sphinx-autobuild"] if serve else []
-    session.install(*BUILD_REQUIREMENTS, *extra_installs)
[…]
