### 2024-11-26 160385b
✨ enable testing on Python 3.13
Signed-off-by: burgholzer <burgholzer@me.com>
