### 2024-11-26 5872967
🔧 use dependency groups in cibuildwheel and fully test package
Signed-off-by: burgholzer <burgholzer@me.com>
