+++ pyproject.toml
@@ -220,7 +220,8 @@ ignore = ["GH200"]
 build = "cp3*"
 skip = "*-musllinux_*"
 archs = "auto64"
-test-command = "python -c \"from mqt import core\""
+test-groups = ["test"]
+test-command = "pytest {package}/test/python"
 build-frontend = "build[uv]"
 enable = ["cpython-freethreading"]
 manylinux-x86_64-image = "manylinux_2_28"
@@ -242,6 +243,12 @@ environment = { CMAKE_GENERATOR = "Ninja" }
 [[tool.cibuildwheel.overrides]]
 select = "*-macosx_arm64"
 environment = { MACOSX_DEPLOYMENT_TARGET = "11.0" }
+
+# Some of our extra dependencies do not support free-threading yet.
+[[tool.cibuildwheel.overrides]]
+select = "cp3*t-manylinux*"
+test-groups = []
+test-command = "python -c \"from mqt import core\""
 
 
 [tool.uv]
