### 2024-11-26 17e701b
♻️ restructure dependency groups and update lock
Signed-off-by: burgholzer <burgholzer@me.com>
