+++ pyproject.toml
@@ -40,8 +40,6 @@ evaluation = [
 ]
 qiskit = [
   "qiskit[qasm3-import]>=1.0.0",
-  "scipy>=1.14.1; python_version >= '3.13'",
-  "numpy>=2.1; python_version >= '3.13'",
 ]
 
 [project.scripts]
@@ -259,28 +257,33 @@ mqt-core = { workspace = true }
 
 [dependency-groups]
 build = [
-    "pybind11>=2.13.5",
-    "scikit-build-core>=0.10.1",
-    "setuptools-scm>=8.1",
+  "pybind11>=2.13.5",
+  "scikit-build-core>=0.10.1",
+  "setuptools-scm>=8.1",
 ]
 docs = [
-    "furo>=2023.9.10",
-    "mqt-core[evaluation,qiskit]",
-    "myst-nb>=1.1.0",
-    "qiskit[visualization]>=1.0",
-    "setuptools-scm>=8.1",
-    "sphinx-autoapi>=3",
-    "sphinx-copybutton>=0.5",
-    "sphinx-design>=0.6",
-    "sphinxcontrib-bibtex>=2.4.2",
-    "sphinxcontrib-svg2pdfconverter>=1.2",
-    "sphinxext-opengraph>=0.9",
+  "furo>=2023.9.10",
+  "myst-nb>=1.1.0",
+  "setuptools-scm>=8.1",
+  "sphinx-autoapi>=3",
+  "sphinx-copybutton>=0.5",
+  "sphinx-design>=0.6",
+  "sphinxcontrib-bibtex>=2.4.2",
+  "sphinxcontrib-svg2pdfconverter>=1.2",
+  "sphinxext-opengraph>=0.9",
+  "pandas[output-formatting]>=2.1.2",
+  "qiskit[qasm3-import,visualization]>=1.0.0",
+  "scipy>=1.14.1; python_version >= '3.13'",
+  "numpy>=2.1; python_version >= '3.13'",
 ]
 test = [
   "pytest>=7.2",
   "pytest-console-scripts>=1.4",
   "pytest-cov>=4",
-  "mqt-core[evaluation,qiskit]",
+  "pandas[output-formatting]>=2.1.2",
+  "qiskit[qasm3-import]>=1.0.0",
+  "scipy>=1.14.1; python_version >= '3.13'",
+  "numpy>=2.1; python_version >= '3.13'",
 ]
 dev = [
   {include-group = "build"},
