### 2024-11-26 52f909b
🔧 refresh lockfile after minimums session
Signed-off-by: burgholzer <burgholzer@me.com>
