### 2024-11-26 25d9224
⬆️ update development dependency lower bounds
Signed-off-by: burgholzer <burgholzer@me.com>
