+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["scikit-build-core>=0.10.1", "setuptools-scm>=8.1", "pybind11>=2.13.5"]
+requires = ["scikit-build-core>=0.10.7", "setuptools-scm>=8.1", "pybind11>=2.13.6"]
 build-backend = "scikit_build_core.build"
 
 [project]
@@ -257,29 +257,32 @@ mqt-core = { workspace = true }
 
 [dependency-groups]
 build = [
-  "pybind11>=2.13.5",
-  "scikit-build-core>=0.10.1",
+  "pybind11>=2.13.6",
+  "scikit-build-core>=0.10.7",
   "setuptools-scm>=8.1",
 ]
 docs = [
-  "furo>=2023.9.10",
-  "myst-nb>=1.1.0",
+  "furo>=2024.8.6",
+  "ipython>=8.29; python_version >= '3.10'",
+  "myst-nb>=1.1.2",
+  "myst-parser>=4; python_version >= '3.10'",
   "setuptools-scm>=8.1",
-  "sphinx-autoapi>=3",
-  "sphinx-copybutton>=0.5",
-  "sphinx-design>=0.6",
-  "sphinxcontrib-bibtex>=2.4.2",
-  "sphinxcontrib-svg2pdfconverter>=1.2",
-  "sphinxext-opengraph>=0.9",
+  "sphinx>=8; python_version >= '3.10'",
+  "sphinx-autoapi>=3.3.3",
+  "sphinx-copybutton>=0.5.2",
+  "sphinx-design>=0.6.1",
+  "sphinxcontrib-bibtex>=2.6.3",
+  "sphinxcontrib-svg2pdfconverter>=1.2.3",
+  "sphinxext-opengraph>=0.9.1",
   "pandas[output-formatting]>=2.1.2",
   "qiskit[qasm3-import,visualization]>=1.0.0",
   "scipy>=1.14.1; python_version >= '3.13'",
   "numpy>=2.1; python_version >= '3.13'",
 ]
 test = [
-  "pytest>=7.2",
-  "pytest-console-scripts>=1.4",
-  "pytest-cov>=4",
+  "pytest>=8.3.3",
+  "pytest-console-scripts>=1.4.1",
+  "pytest-cov>=6",
   "pandas[output-formatting]>=2.1.2",
   "qiskit[qasm3-import]>=1.0.0",
   "scipy>=1.14.1; python_version >= '3.13'",
