### 2024-11-26 b2a2f74
♻️ do not expose `CircuitInputType` in `mqt.core` package
Signed-off-by: burgholzer <burgholzer@me.com>
