### 2024-11-26 5573f86
💚 update RtD configuration
Signed-off-by: burgholzer <burgholzer@me.com>
