+++ .readthedocs.yaml
@@ -21,19 +21,16 @@ build:
     - asdf plugin add uv
     - asdf install uv latest
     - asdf global uv latest
-    - uv venv
     # Unshallow the git clone and fetch tags to get proper version information
     - git fetch --unshallow --tags
-    # Install the project and its documentation dependencies
-    - uv pip install .[docs] readthedocs-sphinx-ext
     # Run the html builder
-    - . .venv/bin/activate && .venv/bin/python -m sphinx -T -b html -d docs/_build/doctrees -D language=en docs $READTHEDOCS_OUTPUT/html
+    - uv run --frozen --no-dev --group docs -m sphinx -T -b html -d docs/_build/doctrees -D language=en docs $READTHEDOCS_OUTPUT/html
     # Run the htmlzip builder and create a zip file
-    - . .venv/bin/activate && .venv/bin/python -m sphinx -T -b dirhtml -d docs/_build/doctrees -D language=en docs docs/_build/dirhtml
+    - uv run --frozen --no-dev --group docs -m sphinx -T -b dirhtml -d docs/_build/doctrees -D language=en docs docs/_build/dirhtml
     - mkdir -p $READTHEDOCS_OUTPUT/htmlzip
     - zip -r $READTHEDOCS_OUTPUT/htmlzip/html.zip docs/_build/dirhtml/*
     # Run the latex builder and create a pdf file
-    - . .venv/bin/activate && .venv/bin/python -m sphinx -T -b latex -d docs/_build/doctrees -D language=en docs docs/_build/latex
+    - uv run --frozen --no-dev --group docs -m sphinx -T -b latex -d docs/_build/doctrees -D language=en docs docs/_build/latex
     - cd docs/_build/latex && latexmk -pdf -f -dvi- -ps- -interaction=nonstopmode -jobname=$READTHEDOCS_PROJECT
     - mkdir -p $READTHEDOCS_OUTPUT/pdf
     - cp docs/_build/latex/$READTHEDOCS_PROJECT.pdf $READTHEDOCS_OUTPUT/pdf/$READTHEDOCS_PROJECT.pdf
