### 2024-11-26 a3b84ee
🔧 use major tag for `attest-build-provenance`
Signed-off-by: burgholzer <burgholzer@me.com>
