### 2024-11-26 944ab46
🔧 adjust cibuildwheel test-command
Signed-off-by: burgholzer <burgholzer@me.com>
