+++ pyproject.toml
@@ -219,7 +219,7 @@ build = "cp3*"
 skip = "*-musllinux_*"
 archs = "auto64"
 test-groups = ["test"]
-test-command = "pytest {package}/test/python"
+test-command = "pytest {project}/test/python"
 build-frontend = "build[uv]"
 enable = ["cpython-freethreading"]
 manylinux-x86_64-image = "manylinux_2_28"
