### 2024-11-26 e314f1a
🩹 fix freethreading test skip condition
Signed-off-by: burgholzer <burgholzer@me.com>
