### 2024-11-27 f506140
🎨 `emplace_back` instead of `push_back`
Signed-off-by: burgholzer <burgholzer@me.com>
