### 2024-11-27 3a28015
✨ Adopt PEP 735 dependency groups (#762)
## Description

This PR updates the project to use the dependency groups feature that
was proposed in [PEP 735](https://peps.python.org/pep-0735/) and is
available in `uv` since `0.4.27` in favour of optional dependencies.
It adds three explicit dependency groups:
- `build`, containing the build time dependencies,
- `test`, containing the test dependencies (including coverage), and
- `docs`, containing the docs dependencies.

In addition,
- enables testing on Python 3.13 (after some dependency adjustments),
- updates lower bounds for many development de
[…]
