+++ .github/workflows/cd.yml
@@ -42,7 +42,6 @@ jobs:
       - uses: pypa/gh-action-pypi-publish@release/v1
         with:
           repository-url: https://test.pypi.org/legacy/
-          attestations: true
 
   # Downloads the previously generated artifacts and deploys to PyPI on published releases.
   deploy:
@@ -60,9 +59,7 @@ jobs:
           path: dist
           merge-multiple: true
       - name: Generate artifact attestation for sdist and wheel(s)
-        uses: actions/attest-build-provenance@v1.4.4
+        uses: actions/attest-build-provenance@v1
         with:
           subject-path: "dist/*"
       - uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          attestations: true

[…]
