+++ include/qir/qir_dd_backend.hpp
@@ -5,18 +5,18 @@
 #include "qir.h"
 
 struct BigIntImpl {
-  uint64_t refcount;
+  int32_t refcount;
   int64_t i;
 };
 struct ResultImpl {
-  uint64_t refcount;
+  int32_t refcount;
   bool r;
 };
 struct QubitImpl {
   qc::Qubit id;
 };
 struct StringImpl {
-  uint64_t refcount;
+  int32_t refcount;
   std::string content;
   friend auto operator<<(std::ostream& os,
                          const StringImpl& s) -> std::ostream& {
@@ -24,17 +24,17 @@ struct StringImpl {
   }
 };
 struct TupleImpl {
-  uint64_t refcount;
+  int32_t refcount;
   // todo
 };
 struct ArrayImpl {
-  uint64_t refcount;
-  uint64_t aliasCount;
+  int32_t refcount;
+  int32_t aliasCount;
   std::vector<int8_t> data;
-  uint32_t elementSize;
+  int64_t elementSize;
 };
 struct CallablImpl {
-  uint64_t refcount;
+  int32_t refcount;
   // todo
 };
 
@@ -47,20 +47,16 @@ namespace mqt {
  */
 class QIR_DD_Backend {
 public:
-  static constexpr auto* RESULT_ZERO_ADDRESS =
-      reinterpret_cast<Result*>(0x10000);
-  static constexpr auto* RESULT_ONE_ADDRESS =
-      reinterpret_cast<Result*>(0x10001);
+  static constexpr auto RESULT_ZERO_ADDRESS = 0x10000;
+  static constexpr auto RESULT_ONE_ADDRESS = 0x10001;
 
 private:
-  static constexpr auto MIN_DYN_QUBIT_ADDRESS =
-      reinterpret_cast<Qubit*>(0x10000);
+  static constexpr auto MIN_DYN_QUBIT_ADDRESS = 0x10000;
   enum class AddressMode : uint8_t { UNKNOWN, DYNAMIC, STATIC };
 
   AddressMode addressMode;
   std::unordered_map<Qubit*, qc::Qubit> qRegister;
-  static constexpr auto MIN_DYN_RESULT_ADDRESS =
-      reinterpret_cast<Result*>(0x10000);
+  static constexpr auto MIN_DYN_RESULT_ADDRESS = 0x10000;
   std::unordered_map<Result*, Result> rRegister;
   Qubit* currentMaxQubitAddress;
   qc::Qubit currentMaxQubitId;
@@ -74,11 +70,13 @@ private:
   explicit QIR_DD_Backend(uint64_t randomSeed);
 
   [[nodiscard]] static auto generateRandomSeed() -> uint64_t;
-  template <typename... Args>
-  auto translateAddresses(const Args... qubits) -> std::vector<qc::Qubit>;
-  template <typename... Params, typename... Args>
-  auto createOperation(qc::OpType op, const Params... params,
-                       const Args... qubits) const -> qc::StandardOperation;
+  template <size_t SIZE>
+  auto translateAddresses(std::array<Qubit*, SIZE> qubits)
+      -> std::array<qc::Qubit, SIZE>;
+  template <size_t P_NUM, size_t SIZE>
+  auto
+  createOperation(qc::OpType op, std::array<double, P_NUM> params,
+                  std::array<Qubit*, SIZE> qubits) -> qc::StandardOperation;
   auto enlargeState(const qc::StandardOperation& operation) -> void;
 
 public:
@@ -90,11 +88,16 @@ public:
   QIR_DD_Backend(const QIR_DD_Backend&) = delete;
   QIR_DD_Backend& operator=(const QIR_DD_Backend&) = delete;
 
-  template <typename... Params, typename... Args>
[…]
