+++ src/qir/qir_dd_backend.cpp
@@ -3,7 +3,6 @@
 #include "dd/Operations.hpp"
 #include "qir/qir.h"
 
-#include <exception>
 #include <stdexcept>
 
 namespace mqt {
@@ -20,33 +19,36 @@ auto QIR_DD_Backend::generateRandomSeed() -> uint64_t {
 
 QIR_DD_Backend::QIR_DD_Backend() : QIR_DD_Backend(generateRandomSeed()) {}
 
+// NOLINTBEGIN(performance-no-int-to-ptr)
 QIR_DD_Backend::QIR_DD_Backend(const uint64_t randomSeed)
     : addressMode(AddressMode::UNKNOWN),
-      currentMaxQubitAddress(MIN_DYN_QUBIT_ADDRESS), currentMaxQubitId(0),
-      currentMaxResultAddress(MIN_DYN_RESULT_ADDRESS), numQubitsInQState(0),
-      qState(dd::vEdge::one()), mt(randomSeed) {
+      currentMaxQubitAddress(reinterpret_cast<Qubit*>(MIN_DYN_QUBIT_ADDRESS)),
+      currentMaxQubitId(0), currentMaxResultAddress(reinterpret_cast<Result*>(
+                                MIN_DYN_RESULT_ADDRESS)),
+      numQubitsInQState(0), qState(dd::vEdge::one()), mt(randomSeed) {
   qRegister = std::unordered_map<Qubit*, qc::Qubit>();
   rRegister = std::unordered_map<Result*, Result>();
-  rRegister.emplace(RESULT_ZERO_ADDRESS, Result{0, false});
-  rRegister.emplace(RESULT_ONE_ADDRESS, Result{0, true});
+  rRegister.emplace(reinterpret_cast<Result*>(RESULT_ZERO_ADDRESS),
+                    Result{0, false});
+  rRegister.emplace(reinterpret_cast<Result*>(RESULT_ONE_ADDRESS),
+                    Result{0, true});
 }
+// NOLINTEND(performance-no-int-to-ptr)
 
-template <typename... Args>
-auto QIR_DD_Backend::translateAddresses(const Args... qubits)
-    -> std::vector<qc::Qubit> {
+template <size_t SIZE>
+auto QIR_DD_Backend::translateAddresses(std::array<Qubit*, SIZE> qubits)
+    -> std::array<qc::Qubit, SIZE> {
   // extract addresses from opaque qubit pointers
-  const std::vector<Qubit*> addresses = {qubits...};
-  std::vector<qc::Qubit> qubitIds;
-  qubitIds.reserve(addresses.size());
+  std::array<qc::Qubit, SIZE> qubitIds;
   if (addressMode != AddressMode::STATIC) {
     // addressMode == AddressMode::DYNAMIC or AddressMode::UNKNOWN
-    for (Qubit* address : addresses) {
+    for (size_t i = 0; i < SIZE; ++i) {
       try {
-        qubitIds.emplace_back(qRegister.at(address));
+        qubitIds[i] = qRegister.at(qubits[i]);
       } catch (const std::out_of_range&) {
         if (addressMode == AddressMode::DYNAMIC) {
           std::stringstream ss;
-          ss << "Qubit not allocated (not found): " << address;
+          ss << "Qubit not allocated (not found): " << qubits[i];
           throw std::out_of_range(ss.str());
         }
         // addressMode == AddressMode::UNKNOWN
@@ -57,20 +59,21 @@ auto QIR_DD_Backend::translateAddresses(const Args... qubits)
   }
   // addressMode might have changed to STATIC
   if (addressMode == AddressMode::STATIC) {
-    std::transform(addresses.cbegin(), addresses.cend(),
[…]
