### 2024-11-27 fe41152
🐛 Fix segfaults
