+++ src/qir/qir_dd_backend.cpp
@@ -19,24 +19,24 @@ auto QIR_DD_Backend::generateRandomSeed() -> uint64_t {
 
 QIR_DD_Backend::QIR_DD_Backend() : QIR_DD_Backend(generateRandomSeed()) {}
 
-// NOLINTBEGIN(performance-no-int-to-ptr)
 QIR_DD_Backend::QIR_DD_Backend(const uint64_t randomSeed)
     : addressMode(AddressMode::UNKNOWN),
-      currentMaxQubitAddress(reinterpret_cast<Qubit*>(MIN_DYN_QUBIT_ADDRESS)),
-      currentMaxQubitId(0), currentMaxResultAddress(reinterpret_cast<Result*>(
-                                MIN_DYN_RESULT_ADDRESS)),
-      numQubitsInQState(0), qState(dd::vEdge::one()), mt(randomSeed) {
-  qRegister = std::unordered_map<Qubit*, qc::Qubit>();
-  rRegister = std::unordered_map<Result*, Result>();
+      currentMaxQubitAddress(MIN_DYN_QUBIT_ADDRESS), currentMaxQubitId(0),
+      currentMaxResultAddress(MIN_DYN_RESULT_ADDRESS), numQubitsInQState(0),
+      qState(dd::vEdge::one()), mt(randomSeed) {
+  dd.incRef(qState);
+  qRegister = std::unordered_map<const Qubit*, qc::Qubit>();
+  rRegister = std::unordered_map<Result*, ResultStruct>();
+  // NOLINTBEGIN(performance-no-int-to-ptr)
   rRegister.emplace(reinterpret_cast<Result*>(RESULT_ZERO_ADDRESS),
-                    Result{0, false});
+                    ResultStruct{0, false});
   rRegister.emplace(reinterpret_cast<Result*>(RESULT_ONE_ADDRESS),
-                    Result{0, true});
+                    ResultStruct{0, true});
+  // NOLINTEND(performance-no-int-to-ptr)
 }
-// NOLINTEND(performance-no-int-to-ptr)
 
 template <size_t SIZE>
-auto QIR_DD_Backend::translateAddresses(std::array<Qubit*, SIZE> qubits)
+auto QIR_DD_Backend::translateAddresses(std::array<const Qubit*, SIZE> qubits)
     -> std::array<qc::Qubit, SIZE> {
   // extract addresses from opaque qubit pointers
   std::array<qc::Qubit, SIZE> qubitIds;
@@ -70,7 +70,7 @@ auto QIR_DD_Backend::translateAddresses(std::array<Qubit*, SIZE> qubits)
 template <size_t P_NUM, size_t SIZE>
 auto QIR_DD_Backend::createOperation(
     qc::OpType op, std::array<double, P_NUM> params,
-    std::array<Qubit*, SIZE> qubits) -> qc::StandardOperation {
+    std::array<const Qubit*, SIZE> qubits) -> qc::StandardOperation {
   const auto& addresses = translateAddresses(qubits);
   // store parameters into vector
   std::vector<qc::fp> paramVec(params.cbegin(), params.cend());
@@ -86,7 +86,7 @@ auto QIR_DD_Backend::createOperation(
   if (addresses.size() > t) { // create controlled operation
     const auto& controls =
         qc::Controls(addresses.cbegin(), addresses.cend() - t);
-    const auto& targets = qc::Targets(addresses.cbegin() - t, addresses.cend());
+    const auto& targets = qc::Targets(addresses.cbegin() + t, addresses.cend());
     return {controls, targets, op, paramVec};
   }
   if (addresses.size() == t) { // create uncontrolled operation
@@ -103,15 +103,18 @@ auto QIR_DD_Backend::enlargeState(const qc::StandardOperation& operation)
[…]
