### 2024-11-27 8a9406c
✨ expose comparison kind for classic-controlled operations
Signed-off-by: burgholzer <burgholzer@me.com>
