+++ src/python/ir/operations/register_classic_controlled_operation.cpp
@@ -10,16 +10,30 @@
 namespace mqt {
 
 void registerClassicControlledOperation(py::module& m) {
+  py::enum_<qc::ComparisonKind>(m, "ComparisonKind")
+      .value("eq", qc::ComparisonKind::Eq)
+      .value("neq", qc::ComparisonKind::Neq)
+      .value("lt", qc::ComparisonKind::Lt)
+      .value("leq", qc::ComparisonKind::Leq)
+      .value("gt", qc::ComparisonKind::Gt)
+      .value("geq", qc::ComparisonKind::Geq)
+      .export_values()
+      .def("__str__",
+           [](const qc::ComparisonKind& cmp) { return qc::toString(cmp); })
+      .def("__repr__",
+           [](const qc::ComparisonKind& cmp) { return qc::toString(cmp); });
+
   auto ccop = py::class_<qc::ClassicControlledOperation, qc::Operation>(
       m, "ClassicControlledOperation");
 
   ccop.def(
       py::init([](qc::Operation* operation, qc::ClassicalRegister controlReg,
-                  std::uint64_t expectedVal) {
+                  std::uint64_t expectedVal, qc::ComparisonKind cmp) {
         return std::make_unique<qc::ClassicControlledOperation>(
-            operation->clone(), controlReg, expectedVal);
+            operation->clone(), controlReg, expectedVal, cmp);
       }),
-      "operation"_a, "control_register"_a, "expected_value"_a = 1U);
+      "operation"_a, "control_register"_a, "expected_value"_a = 1U,
+      "comparison_kind"_a = qc::ComparisonKind::Eq);
   ccop.def_property_readonly("operation",
                              &qc::ClassicControlledOperation::getOperation,
                              py::return_value_policy::reference_internal);
@@ -27,13 +41,16 @@ void registerClassicControlledOperation(py::module& m) {
       "control_register", &qc::ClassicControlledOperation::getControlRegister);
   ccop.def_property_readonly("expected_value",
                              &qc::ClassicControlledOperation::getExpectedValue);
+  ccop.def_property_readonly(
+      "comparison_kind", &qc::ClassicControlledOperation::getComparisonKind);
   ccop.def("__repr__", [](const qc::ClassicControlledOperation& op) {
     std::stringstream ss;
     const auto& controlReg = op.getControlRegister();
     ss << "ClassicControlledOperation(<...op...>, "
        << "control_register=(" << controlReg.first << ", " << controlReg.second
        << "), "
-       << "expected_value=" << op.getExpectedValue() << ")";
+       << "expected_value=" << op.getExpectedValue() << ", "
+       << "comparison_kind='" << op.getComparisonKind() << "')";
     return ss.str();
   });
 }
