+++ include/mqt-core/ir/QuantumComputation.hpp
@@ -694,29 +694,32 @@ public:
   void classicControlled(const OpType op, const Qubit target,
                          const ClassicalRegister& controlRegister,
                          const std::uint64_t expectedValue = 1U,
+                         const ComparisonKind cmp = ComparisonKind::Eq,
                          const std::vector<fp>& params = {}) {
     classicControlled(op, target, Controls{}, controlRegister, expectedValue,
-                      params);
+                      cmp, params);
   }
   void classicControlled(const OpType op, const Qubit target,
                          const Control control,
                          const ClassicalRegister& controlRegister,
                          const std::uint64_t expectedValue = 1U,
+                         const ComparisonKind cmp = ComparisonKind::Eq,
                          const std::vector<fp>& params = {}) {
     classicControlled(op, target, Controls{control}, controlRegister,
-                      expectedValue, params);
+                      expectedValue, cmp, params);
   }
   void classicControlled(const OpType op, const Qubit target,
                          const Controls& controls,
                          const ClassicalRegister& controlRegister,
                          const std::uint64_t expectedValue = 1U,
+                         const ComparisonKind cmp = ComparisonKind::Eq,
                          const std::vector<fp>& params = {}) {
     checkQubitRange(target, controls);
     checkClassicalRegister(controlRegister);
     std::unique_ptr<Operation> gate =
         std::make_unique<StandardOperation>(controls, target, op, params);
     emplace_back<ClassicControlledOperation>(std::move(gate), controlRegister,
-                                             expectedValue);
+                                             expectedValue, cmp);
   }
 
   /// strip away qubits with no operations applied to them and which do not pop
