+++ pyproject.toml
@@ -185,7 +185,7 @@ isort.required-imports = ["from __future__ import annotations"]
 "test/python/**" = ["T20", "INP001"]
 "docs/**" = ["T20", "INP001"]
 "noxfile.py" = ["T20", "TID251"]
-"*.pyi" = ["D418", "PYI021", "DOC202"]  # pydocstyle
+"*.pyi" = ["D418", "DOC202", "PYI011", "PYI021"]
 "*.ipynb" = [
     "D",    # pydocstyle
     "E402", # Allow imports to appear anywhere in Jupyter notebooks
