### 2024-11-27 60f25e8
🐛 Fix problem in createOperation with measure
