### 2024-11-27 343a893
🐛 Reset backend in initialize function
