+++ src/qir/qir_dd_backend.cpp
@@ -110,7 +110,8 @@ auto QIR_DD_Backend::enlargeState(const qc::StandardOperation& operation)
     -> void {
   const auto& qubits = operation.getUsedQubits();
   const auto maxTarget = *std::max_element(qubits.cbegin(), qubits.cend());
-  if (const auto d = maxTarget - numQubitsInQState + 1; d > 0) {
+  if (maxTarget >= numQubitsInQState) {
+    const auto d = maxTarget - numQubitsInQState + 1;
     dd.decRef(qState);
     qState = dd.kronecker(qState, dd.makeZeroState(d), d);
     dd.incRef(qState);
@@ -212,7 +213,13 @@ bool __quantum__rt__result_equal(Result* result1, Result* result2) {
 void __quantum__rt__result_update_reference_count(Result* result,
                                                   const int32_t k) {
   auto& backend = mqt::QIR_DD_Backend::getInstance();
-  if (result != nullptr) {
+  // NOLINTBEGIN(performance-no-int-to-ptr)
+  if (result != nullptr &&
+      result !=
+          reinterpret_cast<Result*>(mqt::QIR_DD_Backend::RESULT_ZERO_ADDRESS) &&
+      result !=
+          reinterpret_cast<Result*>(mqt::QIR_DD_Backend::RESULT_ONE_ADDRESS)) {
+    // NOLINTEND(performance-no-int-to-ptr)
     auto& refcount = backend.deref(result).refcount;
     refcount += k;
     if (refcount == 0) {
@@ -336,119 +343,200 @@ String* __quantum__rt__qubit_to_string(const Qubit* qubit) {
 }
 
 String* __quantum__rt__range_to_string(Range /*unused*/) {
-  throw std::bad_function_call();
+  std::stringstream ss;
+  ss << __FILE__ << ":" << __LINE__ << ": " << __FUNCTION__
+     << " not implemented.";
+  __quantum__rt__fail(__quantum__rt__string_create(ss.str().c_str()));
 }
 
 String* __quantum__rt__bigint_to_string(BigInt* /*unused*/) {
-  throw std::bad_function_call();
+  std::stringstream ss;
+  ss << __FILE__ << ":" << __LINE__ << ": " << __FUNCTION__
+     << " not implemented.";
+  __quantum__rt__fail(__quantum__rt__string_create(ss.str().c_str()));
 }
 
 // *** BIG INTEGERS ***
 BigInt* __quantum__rt__bigint_create_i64(int64_t /*unused*/) {
-  throw std::bad_function_call();
+  std::stringstream ss;
+  ss << __FILE__ << ":" << __LINE__ << ": " << __FUNCTION__
+     << " not implemented.";
+  __quantum__rt__fail(__quantum__rt__string_create(ss.str().c_str()));
 }
 
 BigInt* __quantum__rt__bigint_create_array(int32_t /*unused*/,
                                            int8_t* /*unused*/) {
-  throw std::bad_function_call();
+  std::stringstream ss;
+  ss << __FILE__ << ":" << __LINE__ << ": " << __FUNCTION__
+     << " not implemented.";
+  __quantum__rt__fail(__quantum__rt__string_create(ss.str().c_str()));
 }
 
 int8_t* __quantum__rt__bigint_get_data(BigInt* /*unused*/) {
-  throw std::bad_function_call();
+  std::stringstream ss;
+  ss << __FILE__ << ":" << __LINE__ << ": " << __FUNCTION__
+     << " not implemented.";
+  __quantum__rt__fail(__quantum__rt__string_create(ss.str().c_str()));
 }
 
 int32_t __quantum__rt__bigint_get_length(BigInt* /*unused*/) {
-  throw std::bad_function_call();
[…]
