### 2024-11-27 97b3c4d
🐛 Fix handling of reset gate
