### 2024-11-27 db3e355
🎨 Implement record two functions
