+++ test/qir/test_qir_backend.cpp
@@ -25,10 +25,10 @@ TEST_F(QIR_DD_BackendTest, BellPairStatic) {
   const auto m1 = __quantum__rt__read_result(r0);
   const auto m2 = __quantum__rt__read_result(r1);
   EXPECT_EQ(m1, m2);
-  // __quantum__rt__result_record_output(r0, "r0");
-  // __quantum__rt__result_record_output(r1, "r1");
-  // EXPECT_TRUE(buffer.str() == "r0: 0\nr1: 0\n" ||
-  //             buffer.str() == "r0: 1\nr1: 1\n");
+  __quantum__rt__result_record_output(r0, "r0");
+  __quantum__rt__result_record_output(r1, "r1");
+  EXPECT_TRUE(buffer.str() == "r0: 0\nr1: 0\n" ||
+              buffer.str() == "r0: 1\nr1: 1\n");
 }
 
 TEST_F(QIR_DD_BackendTest, BellPairDynamic) {
@@ -44,10 +44,10 @@ TEST_F(QIR_DD_BackendTest, BellPairDynamic) {
   const auto m1 = __quantum__rt__read_result(r0);
   const auto m2 = __quantum__rt__read_result(r1);
   EXPECT_EQ(m1, m2);
-  // __quantum__rt__result_record_output(r0, "r0");
-  // __quantum__rt__result_record_output(r1, "r1");
-  // EXPECT_TRUE(buffer.str() == "r0: 0\nr1: 0\n" ||
-  //             buffer.str() == "r0: 1\nr1: 1\n");
+  __quantum__rt__result_record_output(r0, "r0");
+  __quantum__rt__result_record_output(r1, "r1");
+  EXPECT_TRUE(buffer.str() == "r0: 0\nr1: 0\n" ||
+              buffer.str() == "r0: 1\nr1: 1\n");
   __quantum__rt__result_update_reference_count(r0, -1);
   __quantum__rt__result_update_reference_count(r1, -1);
 }
