### 2024-11-27 1f68f3a
🚸 add a convenience method for sampling
Signed-off-by: burgholzer <burgholzer@me.com>
