+++ include/mqt-core/dd/Simulation.hpp
@@ -35,6 +35,22 @@ std::map<std::string, std::size_t>
 simulate(const QuantumComputation* qc, const VectorDD& in, Package<Config>& dd,
          std::size_t shots, std::size_t seed = 0U);
 
+/**
+ * Sample from the output distribution of a quantum computation
+ *
+ * This method classically simulates the quantum computation @p qc and samples
+ * @p shots times from the output distribution. The seed for the random number
+ * generator can be set using the @p seed parameter.
+ *
+ * @param qc The quantum computation to simulate
+ * @param shots The number of shots to sample
+ * @param seed The seed for the random number generator
+ * @return A histogram of the measurement results
+ */
+std::map<std::string, std::size_t> sample(const QuantumComputation& qc,
+                                          std::size_t shots = 1024U,
+                                          std::size_t seed = 0U);
+
 template <class Config>
 void extractProbabilityVector(const QuantumComputation* qc, const VectorDD& in,
                               dd::SparsePVec& probVector, Package<Config>& dd);
