### 2024-11-27 8ef7ca0
✅ add tests for comparison kind functionality
Signed-off-by: burgholzer <burgholzer@me.com>
