### 2024-11-27 996713c
🚨 fix `noreturn` C/C++ specification
