### 2024-11-27 e91f056
🐛 fix vector kronecker product involving terminal node
Signed-off-by: burgholzer <burgholzer@me.com>
