### 2024-11-27 #7478 @danielskatz
👋 @1ucian0 - thanks for letting us know about this. I agree that this is technically a conflict, but also agree that we can waive it, given the details as you mention above.
