### 2024-11-27 #7478 @1ucian0
While not mandatory, it is a usual practice to have a header about the license in the files affected by it. Currently, the files do not have this header. Is it an official policy about it?
