### 2024-11-27 #7478 @1ucian0
> Contribution and authorship: [...] Does the full list of paper authors seem appropriate and complete?

While @burgholzer is the clear main maintainer of the software, I was unable to find @robertwille in the history. Additionally, @ystade seems to have been contributing significantly in the last year. I think the paper could benefit on clarifying the author contributions and the criteria for inclusion.
