### 2024-11-27 c36e65c
🎨 improve algorithm unit tests
Signed-off-by: burgholzer <burgholzer@me.com>
