+++ test/algorithms/test_qpe.cpp
@@ -128,7 +128,7 @@ TEST_P(QPE, QPETest) {
 
   qc::VectorDD e{};
   ASSERT_NO_THROW(
-      { e = simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd); });
+      { e = dd::simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd); });
 
   // account for the eigenstate qubit by adding an offset
   const auto offset = 1ULL << (e.p->v + 1);
@@ -140,7 +140,7 @@ TEST_P(QPE, QPETest) {
   if (exactlyRepresentable) {
     EXPECT_NEAR(probability, 1.0, 1e-8);
   } else {
-    const auto threshold = 4. / (qc::PI * qc::PI);
+    constexpr auto threshold = 4. / (qc::PI * qc::PI);
     // account for the eigenstate qubit in the expected result by shifting and
     // adding 1
     const auto secondAmplitude =
@@ -161,8 +161,7 @@ TEST_P(QPE, IQPETest) {
   ASSERT_EQ(qc.getNqubits(), 2U);
 
   constexpr auto shots = 8192U;
-  auto measurements =
-      simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd, shots);
+  const auto measurements = dd::sample(qc, shots);
 
   // sort the measurements
   using Measurement = std::pair<std::string, std::size_t>;
@@ -176,28 +175,27 @@ TEST_P(QPE, IQPETest) {
                                                       measurements.end(), comp);
 
   std::cout << "Obtained measurements: \n";
-  for (const auto& measurement : ordered) {
-    std::cout << "\t" << measurement.first << ": " << measurement.second << " ("
-              << (measurement.second * 100) / shots << "%)\n";
+  for (const auto& [bitstring, count] : ordered) {
+    std::cout << "\t" << bitstring << ": " << count << " ("
+              << (count * 100) / shots << "%)\n";
   }
 
-  const auto& mostLikely = *ordered.begin();
+  const auto& [mostLikelyResult, mostLikelyCount] = *ordered.begin();
   if (exactlyRepresentable) {
-    EXPECT_EQ(mostLikely.first, expectedResultRepresentation);
-    EXPECT_EQ(mostLikely.second, shots);
+    EXPECT_EQ(mostLikelyResult, expectedResultRepresentation);
+    EXPECT_EQ(mostLikelyCount, shots);
   } else {
     auto it = ordered.begin();
     std::advance(it, 1);
-    const auto& secondMostLikely = *(it);
+    const auto& [secondMostLikelyResult, secondMostLikelyCount] = *(it);
     EXPECT_TRUE(
-        (mostLikely.first == expectedResultRepresentation &&
-         secondMostLikely.first == secondExpectedResultRepresentation) ||
-        (mostLikely.first == secondExpectedResultRepresentation &&
-         secondMostLikely.first == expectedResultRepresentation));
+        (mostLikelyResult == expectedResultRepresentation &&
+         secondMostLikelyResult == secondExpectedResultRepresentation) ||
+        (mostLikelyResult == secondExpectedResultRepresentation &&
+         secondMostLikelyResult == expectedResultRepresentation));
     const auto threshold = 4. / (qc::PI * qc::PI);
-    EXPECT_NEAR(static_cast<double>(mostLikely.second) / shots, threshold,
-                0.02);
[…]
