+++ test/algorithms/test_random_clifford.cpp
@@ -31,9 +31,7 @@ TEST_P(RandomClifford, simulate) {
   auto dd = std::make_unique<dd::Package<>>(nq);
   auto qc = qc::RandomCliffordCircuit(nq, nq * nq, 12345);
   auto in = dd->makeZeroState(nq);
-
-  std::cout << qc << "\n";
-  ASSERT_NO_THROW({ simulate(&qc, in, *dd); });
+  ASSERT_NO_THROW({ dd::simulate(&qc, in, *dd); });
   qc.printStatistics(std::cout);
 }
 
@@ -42,7 +40,6 @@ TEST_P(RandomClifford, buildFunctionality) {
 
   auto dd = std::make_unique<dd::Package<>>(nq);
   auto qc = qc::RandomCliffordCircuit(nq, nq * nq, 12345);
-  std::cout << qc << "\n";
-  ASSERT_NO_THROW({ buildFunctionality(&qc, *dd); });
+  ASSERT_NO_THROW({ dd::buildFunctionality(&qc, *dd); });
   qc.printStatistics(std::cout);
 }
