+++ test/algorithms/test_entanglement.cpp
@@ -14,7 +14,12 @@
 class Entanglement : public testing::TestWithParam<std::size_t> {
 protected:
   void TearDown() override {}
-  void SetUp() override {}
+  void SetUp() override {
+    nq = GetParam();
+    dd = std::make_unique<dd::Package<>>(nq);
+  }
+  std::size_t nq{};
+  std::unique_ptr<dd::Package<>> dd;
 };
 
 INSTANTIATE_TEST_SUITE_P(
@@ -28,27 +33,18 @@ INSTANTIATE_TEST_SUITE_P(
     });
 
 TEST_P(Entanglement, FunctionTest) {
-  const auto nq = GetParam();
-
-  auto dd = std::make_unique<dd::Package<>>(nq);
-  auto qc = qc::Entanglement(nq);
-  auto e = buildFunctionality(&qc, *dd);
-
+  const auto qc = qc::Entanglement(nq);
+  const auto e = dd::buildFunctionality(&qc, *dd);
   ASSERT_EQ(qc.getNops(), nq);
   const auto r = dd->multiply(e, dd->makeZeroState(nq));
-
   ASSERT_EQ(r.getValueByPath(nq, std::string(nq, '0')), dd::SQRT2_2);
   ASSERT_EQ(r.getValueByPath(nq, std::string(nq, '1')), dd::SQRT2_2);
 }
 
 TEST_P(Entanglement, GHZRoutineFunctionTest) {
-  const auto nq = GetParam();
-
-  auto qc = qc::Entanglement(nq);
-  auto dd = std::make_unique<dd::Package<>>(nq);
-  const dd::VectorDD e = simulate(&qc, dd->makeZeroState(qc.getNqubits()), *dd);
+  const auto qc = qc::Entanglement(nq);
+  const auto e = dd::simulate(&qc, dd->makeZeroState(nq), *dd);
   const auto f = dd->makeGHZState(nq);
-
   EXPECT_EQ(e, f);
 }
 

[…]
