### 2024-11-27 c4a17bc
🏁 fix windows compilation
Signed-off-by: burgholzer <burgholzer@me.com>
