### 2024-11-27 48974a5
♻️ Refactor simulation and sampling workflow
Renamed the `simulate` function to `sample` for consistency and clarity. Implemented virtual SWAP gate execution by altering qubit permutations directly. Updated function signatures and internal logic for applying unitary, measurement, and reset operations, supporting the new SWAP handling. Removed redundant permutation management in function calls and streamlined permutation updates.

Signed-off-by: burgholzer <burgholzer@me.com>
