+++ include/mqt-core/dd/Operations.hpp
@@ -174,23 +174,13 @@ getStandardOperationDD(const qc::StandardOperation* op, Package<Config>& dd,
 // the mapping specified by the permutation, e.g.
 //      if perm[0] = 1 and perm[1] = 0
 //      then cx 0 1 will be translated to cx perm[0] perm[1] == cx 1 0
+// An empty permutation marks the identity permutation.
 
 template <class Config>
 qc::MatrixDD getDD(const qc::Operation* op, Package<Config>& dd,
-                   qc::Permutation& permutation, const bool inverse = false) {
+                   const qc::Permutation& permutation = {},
+                   const bool inverse = false) {
   const auto type = op->getType();
-
-  // if a permutation is provided and the current operation is a SWAP, this
-  // routine just updates the permutation
-  if (!permutation.empty() && type == qc::SWAP && !op->isControlled()) {
-    const auto& targets = op->getTargets();
-
-    const auto target0 = targets[0U];
-    const auto target1 = targets[1U];
-    // update permutation
-    std::swap(permutation.at(target0), permutation.at(target1));
-    return dd.makeIdent();
-  }
 
   if (type == qc::Barrier) {
     return dd.makeIdent();
@@ -207,12 +197,8 @@ qc::MatrixDD getDD(const qc::Operation* op, Package<Config>& dd,
   }
 
   if (const auto* standardOp = dynamic_cast<const qc::StandardOperation*>(op)) {
-    auto targets = op->getTargets();
-    auto controls = op->getControls();
-    if (!permutation.empty()) {
-      targets = permutation.apply(targets);
-      controls = permutation.apply(controls);
-    }
+    const auto& targets = permutation.apply(op->getTargets());
+    const auto& controls = permutation.apply(op->getControls());
 
     if (qc::isTwoQubitGate(type)) {
       assert(targets.size() == 2);
@@ -248,27 +234,15 @@ qc::MatrixDD getDD(const qc::Operation* op, Package<Config>& dd,
 }
 
 template <class Config>
-qc::MatrixDD getDD(const qc::Operation* op, Package<Config>& dd,
-                   const bool inverse = false) {
-  qc::Permutation perm{};
-  return getDD(op, dd, perm, inverse);
-}
-
-template <class Config>
-qc::MatrixDD getInverseDD(const qc::Operation* op, Package<Config>& dd) {
-  return getDD(op, dd, true);
-}
-
-template <class Config>
 qc::MatrixDD getInverseDD(const qc::Operation* op, Package<Config>& dd,
-                          qc::Permutation& permutation) {
+                          const qc::Permutation& permutation = {}) {
   return getDD(op, dd, permutation, true);
 }
 
 template <class Config, class Node>
 Edge<Node> applyUnitaryOperation(const qc::Operation* op, Edge<Node> in,
                                  Package<Config>& dd,
-                                 qc::Permutation& permutation) {
+                                 const qc::Permutation& permutation = {}) {
   static_assert(std::is_same_v<Node, dd::vNode> ||
                 std::is_same_v<Node, dd::mNode>);
   assert(op->isUnitary());
@@ -280,66 +254,63 @@ Edge<Node> applyUnitaryOperation(const qc::Operation* op, Edge<Node> in,
 }
 
 template <class Config>
-qc::VectorDD
-applyMeasurement(const qc::Operation* op, qc::VectorDD in, Package<Config>& dd,
-                 const qc::Permutation& permutation, std::mt19937_64& rng,
-                 std::vector<bool>& measurements) {
+qc::VectorDD applyMeasurement(const qc::Operation* op, qc::VectorDD in,
+                              Package<Config>& dd, std::mt19937_64& rng,
+                              std::vector<bool>& measurements,
[…]
