+++ src/dd/Simulation.cpp
@@ -15,16 +15,17 @@
 #include <map>
 #include <random>
 #include <string>
+#include <utility>
 #include <vector>
 
 namespace dd {
 template <class Config>
 std::map<std::string, std::size_t>
-simulate(const QuantumComputation* qc, const VectorDD& in, Package<Config>& dd,
-         std::size_t shots, std::size_t seed) {
-  bool isDynamicCircuit = false;
-  bool hasMeasurements = false;
-  bool measurementsLast = true;
+sample(const QuantumComputation* qc, const VectorDD& in, Package<Config>& dd,
+       std::size_t shots, std::size_t seed) {
+  auto isDynamicCircuit = false;
+  auto hasMeasurements = false;
+  auto measurementsLast = true;
 
   std::mt19937_64 mt{};
   if (seed != 0U) {
@@ -88,6 +89,13 @@ simulate(const QuantumComputation* qc, const VectorDD& in, Package<Config>& dd,
         continue;
       }
 
+      // SWAP gates can be executed virtually by changing the permutation
+      if (op->getType() == OpType::SWAP && !op->isControlled()) {
+        const auto& targets = op->getTargets();
+        std::swap(permutation.at(targets[0U]), permutation.at(targets[1U]));
+        continue;
+      }
+
       e = applyUnitaryOperation(op.get(), e, dd, permutation);
     }
 
@@ -143,23 +151,35 @@ simulate(const QuantumComputation* qc, const VectorDD& in, Package<Config>& dd,
     auto e = in;
 
     for (const auto& op : *qc) {
+      if (op->isUnitary()) {
+        // SWAP gates can be executed virtually by changing the permutation
+        if (op->getType() == OpType::SWAP && !op->isControlled()) {
+          const auto& targets = op->getTargets();
+          std::swap(permutation.at(targets[0U]), permutation.at(targets[1U]));
+          continue;
+        }
+
+        e = applyUnitaryOperation(op.get(), e, dd, permutation);
+        continue;
+      }
+
       if (op->getType() == Measure) {
-        e = applyMeasurement(op.get(), e, dd, permutation, mt, measurements);
+        e = applyMeasurement(op.get(), e, dd, mt, measurements, permutation);
         continue;
       }
 
       if (op->getType() == Reset) {
-        e = applyReset(op.get(), e, dd, permutation, mt);
+        e = applyReset(op.get(), e, dd, mt, permutation);
         continue;
       }
 
       if (op->isClassicControlledOperation()) {
-        e = applyClassicControlledOperation(op.get(), e, dd, permutation,
-                                            measurements);
+        e = applyClassicControlledOperation(op.get(), e, dd, measurements,
+                                            permutation);
         continue;
       }
 
-      e = applyUnitaryOperation(op.get(), e, dd, permutation);
+      qc::unreachable();
     }
 
     // reduce reference count of measured state
@@ -183,7 +203,7 @@ std::map<std::string, std::size_t> sample(const QuantumComputation& qc,
                                           const std::size_t seed) {
   const auto nqubits = qc.getNqubits();
   Package<> dd(nqubits);
-  return simulate(&qc, dd.makeZeroState(nqubits), dd, shots, seed);
+  return sample(&qc, dd.makeZeroState(nqubits), dd, shots, seed);
 }
 
 template <class Config>
@@ -208,16 +228,28 @@ void extractProbabilityVectorRecursive(const QuantumComputation* qc,
   for (auto it = currentIt; it != qc->end(); ++it) {
     const auto& op = (*it);
 
+    // any standard operation is applied here
+    if (op->isUnitary()) {
+      // SWAP gates can be executed virtually by changing the permutation
+      if (op->getType() == OpType::SWAP && !op->isControlled()) {
+        const auto& targets = op->getTargets();
[…]
