### 2024-11-27 ba0ba11
🚨 clang-tidy fixes for evaluation
Signed-off-by: burgholzer <burgholzer@me.com>
