+++ eval/eval_dd_package.cpp
@@ -21,7 +21,6 @@
 #include <fstream>
 #include <ios>
 #include <iostream>
-#include <map>
 #include <memory>
 #include <nlohmann/json.hpp>
 #include <string>
@@ -44,7 +43,7 @@ struct Experiment {
   [[nodiscard]] virtual bool success() const noexcept { return false; }
 };
 
-struct SimulationExperiment : public Experiment {
+struct SimulationExperiment final : public Experiment {
   SimulationExperiment() = default;
   qc::VectorDD sim{};
 
@@ -53,7 +52,7 @@ struct SimulationExperiment : public Experiment {
   }
 };
 
-struct FunctionalityConstructionExperiment : public Experiment {
+struct FunctionalityConstructionExperiment final : public Experiment {
   qc::MatrixDD func{};
 
   [[nodiscard]] bool success() const noexcept override {
@@ -61,6 +60,7 @@ struct FunctionalityConstructionExperiment : public Experiment {
   }
 };
 
+namespace {
 template <class Config>
 MatrixDD buildFunctionality(const qc::Grover* qc, Package<Config>& dd) {
   QuantumComputation groverIteration(qc->getNqubits());
@@ -144,8 +144,7 @@ MatrixDD buildFunctionalityRecursive(const qc::Grover* qc,
 
 std::unique_ptr<SimulationExperiment>
 benchmarkSimulate(const QuantumComputation& qc) {
-  std::unique_ptr<SimulationExperiment> exp =
-      std::make_unique<SimulationExperiment>();
+  auto exp = std::make_unique<SimulationExperiment>();
   const auto nq = qc.getNqubits();
   exp->dd = std::make_unique<Package<>>(nq);
   const auto start = std::chrono::high_resolution_clock::now();
@@ -161,8 +160,7 @@ benchmarkSimulate(const QuantumComputation& qc) {
 std::unique_ptr<FunctionalityConstructionExperiment>
 benchmarkFunctionalityConstruction(const QuantumComputation& qc,
                                    const bool recursive = false) {
-  std::unique_ptr<FunctionalityConstructionExperiment> exp =
-      std::make_unique<FunctionalityConstructionExperiment>();
+  auto exp = std::make_unique<FunctionalityConstructionExperiment>();
   const auto nq = qc.getNqubits();
   exp->dd = std::make_unique<Package<>>(nq);
   const auto start = std::chrono::high_resolution_clock::now();
@@ -186,23 +184,14 @@ benchmarkFunctionalityConstruction(const QuantumComputation& qc,
   exp->stats = dd::getStatistics(exp->dd.get());
   return exp;
 }
-
-std::map<std::string, std::size_t>
-benchmarkSimulateWithShots(const qc::QuantumComputation& qc,
-                           const std::size_t shots) {
-  auto nq = qc.getNqubits();
-  auto dd = std::make_unique<dd::Package<>>(nq);
-  auto in = dd->makeZeroState(nq);
-  auto measurements = simulate(&qc, in, *dd, shots);
-  return measurements;
-}
+} // namespace
 
 static constexpr std::size_t SEED = 42U;
 
 class BenchmarkDDPackage {
 protected:
   void verifyAndSave(const std::string& name, const std::string& type,
-                     qc::QuantumComputation& qc, const Experiment& exp) {
+                     qc::QuantumComputation& qc, const Experiment& exp) const {
 
     const std::string& filename = "results_" + inputFilename + ".json";
 
@@ -234,8 +223,8 @@ protected:
 
[…]
