### 2024-11-27 725039b
🚨 clang-tidy fixes for DD package
Signed-off-by: burgholzer <burgholzer@me.com>
