### 2024-11-27 4592c45
🏁 (finally) fix Windows compilation
Signed-off-by: burgholzer <burgholzer@me.com>
