### 2024-11-27 e00ed14
🎨 optimize conditional checks
Signed-off-by: burgholzer <burgholzer@me.com>
