### 2024-11-27 a453e35
✅ add tests for virtual SWAPs
Signed-off-by: burgholzer <burgholzer@me.com>
