### 2024-11-27 b469a8e
⚗️ slightly tweak reference counting
Signed-off-by: burgholzer <burgholzer@me.com>
