### 2024-11-27 940c70d
🎨 small code style improvement
Signed-off-by: burgholzer <burgholzer@me.com>
