### 2024-11-27 4dec1b7
🎨 some more const
Signed-off-by: burgholzer <burgholzer@me.com>
