### 2024-11-27 efe28ae
🎨 slight coverage improvement
Signed-off-by: burgholzer <burgholzer@me.com>
