### 2024-11-28 0418edc
⚗️ try dynamic allocation of DD package
Signed-off-by: burgholzer <burgholzer@me.com>
