### 2024-11-28 6733f78
🚸♻️ convenience methods and updates for DD functionality (#674)
## Description

This pull request includes several changes aimed at improving code
efficiency and readability across the DD package.

### Utility (functions):

* Introduced new utility functions in `Operations.hpp` for applying
unitary operations, measurements, resets, and classically controlled
operations on quantum states.
* DD routines for state preparation now automatically increase the
respective refcount, so this does not have to be done manually anymore.
* The classic-controlled Operation class now exposes its comparison ki
[…]
