+++ test/dd/test_dd_functionality.cpp
@@ -2,11 +2,13 @@
 #include "circuit_optimizer/CircuitOptimizer.hpp"
 #include "dd/DDDefinitions.hpp"
 #include "dd/FunctionalityConstruction.hpp"
+#include "dd/Node.hpp"
 #include "dd/Operations.hpp"
 #include "dd/Package.hpp"
 #include "dd/Simulation.hpp"
 #include "ir/Permutation.hpp"
 #include "ir/QuantumComputation.hpp"
+#include "ir/operations/ClassicControlledOperation.hpp"
 #include "ir/operations/Control.hpp"
 #include "ir/operations/OpType.hpp"
 #include "ir/operations/StandardOperation.hpp"
@@ -513,4 +515,75 @@ TEST_F(DDFunctionality, FuseSingleQubitGatesAcrossOtherGates) {
   qc.print(std::cout);
   EXPECT_EQ(qc.getNops(), 2);
   EXPECT_EQ(e, f);
+}
+
+TEST_F(DDFunctionality, classicControlledOperationConditions) {
+  const auto cmpKinds = {ComparisonKind::Eq, ComparisonKind::Neq,
+                         ComparisonKind::Lt, ComparisonKind::Leq,
+                         ComparisonKind::Gt, ComparisonKind::Geq};
+  for (const auto kind : cmpKinds) {
+    QuantumComputation qc(1U, 1U);
+    // ensure that the state is |1>.
+    qc.x(0);
+    // measure the qubit to get a classical `1` result to condition on.
+    qc.measure(0, 0);
+    // apply a classic-controlled X gate whenever the measured result compares
+    // as specified by kind with the previously measured result.
+    qc.classicControlled(qc::X, 0, {0, 1U}, 1U, kind);
+    // measure into the same register to check the result.
+    qc.measure(0, 0);
+
+    constexpr auto shots = 16U;
+    const auto hist = dd::sample(qc, shots);
+
+    EXPECT_EQ(hist.size(), 1);
+    const auto& [key, value] = *hist.begin();
+    EXPECT_EQ(value, shots);
+    if (kind == ComparisonKind::Eq || kind == ComparisonKind::Leq ||
+        kind == ComparisonKind::Geq) {
+      EXPECT_EQ(key, "0");
+    } else {
+      EXPECT_EQ(key, "1");
+    }
+  }
+}
+
+TEST_F(DDFunctionality, vectorKroneckerWithTerminal) {
+  const auto root = dd::vEdge::one();
+  const auto zeroState = dd->makeZeroState(1);
+  const auto extendedRoot = dd->kronecker(zeroState, root, 0);
+  EXPECT_EQ(zeroState, extendedRoot);
+}
+
+TEST_F(DDFunctionality, dynamicCircuitSimulationWithSWAP) {
+  QuantumComputation qc(2, 2);
+  qc.x(0);
+  qc.swap(0, 1);
+  qc.measure(1, 0);
+  qc.classicControlled(qc::X, 0, {0, 1U});
+  qc.measure(0, 1);
+
+  constexpr auto shots = 16U;
+  const auto hist = dd::sample(qc, shots);
+  EXPECT_EQ(hist.size(), 1);
+  const auto& [key, value] = *hist.begin();
+  EXPECT_EQ(value, shots);
+  EXPECT_EQ(key, "11");
+}
+
+TEST_F(DDFunctionality, dynamicCircuitProbabilityVectorExtractionWithSWAP) {
+  QuantumComputation qc(2, 2);
+  qc.x(0);
+  qc.swap(0, 1);
+  qc.measure(1, 0);
+  qc.reset(1);
+  qc.measure(1, 1);
+
[…]
