### 2024-11-28 7a818a8
🎨 also delete move constructor and move assignment
