### 2024-11-28 08f3c79
♻️ adjust handling of operations
