### 2024-11-28 13f33d7
🎨 further DD usability improvements (#764)
## Description

This is a direct follow-up to #674 and adds one more convenience routine
that avoids users having to perform reference counting on their own.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's sty
[…]
