+++ src/dd/Simulation.cpp
@@ -285,10 +285,7 @@ void extractProbabilityVectorRecursive(const QuantumComputation* qc,
       if (RealNumber::approximatelyEquals(pone, 1.)) {
         const qc::MatrixDD xGate =
             dd.makeGateDD(X_MAT, static_cast<Qubit>(permutation.at(target)));
-        const qc::VectorDD resetState = dd.multiply(xGate, state);
-        dd.incRef(resetState);
-        dd.decRef(state);
-        state = resetState;
+        state = dd.applyOperation(xGate, state);
         continue;
       }
 
