### 2024-11-28 2ab1e35
🎨 Make DD Package unique pointer in order to allocate mem in heap
