### 2024-11-28 1ed4ee7
🎨 Unroll loop
