+++ src/qir/qir_dd_backend.cpp
@@ -23,7 +23,6 @@
 #include <stdexcept>
 #include <string>
 #include <unordered_map>
-#include <utility>
 #include <vector>
 
 namespace mqt {
@@ -63,26 +62,31 @@ auto QIR_DD_Backend::translateAddresses(std::array<const Qubit*, SIZE> qubits)
   std::array<qc::Qubit, SIZE> qubitIds{};
   if (addressMode != AddressMode::STATIC) {
     // addressMode == AddressMode::DYNAMIC or AddressMode::UNKNOWN
-    for (size_t i = 0; i < SIZE; ++i) {
-      try {
-        qubitIds[i] = qRegister.at(qubits[i]);
-      } catch (const std::out_of_range&) {
-        if (addressMode == AddressMode::DYNAMIC) {
-          std::stringstream ss;
-          ss << __FILE__ << ":" << __LINE__
-             << ": Qubit not allocated (not found): " << qubits[i];
-          throw std::out_of_range(ss.str());
-        }
-        // addressMode == AddressMode::UNKNOWN
-        addressMode = AddressMode::STATIC;
-        break;
+    try {
+      Utils::transform(
+          [&](const auto q) {
+            try {
+              return qRegister.at(q);
+            } catch (const std::out_of_range&) {
+              std::stringstream ss;
+              ss << __FILE__ << ":" << __LINE__
+                 << ": Qubit not allocated (not found): " << q;
+              throw std::out_of_range(ss.str());
+            }
+          },
+          qubits, qubitIds);
+    } catch (std::out_of_range& e) {
+      if (addressMode == AddressMode::DYNAMIC) {
+        throw e; // rethrow
       }
+      // addressMode == AddressMode::UNKNOWN
+      addressMode = AddressMode::STATIC;
     }
   }
   // addressMode might have changed to STATIC
   if (addressMode == AddressMode::STATIC) {
     Utils::transform(
-        [](auto q) {
+        [](const auto q) {
           return static_cast<qc::Qubit>(reinterpret_cast<uintptr_t>(q));
         },
         qubits, qubitIds);
@@ -149,19 +153,6 @@ auto QIR_DD_Backend::apply(const qc::OpType op,
   qState = dd::applyUnitaryOperation(&operation, qState, *dd);
 }
 
-template <size_t... I>
-auto loop(const std::array<qc::Qubit, sizeof...(I)>& targets,
-          std::array<Result*, sizeof...(I)>& results, dd::vEdge& qState,
-          dd::Package<>& dd, std::mt19937_64& mt, QIR_DD_Backend& inst,
-          std::index_sequence<I...>) -> void {
-  dd::Qubit q;
-  char result;
-  ((q = static_cast<dd::Qubit>(targets[I]),
-    result = dd.measureOneCollapsing(qState, q, mt),
-    inst.deref(results[I]).r = result == '1'),
-   ...);
-}
-
 template <size_t SIZE>
 auto QIR_DD_Backend::measure(std::array<const Qubit*, SIZE> qubits,
                              std::array<Result*, SIZE> results) -> void {
@@ -169,13 +160,13 @@ auto QIR_DD_Backend::measure(std::array<const Qubit*, SIZE> qubits,
   const auto maxQubit = *std::max_element(targets.cbegin(), targets.cend());
   enlargeState(maxQubit);
   // measure qubits
-  loop(targets, results, qState, *dd, mt, *this,
-       std::make_index_sequence<SIZE>{});
-  // for (size_t i = 0; i < SIZE; ++i) {
-  //   const auto q = static_cast<dd::Qubit>(targets[i]);
-  //   const auto& result = dd->measureOneCollapsing(qState, q, mt);
[…]
