### 2024-11-28 #7478 @burgholzer
> > Contribution and authorship: [...] Does the full list of paper authors seem appropriate and complete?
>
> While @burgholzer is the clear main maintainer of the software, I was unable to find @robertwille in the history. Additionally, @ystade seems to have been contributing significantly in the last year. I think the paper could benefit on clarifying the author contributions and the criteria for inclusion.

You are definitely right, and I think it would be a good idea to update the author list for this submission. Given his significant code contributions over the past year, I would like to add Yannick to the middle of the author list.
Robert is included in the author list based on his substantial impact in shaping mqt-core and the research methods that have been developed as part or on top of mqt-core. Despite not showing up in terms of commits, the countless discussions are very much reflected in the code as it stands and the software would not exist without his support.

@danielskatz Similar question to above: Do I simply update the `joss-paper` branch with the updated metadata?
