### 2024-11-29 b5b5677
🎨 Simplify helper functions
