+++ include/qir/qir_dd_backend.hpp
@@ -55,39 +55,93 @@ namespace mqt {
 
 class Utils {
 private:
-  template <typename Func, typename Store, typename S, typename R,
-            std::size_t... I>
-  constexpr static void
-  transformStoreImpl(Func&& func, Store&& store,
-                     const std::array<S, sizeof...(I)>& source,
-                     std::array<R, sizeof...(I)>& result,
-                     std::index_sequence<I...> /*anonymous*/) {
-    ((std::forward<Store>(store)(result[I],
-                                 std::forward<Func>(func)(source[I]))),
-     ...);
+  template <typename Func, typename S, typename T, std::size_t... I>
+  constexpr static void apply2Impl(Func&& func,
+                                   std::array<S, sizeof...(I)>& arg1,
+                                   std::array<T, sizeof...(I)>& arg2,
+                                   std::index_sequence<I...> /*anonymous*/) {
+    ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
+  }
+  template <typename Func, typename S, typename T, std::size_t... I>
+  constexpr static void apply2Impl(Func&& func,
+                                   const std::array<S, sizeof...(I)>& arg1,
+                                   std::array<T, sizeof...(I)>& arg2,
+                                   std::index_sequence<I...> /*anonymous*/) {
+    ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
+  }
+  template <typename Func, typename S, typename T, std::size_t... I>
+  constexpr static void apply2Impl(Func&& func,
+                                   std::array<S, sizeof...(I)>& arg1,
+                                   const std::array<T, sizeof...(I)>& arg2,
+                                   std::index_sequence<I...> /*anonymous*/) {
+    ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
+  }
+  template <typename Func, typename S, typename T, std::size_t... I>
+  constexpr static void apply2Impl(Func&& func,
+                                   const std::array<S, sizeof...(I)>& arg1,
+                                   const std::array<T, sizeof...(I)>& arg2,
+                                   std::index_sequence<I...> /*anonymous*/) {
+    ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
   }
 
 public:
   /// Helper function to apply a function to each element of the array and store
   /// the result in another equally sized array.
   template <typename Func, typename S, typename R, std::size_t N>
+  constexpr static void transform(Func&& func, std::array<S, N>& source,
+                                  std::array<R, N>& result) {
+    apply2(
+        [&func](auto& value, auto& container) {
+          container = std::forward<Func>(func)(value);
+        },
+        source, result);
+  }
+  /// Helper function to apply a function to each element of the array and store
+  /// the result in another equally sized array.
+  template <typename Func, typename S, typename R, std::size_t N>
   constexpr static void transform(Func&& func, const std::array<S, N>& source,
                                   std::array<R, N>& result) {
-    transformStoreImpl(
-        std::forward<Func>(func),
-        [](auto& container, const auto value) { container = value; }, source,
-        result, std::make_index_sequence<N>{});
+    apply2(
+        [&func](auto& value, auto& container) {
+          container = std::forward<Func>(func)(value);
+        },
+        source, result);
   }
   /// Helper function to apply a function to each element of the array and store
   /// the result with the help of the store function in another equally sized
   /// array.
-  template <typename Func, typename Store, typename S, typename R,
-            std::size_t N>
-  constexpr static void transformStore(Func&& func, Store&& store,
-                                       const std::array<S, N>& source,
-                                       std::array<R, N>& result) {
-    transformStoreImpl(std::forward<Func>(func), std::forward<Store>(store),
[…]
