### 2024-11-29 0654518
🎨 Template magic works
