+++ include/qir/qir_dd_backend.hpp
@@ -14,6 +14,7 @@
 #include <ostream>
 #include <random>
 #include <string>
+#include <tuple>
 #include <unordered_map>
 #include <utility>
 #include <vector>
@@ -55,39 +56,54 @@ namespace mqt {
 
 class Utils {
 private:
-  template <typename Func, typename S, typename T, std::size_t... I>
+  template <typename Func, typename S, typename T, size_t... I>
   constexpr static void apply2Impl(Func&& func,
                                    std::array<S, sizeof...(I)>& arg1,
                                    std::array<T, sizeof...(I)>& arg2,
                                    std::index_sequence<I...> /*anonymous*/) {
     ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
   }
-  template <typename Func, typename S, typename T, std::size_t... I>
+  template <typename Func, typename S, typename T, size_t... I>
   constexpr static void apply2Impl(Func&& func,
                                    const std::array<S, sizeof...(I)>& arg1,
                                    std::array<T, sizeof...(I)>& arg2,
                                    std::index_sequence<I...> /*anonymous*/) {
     ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
   }
-  template <typename Func, typename S, typename T, std::size_t... I>
+  template <typename Func, typename S, typename T, size_t... I>
   constexpr static void apply2Impl(Func&& func,
                                    std::array<S, sizeof...(I)>& arg1,
                                    const std::array<T, sizeof...(I)>& arg2,
                                    std::index_sequence<I...> /*anonymous*/) {
     ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
   }
-  template <typename Func, typename S, typename T, std::size_t... I>
+  template <typename Func, typename S, typename T, size_t... I>
   constexpr static void apply2Impl(Func&& func,
                                    const std::array<S, sizeof...(I)>& arg1,
                                    const std::array<T, sizeof...(I)>& arg2,
                                    std::index_sequence<I...> /*anonymous*/) {
     ((std::forward<Func>(func)(arg1[I], arg2[I])), ...);
+  }
+  template <size_t I, size_t N, typename T, typename... Args>
+  constexpr static void fillArray(std::array<T, N>& arr, T head, Args... tail) {
+    arr[I] = head;
+    if constexpr (N - I > 1) {
+      fillArray<I + 1>(arr, tail...);
+    }
+  }
+  template <size_t N, typename T, typename... Args>
+  constexpr static auto skipFirstNArgs(T head, Args... tail) {
+    if constexpr (N == 0) {
+      return std::make_tuple<>(head, tail...);
+    } else {
+      return skipFirstNArgs<N - 1>(tail...);
+    }
   }
 
 public:
   /// Helper function to apply a function to each element of the array and store
   /// the result in another equally sized array.
-  template <typename Func, typename S, typename R, std::size_t N>
+  template <typename Func, typename S, typename R, size_t N>
   constexpr static void transform(Func&& func, std::array<S, N>& source,
                                   std::array<R, N>& result) {
     apply2(
@@ -98,7 +114,7 @@ public:
   }
   /// Helper function to apply a function to each element of the array and store
   /// the result in another equally sized array.
-  template <typename Func, typename S, typename R, std::size_t N>
+  template <typename Func, typename S, typename R, size_t N>
   constexpr static void transform(Func&& func, const std::array<S, N>& source,
                                   std::array<R, N>& result) {
     apply2(
[…]
