+++ src/qir/qir_dd_backend.cpp
@@ -56,7 +56,7 @@ QIR_DD_Backend::QIR_DD_Backend(const uint64_t randomSeed)
 }
 
 template <size_t SIZE>
-auto QIR_DD_Backend::translateAddresses(std::array<const Qubit*, SIZE> qubits)
+auto QIR_DD_Backend::translateAddresses(std::array<Qubit*, SIZE> qubits)
     -> std::array<qc::Qubit, SIZE> {
   // extract addresses from opaque qubit pointers
   std::array<qc::Qubit, SIZE> qubitIds{};
@@ -97,7 +97,7 @@ auto QIR_DD_Backend::translateAddresses(std::array<const Qubit*, SIZE> qubits)
 template <size_t P_NUM, size_t SIZE>
 auto QIR_DD_Backend::createOperation(
     qc::OpType op, std::array<double, P_NUM> params,
-    std::array<const Qubit*, SIZE> qubits) -> qc::StandardOperation {
+    std::array<Qubit*, SIZE> qubits) -> qc::StandardOperation {
   const auto& addresses = translateAddresses(qubits);
   // store parameters into vector
   const std::vector<qc::fp> paramVec(params.cbegin(), params.cend());
@@ -143,18 +143,21 @@ auto QIR_DD_Backend::enlargeState(const std::uint64_t maxQubit) -> void {
   }
 }
 
-template <size_t P_NUM, size_t SIZE>
-auto QIR_DD_Backend::apply(const qc::OpType op,
-                           std::array<double, P_NUM> params,
-                           std::array<const Qubit*, SIZE> qubits) -> void {
-  const auto& operation = createOperation<P_NUM, SIZE>(op, params, qubits);
+template <typename... Args>
+auto QIR_DD_Backend::apply(const qc::OpType op, Args... args) -> void {
+  constexpr auto nParams = (std::is_same_v<qc::fp, Args> + ...);
+  constexpr auto nQubits = (std::is_same_v<Qubit*, Args> + ...);
+  const auto params = Utils::getFirstNArgs<nParams, qc::fp>(args...);
+  const auto qubits = Utils::getNAfterMArgs<nParams, nQubits, Qubit*>(args...);
+
+  const auto& operation = createOperation<nParams, nQubits>(op, params, qubits);
   const auto& usedQubits = operation.getUsedQubits();
   enlargeState(*usedQubits.crbegin());
   qState = dd::applyUnitaryOperation(&operation, qState, *dd);
 }
 
 template <size_t SIZE>
-auto QIR_DD_Backend::measure(std::array<const Qubit*, SIZE> qubits,
+auto QIR_DD_Backend::measure(std::array<Qubit*, SIZE> qubits,
                              std::array<Result*, SIZE> results) -> void {
   const auto& targets = translateAddresses(qubits);
   const auto maxQubit = *std::max_element(targets.cbegin(), targets.cend());
@@ -170,7 +173,7 @@ auto QIR_DD_Backend::measure(std::array<const Qubit*, SIZE> qubits,
 }
 
 template <size_t SIZE>
-auto QIR_DD_Backend::reset(std::array<const Qubit*, SIZE> qubits) -> void {
+auto QIR_DD_Backend::reset(std::array<Qubit*, SIZE> qubits) -> void {
   const auto& targets = translateAddresses(qubits);
   const auto maxQubit = *std::max_element(targets.cbegin(), targets.cend());
   enlargeState(maxQubit);
[…]
