### 2024-11-29 936205c
🎨 Use forward refs instead of const varsion of function
