### 2024-11-29 8a82504
🎨 Finally make apply pretty
