### 2024-11-29 19d40f2
🎨 Also make the measure function pretty.
