### 2024-11-30 0e966c7
⬆️🪝 Update pre-commit hook astral-sh/ruff-pre-commit to v0.8.1
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
