+++ .pre-commit-config.yaml
@@ -59,7 +59,7 @@ repos:
 
   # Python linting using ruff
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.8.0
+    rev: v0.8.1
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes"]
