### 2024-11-30 09763ae
⬆️🪝 Update pre-commit hook rbubley/mirrors-prettier to v3.4.1
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
