### 2024-11-30 #7478 @danielskatz
@burgholzer - yes, you just update the paper content, and since we're in the middle of the review, make clear to the reviewers what changes you are making so they can factor that into their ongoing reviews.
