### 2024-11-30 #7478 @danielskatz
👋 @edyounis - please go ahead and get started on your review when you have the chance - see [above](https://github.com/openjournals/joss-reviews/issues/7478#issuecomment-2474721914) for instructions on generating your review checklist.
