### 2024-12-02 2574cbb
🎨 Make template magic pretty
