+++ include/qir/qir_dd_backend.hpp
@@ -15,6 +15,7 @@
 #include <random>
 #include <string>
 #include <tuple>
+// ReSharper disable once CppUnusedIncludeDirective
 #include <type_traits>
 #include <unordered_map>
 #include <utility>
@@ -55,13 +56,63 @@ struct CallablImpl {
 
 namespace mqt {
 
+// Primary template
+template <typename T> static constexpr bool IS_STD_ARRAY_V = false;
+// Specialization for std::array
+template <typename T, std::size_t N>
+static constexpr bool IS_STD_ARRAY_V<std::array<T, N>> = true;
+
+// Primary template
+template <typename T, typename... Args> struct SizeOfPackOfType;
+// Base case: no matching types
+template <typename T> struct SizeOfPackOfType<T> {
+  static constexpr size_t VALUE = 0;
+};
+// Recursive case: first type does not match
+template <typename T, typename U, typename... Args>
+struct SizeOfPackOfType<T, U, Args...> {
+  static constexpr size_t VALUE = SizeOfPackOfType<T>::VALUE;
+};
+// Recursive case: first type matches
+template <typename T, typename... Args> struct SizeOfPackOfType<T, T, Args...> {
+  static constexpr size_t VALUE = 1 + SizeOfPackOfType<T, Args...>::VALUE;
+};
+// Helper variable template
+template <typename T, typename... Args>
+static inline constexpr size_t SIZE_OF_PACK_OF_TYPE_V =
+    SizeOfPackOfType<T, Args...>::VALUE;
+
+// Primary template
+template <template <typename, typename...> class V, typename T,
+          typename... Args>
+struct SkipUntilType;
+// Base case: no matching types
+template <template <typename, typename...> class V, typename T>
+struct SkipUntilType<V, T> {
+  static constexpr size_t VALUE = 0;
+};
+// Recursive case: skip until T is found
+template <template <typename, typename...> class V, typename T, typename U,
+          typename... Args>
+struct SkipUntilType<V, T, U, Args...> {
+  static constexpr size_t VALUE = SkipUntilType<V, T, Args...>::VALUE;
+};
+// Recursive case: T is found
+template <template <typename, typename...> class V, typename T,
+          typename... Args>
+struct SkipUntilType<V, T, T, Args...> {
+  static constexpr size_t VALUE = V<T, T, Args...>::VALUE;
+};
+// Helper type template
+template <template <typename, typename...> class V, typename T,
+          typename... Args>
+static inline constexpr size_t SKIP_UNTIL_TYPE_V =
+    SkipUntilType<V, T, Args...>::VALUE;
+
 class Utils {
 private:
-  template <typename T> constexpr static bool is_std_array_v = false;
-  template <typename T, std::size_t N>
-  constexpr static bool is_std_array_v<std::array<T, N>> = true;
   template <typename Func, typename S, typename T, size_t... I>
-  constexpr static void
+  static constexpr void
   apply2Impl(Func&& func, S&& arg1, T&& arg2,
              [[maybe_unused]] std::index_sequence<I...> _) {
     ((std::forward<Func>(func)(std::forward<S>(arg1)[I],
@@ -69,18 +120,27 @@ private:
      ...);
   }
   template <size_t I, size_t N, typename T, typename... Args>
-  constexpr static void fillArray(std::array<T, N>& arr, T head, Args... tail) {
+  static constexpr void fillArray(std::array<T, N>& arr, T head, Args... tail) {
     arr[I] = head;
     if constexpr (N - I > 1) {
       fillArray<I + 1>(arr, tail...);
     }
   }
   template <size_t N, typename T, typename... Args>
-  constexpr static auto skipFirstNArgs(T head, Args... tail) {
+  static constexpr auto skipNArgs(T head, Args... tail) {
[…]
