### 2024-12-02 43453bd
✨ Implement more functions
