### 2024-12-02 bbbeb9d
🚚 Add other examples
