+++ .pre-commit-config.yaml
@@ -47,6 +47,7 @@ repos:
     rev: v1.24.6
     hooks:
       - id: typos
+        exclude: .*\.ll
 
   # Clang-format the C++ part of the code base automatically
   - repo: https://github.com/pre-commit/mirrors-clang-format
