### 2024-12-02 3481566
👥 add Yannick (@ystade) as author to JOSS paper
Signed-off-by: burgholzer <burgholzer@me.com>
