### 2024-12-02 99aa517
👥 add Yannick (@ystade) as author to JOSS paper
Signed-off-by: burgholzer <burgholzer@me.com>
