### 2024-12-02 #7478 @burgholzer
> > > Contribution and authorship: [...] Does the full list of paper authors seem appropriate and complete?
> >
> >
> > While @burgholzer is the clear main maintainer of the software, I was unable to find @robertwille in the history. Additionally, @ystade seems to have been contributing significantly in the last year. I think the paper could benefit on clarifying the author contributions and the criteria for inclusion.
>
> You are definitely right, and I think it would be a good idea to update the author list for this submission. Given his significant code contributions over the past year, I would like to add Yannick to the middle of the author list. Robert is included in the author list based on his substantial impact in shaping mqt-core and the research methods that have been developed as part or on top of mqt-core. Despite not showing up in terms of commits, the countless discussions are very much reflected in the code as it stands and the software would not exist without his support.

Based on the above, I added Yannick as a co-author in https://github.com/cda-tum/mqt-core/commit/3481566031eb2ff1d7be02740be34434880beeae
