### 2024-12-02 95d2417
🚨 extract file-wide ruff ignore
Signed-off-by: burgholzer <burgholzer@me.com>
