### 2024-12-02 43a182c
📄 add license headers to all relevant files
Signed-off-by: burgholzer <burgholzer@me.com>
