+++ src/dd/FunctionalityConstruction.cpp
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2024 Chair for Design Automation, TUM
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
 #include "dd/FunctionalityConstruction.hpp"
 
 #include "dd/Package.hpp"
