+++ src/python/ir/operations/register_compound_operation.cpp
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2024 Chair for Design Automation, TUM
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
 #include "ir/operations/CompoundOperation.hpp"
 #include "ir/operations/Operation.hpp"
 #include "python/pybind11.hpp"
