+++ test/python/test_evaluation.py
@@ -1,3 +1,10 @@
+# Copyright (c) 2024 Chair for Design Automation, TUM
+# All rights reserved.
+#
+# SPDX-License-Identifier: MIT
+#
+# Licensed under the MIT License
+
 """Test the evaluation of JSON results."""
 
 from __future__ import annotations
