+++ src/mqt/core/plugins/__init__.py
@@ -1,3 +1,10 @@
+# Copyright (c) 2023 - 2024 Chair for Design Automation, TUM
+# All rights reserved.
+#
+# SPDX-License-Identifier: MIT
+#
+# Licensed under the MIT License
+
 """Plugins for the core package."""
 
 from __future__ import annotations
