### 2024-12-02 43956ec
🔧 adjust config
Signed-off-by: burgholzer <burgholzer@me.com>
