+++ src/zx/Rules.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 - 2024 Chair for Design Automation, TUM
+ * Copyright (c) 2024 Chair for Design Automation, TUM
  * All rights reserved.
  *
  * SPDX-License-Identifier: MIT
