### 2024-12-02 c606fc0
📄 add license headers (#770)
## Description

This PR adds license headers to all relevant files in the repository.
To reduce the maintenance burden, a pre-commit check is set up that
automatically checks for the presence of respective headers and adds
them if missing.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The p
[…]
