+++ eval/eval_dd_package.cpp
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2024 Chair for Design Automation, TUM
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
 #include "algorithms/BernsteinVazirani.hpp"
 #include "algorithms/Entanglement.hpp"
 #include "algorithms/Grover.hpp"
