+++ src/ir/parsers/qasm3_parser/passes/TypeCheckPass.cpp
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2024 Chair for Design Automation, TUM
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
 #include "ir/parsers/qasm3_parser/passes/TypeCheckPass.hpp"
 
 #include "ir/parsers/qasm3_parser/Exception.hpp"
