### 2024-12-02 #7478 @burgholzer
> > While not mandatory, it is a usual practice to have a header about the license in the files affected by it. Currently, the files do not have this header. Is it an official policy about it?
>
> I'd be happy to add license headers to the project. Certainly does not hurt.

Added license headers to the files in https://github.com/cda-tum/mqt-core/commit/c606fc058ea52cc5628ea2933eb22aee43988897 and merged the latest `main` branch into the `joss-paper` branch in https://github.com/cda-tum/mqt-core/commit/e84b179248548e4674ffbaadb042db6ce2f6bb8b
