### 2024-12-02 #7478 @josh146
> is there anything blocking your progress (other than time)?

Just time, had quite a busy week but should have some more time now!
