### 2024-12-03 #7478 @edyounis
> Installation: Does installation proceed as outlined in the documentation?

Installation went well, but it would be nice in the documentation to have a "Run this command to ensure installation worked..." I just did `python -c 'import mqt'` and saw no error. I don't think this is important or required, but is always nice to see.
