### 2024-12-03 #7478 @edyounis
> Functionality: Have the functional claims of the software been confirmed?

You mentioned a ZX-calculus module in the paper, but I could not find the relevant documentation. Can you point me in the right direction?
