### 2024-12-03 #7478 @edyounis
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

Regarding this requirement, the paper may benefit from a related works section.
