### 2024-12-03 #7478 @josh146
@burgholzer thanks for clarifying the author list above! I had the same question as @1ucian0. Separately, I was wondering about the inclusion of Tom Peham, who also looks to have contributed significantly, but not recently -- I assume that these code additions may not survive in the current version of MQT submitted to JOSS?

- Installation: I installed via pip, but did not attempt to install from source or attempt to integrate it into an existing project.

- Documentation: I've gone through the quickstart, and all examples work great. I have a few minor comments/questions here:

  - Something I struggled to understand was the notation/differentiation between the algorithmic qubit mapping, and the output measurement mapping. Maybe something that could be clarified a bit in the documentation?

  - I also struggled to parse/fully understand the 'Benchmarking the DD Package' part of the documentation. This is partly because I was missing the context of what you use the DD package for (as decision diagrams weren't introduced in the quickstart), as well as the motivation for benchmarking (comparing changes in MQT-core, comparing changes in MQT-core workflows, etc.).

    Separately, I looked though the `mqt.core.dd` docstrings, but couldn't find an example to showcase the decision diagram features. But perhaps I am missing something :)

  - I have similar questions about the ZX functionality.

- I found the contributing section really well done, and liked that it had a pull request workflow!

Regarding the paper, I thought it was succinct and well written, with a clear statement of need. Based on the review checklist, it is perhaps missing a section for

>  State of the field: Do the authors describe how this software compares to other commonly-used packages?

In addition, a small workflow example demonstrating the ZX and DD usage would be beneficial.
