### 2024-12-06 #7478 @burgholzer
Thanks for all your comments, and sorry for the lack of replies these past days. I have been quite occupied with meetings. I'll get back to the individual points at the beginning of next week, but let me try to provide a brief overall comment on the feedback above:

The DD and ZX packages very much lack public documentation at the moment, as they are not exposed to Python as part of mqt-core but rather used as data structures in our top-level projects (particularly in [MQT QCEC](https://github.com/cda-tum/mqt-qcec) and [MQT DDSIM](https://github.com/cda-tum/mqt-ddsim)) that have their own documentation. Historically, we mostly documented the public-facing Python API.
However, as @josh146's confusion about the DD benchmarking section makes clear, the repository could definitely benefit from some dedicated documentation on these features.
I'll use this opportunity to create such documentation.
Once done, I will get back to the thread here.
I'll also add a brief related work section to the paper, which totally makes sense 👍🏼
