### 2024-12-06 #7478 @danielskatz
> > 👋 @1ucian0 - How is your review coming? Are there things blocking you?
>
> The compiled version of the paper in [#7478 (comment)](https://github.com/openjournals/joss-reviews/issues/7478#issuecomment-2474717495) is not up-to-date. Where can I find the new version?

You can always generate this, as I will do in the next comment
