### 2024-12-07 85d6551
⬆️🪝 Update pre-commit hook crate-ci/typos to v1.28.2
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
