### 2024-12-07 9c48337
⬆️👨‍💻 Update actions/attest-build-provenance action to v2
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
