### 2024-12-07 467c680
⬆️🪝 Update pre-commit hook rbubley/mirrors-prettier to v3.4.2 (#775)
This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
|
[rbubley/mirrors-prettier](https://redirect.github.com/rbubley/mirrors-prettier)
| repository | patch | `v3.4.1` -> `v3.4.2` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
there, instead [create a Discussion in the Renovate
repository](https://redirect.github.com/renovatebot/renovate/discussions/new)
if you have any questions.

---

##
[…]
