### 2024-12-08 725fe4f
⚗️ try the new individual Python tests workflow
Signed-off-by: burgholzer <burgholzer@me.com>
