+++ .github/workflows/ci.yml
@@ -36,7 +36,9 @@ jobs:
     name: 🐍 Test
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-python-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.5
+    uses: nquetschlich/mqt-workflows/.github/workflows/reusable-python-ci.yml@parallel_workflow
+    with:
+      run-tests-individually: true
 
   code-ql:
     name: 📝 CodeQL
