### 2024-12-08 d97ff3b
🛂 fix permissions
Signed-off-by: burgholzer <burgholzer@me.com>
