### 2024-12-13 596b285
feat: :sparkles: Add docker image for Catalyst
