+++ mlir/docker/run.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+docker run --rm -it -v $(pwd)/..:/home/mqt/mqt-mlir mqt-catalyst
