### 2024-12-14 6d2ab70
⬆️🪝 Update pre-commit hook astral-sh/ruff-pre-commit to v0.8.3
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
