### 2024-12-15 792abc6
:recycle: refractor dd operations
refractor functions to receive Operation as const referece instead of pointer

refractor getStandardOperationDD to receive targets as vector, remove overload that takes two targets and inline helper function as it now only has a single use
