+++ include/mqt-core/dd/Operations.hpp
@@ -23,6 +23,7 @@
 #include "ir/operations/Operation.hpp"
 #include "ir/operations/StandardOperation.hpp"
 
+#include <algorithm>
 #include <cassert>
 #include <cmath>
 #include <cstddef>
@@ -35,110 +36,102 @@
 
 namespace dd {
 
-inline std::pair<qc::OpType, std::vector<fp>>
-getOperationParameters(const qc::StandardOperation* op, qc::Qubit* target0,
-                       qc::Qubit* target1, bool inverse) {
-  auto type = op->getType();
-  const auto& parameter = op->getParameter();
-
-  switch (type) {
-  // self-inverse Operations
-  case qc::I:
-  case qc::H:
-  case qc::X:
-  case qc::Y:
-  case qc::Z:
-  case qc::SWAP:
-  case qc::ECR:
-    return {type, {}};
-  // inverse by using the inverse type
-  case qc::iSWAP:
-  case qc::iSWAPdg:
-  case qc::Peres:
-  case qc::Peresdg:
-  case qc::S:
-  case qc::Sdg:
-  case qc::T:
-  case qc::Tdg:
-  case qc::V:
-  case qc::Vdg:
-  case qc::SX:
-  case qc::SXdg:
-    if (inverse) {
-      type = static_cast<qc::OpType>(+type ^ qc::OpTypeInv);
-    }
-    return {type, {}};
-  // inverse by negating the parameter
-  case qc::RXX:
-  case qc::RYY:
-  case qc::RZZ:
-  case qc::RZX:
-  case qc::RX:
-  case qc::RY:
-  case qc::RZ:
-  case qc::P:
-    if (inverse) {
-      return {type, {-parameter[0U]}};
-    }
-    return {type, {parameter[0U]}};
-  // inverse by negating the first of two parameter
-  case qc::XXminusYY:
-  case qc::XXplusYY:
-    if (inverse) {
-      return {type, {-parameter[0U], parameter[1U]}};
-    }
-    return {type, {parameter[0U], parameter[1U]}};
-
-  case qc::DCX:
-    if (inverse) {
-      if (target0 == nullptr || target1 == nullptr) {
-        throw qc::QFRException("Invalid target qubits for DCX");
-      }
-
-      // DCX is not self-inverse, but the inverse is just swapping the targets
-      std::swap(*target0, *target1);
-    }
-    return {type, {}};
-
-  case qc::U:
-    if (inverse) {
-      return {type, {-parameter[1U], -parameter[2U], -parameter[0U]}};
-    }
-
-    return {type, {parameter[2U], parameter[1U], parameter[0U]}};
-  case qc::U2:
-    if (inverse) {
-      return {type, {-parameter[1U] + PI, -parameter[1U] - PI}};
-    }
-    return {type, {parameter[1U], parameter[1U]}};
-  default:
-    std::ostringstream oss{};
-    oss << "DD for gate " << op->getName() << " not available!";
-    throw qc::QFRException(oss.str());
-  }
-}
-
-// single-target Operations
 template <class Config>
-qc::MatrixDD getStandardOperationDD(const qc::StandardOperation* op,
+qc::MatrixDD getStandardOperationDD(const qc::StandardOperation& op,
                                     Package<Config>& dd,
                                     const qc::Controls& controls,
-                                    qc::Qubit target, const bool inverse) {
-  auto [type, params] = getOperationParameters(op, &target, nullptr, inverse);
+                                    const std::vector<qc::Qubit>& targets,
+                                    const bool inverse) {
 
-  return dd.makeGateDD(opToSingleGateMatrix(type, params), controls, target);
-}
+  auto type = op.getType();
[…]
