### 2024-12-15 db945c8
:recycle: retrieve DDs over standardoperations in tests
