+++ test/dd/test_package.cpp
@@ -13,11 +13,13 @@
 #include "dd/GateMatrixDefinitions.hpp"
 #include "dd/MemoryManager.hpp"
 #include "dd/Node.hpp"
+#include "dd/Operations.hpp"
 #include "dd/Package.hpp"
 #include "dd/RealNumber.hpp"
 #include "dd/statistics/PackageStatistics.hpp"
 #include "ir/operations/Control.hpp"
 #include "ir/operations/OpType.hpp"
+#include "ir/operations/StandardOperation.hpp"
 
 #include <array>
 #include <cmath>
@@ -49,8 +51,8 @@ TEST(DDPackageTest, TrivialTest) {
   auto dd = std::make_unique<dd::Package<>>(2);
   EXPECT_EQ(dd->qubits(), 2);
 
-  auto xGate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::X), 0);
-  auto hGate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::H), 0);
+  auto xGate = dd::getDD(qc::StandardOperation(0, qc::X), *dd);
+  auto hGate = dd::getDD(qc::StandardOperation(0, qc::H), *dd);
 
   ASSERT_EQ(hGate.getValueByPath(1, "0"), dd::SQRT2_2);
 
@@ -68,8 +70,8 @@ TEST(DDPackageTest, TrivialTest) {
 TEST(DDPackageTest, BellState) {
   auto dd = std::make_unique<dd::Package<>>(2);
 
-  auto hGate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::H), 1);
-  auto cxGate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::X), 1_pc, 0);
+  auto hGate = dd::getDD(qc::StandardOperation(1, qc::H), *dd);
+  auto cxGate = dd::getDD(qc::StandardOperation(1_pc, 0, qc::X), *dd);
   auto zeroState = dd->makeZeroState(2);
 
   auto bellState = dd->multiply(dd->multiply(cxGate, hGate), zeroState);
@@ -136,12 +138,12 @@ TEST(DDPackageTest, BellState) {
 TEST(DDPackageTest, QFTState) {
   auto dd = std::make_unique<dd::Package<>>(3);
 
-  auto h0Gate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::H), 0);
-  auto s0Gate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::S), 1_pc, 0);
-  auto t0Gate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::T), 2_pc, 0);
-  auto h1Gate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::H), 1);
-  auto s1Gate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::S), 2_pc, 1);
-  auto h2Gate = dd->makeGateDD(dd::opToSingleGateMatrix(qc::H), 2);
+  auto h0Gate = dd::getDD(qc::StandardOperation(0, qc::H), *dd);
+  auto s0Gate = dd::getDD(qc::StandardOperation(1_pc, 0, qc::S), *dd);
+  auto t0Gate = dd::getDD(qc::StandardOperation(2_pc, 0, qc::T), *dd);
+  auto h1Gate = dd::getDD(qc::StandardOperation(1, qc::H), *dd);
+  auto s1Gate = dd::getDD(qc::StandardOperation(2_pc, 1, qc::S), *dd);
+  auto h2Gate = dd::getDD(qc::StandardOperation(2, qc::H), *dd);
   auto swapGate = dd->makeTwoQubitGateDD(dd::opToTwoQubitGateMatrix(qc::SWAP),
[…]
