### 2024-12-15 d442811
:recycle: retrieve DDs over standardoperations in tests
