### 2024-12-15 337ba29
:recycle: move helpers in GateMatrixDefinition from .hpp to .cpp
