### 2024-12-15 cc0ccf8
:recycle: move helpers in GateMatrixDefinition from .hpp to .cpp
