+++ src/dd/GateMatrixDefinitions.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Chair for Design Automation, TUM
+ * Copyright (c) 2025 Chair for Design Automation, TUM
  * All rights reserved.
  *
  * SPDX-License-Identifier: MIT
@@ -9,59 +9,218 @@
 
 #include "dd/GateMatrixDefinitions.hpp"
 
+#include "Definitions.hpp"
 #include "dd/DDDefinitions.hpp"
 #include "ir/operations/OpType.hpp"
 
+#include <cassert>
+#include <cmath>
 #include <complex>
-#include <unordered_map>
 #include <vector>
+
+namespace {
+using namespace dd;
+
+inline GateMatrix uMat(const fp lambda, const fp phi, const fp theta) {
+  return GateMatrix{{{std::cos(theta / 2.), 0.},
+                     {-std::cos(lambda) * std::sin(theta / 2.),
+                      -std::sin(lambda) * std::sin(theta / 2.)},
+                     {std::cos(phi) * std::sin(theta / 2.),
+                      std::sin(phi) * std::sin(theta / 2.)},
+                     {std::cos(lambda + phi) * std::cos(theta / 2.),
+                      std::sin(lambda + phi) * std::cos(theta / 2.)}}};
+}
+
+inline GateMatrix u2Mat(const fp lambda, const fp phi) {
+  return GateMatrix{
+      SQRT2_2,
+      {-std::cos(lambda) * SQRT2_2, -std::sin(lambda) * SQRT2_2},
+      {std::cos(phi) * SQRT2_2, std::sin(phi) * SQRT2_2},
+      {std::cos(lambda + phi) * SQRT2_2, std::sin(lambda + phi) * SQRT2_2}};
+}
+
+inline GateMatrix pMat(const fp lambda) {
+  return GateMatrix{1, 0, 0, {std::cos(lambda), std::sin(lambda)}};
+}
+
+inline GateMatrix rxMat(const fp lambda) {
+  return GateMatrix{{{std::cos(lambda / 2.), 0.},
+                     {0., -std::sin(lambda / 2.)},
+                     {0., -std::sin(lambda / 2.)},
+                     {std::cos(lambda / 2.), 0.}}};
+}
+
+inline GateMatrix ryMat(const fp lambda) {
+  return GateMatrix{{{std::cos(lambda / 2.), 0.},
+                     {-std::sin(lambda / 2.), 0.},
+                     {std::sin(lambda / 2.), 0.},
+                     {std::cos(lambda / 2.), 0.}}};
+}
+
+inline GateMatrix rzMat(const fp lambda) {
+  return GateMatrix{{{std::cos(lambda / 2.), -std::sin(lambda / 2.)},
+                     0,
+                     0,
+                     {std::cos(lambda / 2.), std::sin(lambda / 2.)}}};
+}
+
+inline TwoQubitGateMatrix rxxMat(const fp theta) {
+  const auto cosTheta = std::cos(theta / 2.);
+  const auto sinTheta = std::sin(theta / 2.);
+
+  return TwoQubitGateMatrix{{{cosTheta, 0, 0, {0., -sinTheta}},
+                             {0, cosTheta, {0., -sinTheta}, 0},
+                             {0, {0., -sinTheta}, cosTheta, 0},
+                             {std::complex{0., -sinTheta}, 0, 0, cosTheta}}};
[…]
