### 2024-12-15 d678b71
:recycle: rename opToSingleGateMatrix to opToSingleQubitGateMatrix
