### 2024-12-17 9a4a40c
📝 update installation guide
Signed-off-by: burgholzer <burgholzer@me.com>
