+++ docs/installation.md
@@ -1,13 +1,68 @@
 # Installation
 
 MQT Core is mainly developed as a C++17 library with Python bindings.
-The resulting Python package is available on [PyPI](https://pypi.org/project/mqt.core/) and can be installed via `pip` for all major operating systems and all modern Python versions.
+The resulting Python package is available on [PyPI](https://pypi.org/project/mqt.core/) and can be installed on all major operating systems using all modern Python versions.
+
+:::::{tip}
+We highly recommend using [`uv`](https://docs.astral.sh/uv/) for working with Python projects.
+It is an extremely fast Python package and project manager, written in Rust and developed by [Astral](https://astral.sh/) (the same team behind [`ruff`](https://docs.astral.sh/ruff/)).
+It can act as a drop-in replacement for `pip` and `virtualenv`, and provides a more modern and faster alternative to the traditional Python package management tools.
+It automatically handles the creation of virtual environments and the installation of packages, and is much faster than `pip`.
+Additionally, it can even set up Python for you if it is not installed yet.
+
+If you do not have `uv` installed yet, you can install it via:
+
+::::{tab-set}
+:::{tab-item} macOS and Linux
 
 ```console
-(.venv) $ pip install mqt.core
+$ curl -LsSf https://astral.sh/uv/install.sh | sh
 ```
 
+:::
+:::{tab-item} Windows
+
+```console
+$ powershell -ExecutionPolicy ByPass -c "irm https://astral.sh/uv/install.ps1 | iex"
+```
+
+::::
+
+Check out their excellent [documentation](https://docs.astral.sh/uv/) for more information.
+
+:::::
+
+::::{tab-set}
+:sync-group: installer
+
+:::{tab-item} uv _(recommended)_
+:sync: uv
+
+```console
+$ uv pip install mqt.core
+```
+
+:::
+
+:::{tab-item} pip
+:sync: pip
+
+```console
+(.venv) $ python -m pip install mqt.core
+```
+
+:::
+::::
+
 In most practical cases (under 64-bit Linux, MacOS incl. Apple Silicon, and Windows), this requires no compilation and merely downloads and installs a platform-specific pre-built wheel.
+
+Once installed, you can check if the installation was successful by running:
+
+```console
+(.venv) $ python -c "import mqt.core; print(mqt.core.__version__)"
+```
+
+which should print the installed version of the library.
 
 :::{attention}
 As of version 2.7.0, support for Python 3.8 has been officially dropped.
@@ -19,9 +74,27 @@ Thank you for your understanding.
 
 In order to get the best performance and enable platform-specific optimizations that cannot be enabled on portable wheels, it is recommended to build the library from source via:
 
+::::{tab-set}
+:sync-group: installer
+
+:::{tab-item} uv _(recommended)_
+:sync: uv
+
+```console
+$ uv pip install mqt.core --no-binary mqt.core
+```
+
+:::
+
+:::{tab-item} pip
+:sync: pip
+
 ```console
 (.venv) $ pip install mqt.core --no-binary mqt.core
 ```
+
+:::
+::::
 
 This requires a [C++ compiler supporting C++17](https://en.wikipedia.org/wiki/List_of_compilers#C++_compilers) and a minimum [CMake](https://cmake.org/) version of 3.19.
[…]
