### 2024-12-21 4c602a0
⬆️🪝 Update pre-commit hook pre-commit/mirrors-mypy to v1.14.0
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
