### 2024-12-23 76be025
➖ simplify dependencies after `uv` update (#786)
## Description

This PR slighlty simplifies the dependencies setup after the recent `uv`
update which automatically picks newer versions for newer Python
versions. This allows simplifying the configuration for some of our
dependencies.
The pandas lower bound in the test dependencies cannot be dropped yet as
`uv` still picks an old pandas version without wheels for Python 3.13
when resolving for minimum versions.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-
[…]
