+++ pyproject.toml
@@ -37,7 +37,6 @@ dynamic = ["version"]
 [project.optional-dependencies]
 evaluation = [
   "pandas[output-formatting]>=2.1.2",
-  "pandas[output-formatting]>=2.2.3; python_version >= '3.13'",
 ]
 qiskit = [
   "qiskit[qasm3-import]>=1.0.0",
@@ -268,11 +267,8 @@ build = [
 ]
 docs = [
   "furo>=2024.8.6",
-  "ipython>=8.29; python_version >= '3.10'",
   "myst-nb>=1.1.2",
-  "myst-parser>=4; python_version >= '3.10'",
   "setuptools-scm>=8.1",
-  "sphinx>=8; python_version >= '3.10'",
   "sphinx-autoapi>=3.3.3",
   "sphinx-copybutton>=0.5.2",
   "sphinx-design>=0.6.1",
@@ -280,10 +276,7 @@ docs = [
   "sphinxcontrib-svg2pdfconverter>=1.2.3",
   "sphinxext-opengraph>=0.9.1",
   "pandas[output-formatting]>=2.1.2",
-  "pandas[output-formatting]>=2.2.3; python_version >= '3.13'",
   "qiskit[qasm3-import,visualization]>=1.0.0",
-  "scipy>=1.14.1; python_version >= '3.13'",
-  "numpy>=2.1; python_version >= '3.13'",
 ]
 test = [
   "pytest>=8.3.3",
@@ -292,8 +285,6 @@ test = [
   "pandas[output-formatting]>=2.1.2",
   "pandas[output-formatting]>=2.2.3; python_version >= '3.13'",
   "qiskit[qasm3-import]>=1.0.0",
-  "scipy>=1.14.1; python_version >= '3.13'",
-  "numpy>=2.1; python_version >= '3.13'",
 ]
 dev = [
   {include-group = "build"},
