### 2025-01-03 e615b11
➕ add `openqasm-pygments` dependency for pretty printing OpenQASM 3 output as part of the documentation
Signed-off-by: burgholzer <burgholzer@me.com>
