+++ pyproject.toml
@@ -277,6 +277,7 @@ docs = [
   "sphinxext-opengraph>=0.9.1",
   "pandas[output-formatting]>=2.1.2",
   "qiskit[qasm3-import,visualization]>=1.0.0",
+  "openqasm-pygments>=0.1.2",
 ]
 test = [
   "pytest>=8.3.3",
