### 2025-01-03 e10566b
📝 rework the quickstart guide
Signed-off-by: burgholzer <burgholzer@me.com>
