+++ docs/mqt_core_ir.md
@@ -1,0 +1,324 @@
+---
+file_format: mystnb
+kernelspec:
+  name: python3
+mystnb:
+  number_source_lines: true
+---
+
+```{code-cell} ipython3
+:tags: [remove-cell]
+%config InlineBackend.figure_formats = ['svg']
+```
+
+# MQT Core IR
+
+The central interface for working with quantum computations throughout the Munich Quantum Toolkit is the {py:class}`~mqt.core.ir.QuantumComputation` class.
+It effectively represents quantum computations as sequential lists of operation, similar to Qiskit's {py:class}`~qiskit.circuit.QuantumCircuit` class.
+
+The following will demonstrate how to work with the {py:class}`~mqt.core.ir.QuantumComputation` class in Python.
+
+:::{note}
+MQT Core is primarily designed in C++ with a thin Python wrapper.
+Historically, the C++ part of MQT Core was the focus and the Python interface was added later.
+As the standards we hold ourselves to have evolved, the Python interface is much better documented than the C++ interface.
+Contributions to the C++ documentation are welcome. See the [contribution guidelines](contributing.md) for more information.
+:::
+
+## Quickstart
+
+The following code snippet demonstrates how to construct a quantum computation for an
+instance of the Iterative Quantum Phase Estimation algorithm that aims to estimate
+the phase of a unitary operator $U=p(3\pi/8)$ using 3 bits of precision.
+
+```{code-cell} ipython3
+---
+mystnb:
+  text_lexer: 'qasm3'
+---
+from mqt.core.ir import QuantumComputation
+
+from math import pi
+
+theta = 3 * pi / 8
+precision = 3
+
+# Create an empty quantum computation
+qc = QuantumComputation()
+
+# Counting register
+qc.add_qubit_register(1, "q")
+
+# Eigenstate register
+qc.add_qubit_register(1, "psi")
+
+# Classical register for the result, the estimated phase is `0.c_2 c_1 c_0 * pi`
+qc.add_classical_register(precision, "c")
+
+# Prepare psi in the eigenstate |1>
+qc.x(1)
+
+for i in range(precision):
+  # Hadamard on the working qubit
+  qc.h(0)
+
+  # Controlled phase gate
+  qc.cp(2**(precision - i - 1) * theta, 0, 1)
+
+  # Iterative inverse QFT
+  for j in range(i):
+    qc.classic_controlled(op="p", target=0, creg=(j, 1), params=[-pi / 2**(i - j)])
+  qc.h(0)
+
+  # Measure the result
+  qc.measure(0, i)
+
+  # Reset the qubit if not finished
+  if i < precision - 1:
+    qc.reset(0)
+```
+
+The circuit class provides lots of flexibility when it comes to the kind of gates that can be applied.
+Check out the full API documentation of the {py:class}`~mqt.core.ir.QuantumComputation` class for more details.
+
+## Visualizing Circuits
+
+Circuits can be printed in a human-readable, text-based format.
+The output is to be read from top to bottom and left to right.
+Each line represents a single operation in the circuit.
+
+:::{note}
+The first and last lines have a special meaning: the first line contains the initial layout information, while the last line contains the output permutation. This is explained in more detail in the [Layout Information](#layout-information) section.
+:::
+
+```{code-cell} ipython3
+print(qc)
+```
+
+Circuits can also easily be exported to OpenQASM 3 using the {py:meth}`~mqt.core.ir.QuantumComputation.qasm3_str` method.
+
+```{code-cell} ipython3
+---
+mystnb:
+  text_lexer: 'qasm3'
+---
+print(qc.qasm3_str())
+```
+
[…]
