+++ docs/quickstart.md
@@ -1,271 +1,0 @@
----
-file_format: mystnb
-kernelspec:
-  name: python3
-mystnb:
-  number_source_lines: true
----
-
-```{code-cell} ipython3
-:tags: [remove-cell]
-%config InlineBackend.figure_formats = ['svg']
-```
-
-# Quickstart
-
-The central interface for working with quantum circuits in the Munich Quantum Toolkit is the {py:class}`~mqt.core.ir.QuantumComputation` class.
-It represents quantum circuits as a sequential list of operations.
-Operations can be directly applied to the {py:class}`~mqt.core.ir.QuantumComputation`:
-
-```{code-cell} ipython3
-from mqt.core import QuantumComputation
-
-# Build a `QuantumComputation` representing a Bell-state preparation circuit.
-nqubits = 2
-qc = QuantumComputation(nqubits)
-
-qc.h(0)  # Apply Hadamard gate on qubit 0
-qc.cx(0, 1)  # Apply a CNOT (controlled X-Gate) with control on qubit 0 and target on qubit 1
-
-# Get Circuit in OpenQASM 3.0 format
-print(qc.qasm3_str())
-```
-
-The circuit class provides a lot of flexibility as every unitary gate can be declared as a controlled gate:
-
-```{code-cell} ipython3
-from mqt.core.ir.operations import Control
-
-nqubits = 2
-qc = QuantumComputation(nqubits)
-
-# Controlled Hadamard Gate
-qc.ch(0, 1)
-
-# Negatively controlled S-gate: S-Gate on target is performed if control is in |0> state.
-qc.cs(Control(0, Control.Type.Neg), 1)
-
-print(qc.qasm3_str())
-```
-
-Providing a set of `Control` objects allows declaring any (unitary) gate as a multi-controlled gate:
-
-```{code-cell} ipython3
-nqubits = 3
-qc = QuantumComputation(nqubits)
-
-# Toffoli gate in mqt-core:
-qc.mcx({0, 1}, 2)
-
-# Control type can be individually declared
-qc.mcs({Control(0, Control.Type.Neg), Control(1, Control.Type.Pos)}, 2)
-
-print(qc.qasm3_str())
-```
-
-## Layout Information
-
-A {py:class}`~mqt.core.ir.QuantumComputation` also contains information about the mapping of algorithmic (or logical/virtual/circuit) qubits to and from device (or physical) qubits.
-These are contained in the {py:attr}`~mqt.core.ir.QuantumComputation.initial_layout` and {py:attr}`~mqt.core.ir.QuantumComputation.output_permutation` members which are instances of the {py:class}`~mqt.core.ir.Permutation` class. If no layout is given the trivial layout is assumed.
-
-When printing the OpenQASM representation of the {py:class}`~mqt.core.ir.QuantumComputation` the input and output permutations are given as comments in the first two lines of the QASM string. The format is:
-
-`// i Q_0, Q_1, ..., Q_n` ... algorithmic qubit $i$ is mapped to device qubit $Q_i$.
-
-`// o Q_0, Q_1, ..., Q_n` ... the value of algorithmic qubit $i$ (assumed to be stored in classical bit $c[i]$) is measured at device qubit $Q_i$.
-
-```{code-cell} ipython3
-nqubits = 3
-qc = QuantumComputation(nqubits)
-qc.initial_layout[2] = 0
-qc.initial_layout[0] = 1
-qc.initial_layout[1] = 2
-
-qc.output_permutation[2] = 0
-qc.output_permutation[0] = 1
-qc.output_permutation[1] = 2
-
-
[…]
