+++ include/mqt-core/ir/parsers/qasm3_parser/passes/ConstEvalPass.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Chair for Design Automation, TUM
+ * Copyright (c) 2025 Chair for Design Automation, TUM
  * All rights reserved.
  *
  * SPDX-License-Identifier: MIT
