### 2025-01-03 230cc2a
📝 add related work section
Signed-off-by: burgholzer <burgholzer@me.com>
