+++ paper/paper.bib
@@ -93,6 +93,13 @@
   doi                 = {10.1145/3505636},
 }
 
+@software{cuda_q,
+  title               = {CUDA-Q},
+  year                = {2024},
+  doi                 = {10.5281/zenodo.8092233},
+  rights              = {Apache License 2.0},
+}
+
 @article{duncanGraphtheoreticSimplificationQuantum2020,
   title               = {Graph-theoretic {{Simplification}} of {{Quantum Circuits}} with the {{ZX-calculus}}},
   author              = {Duncan, Ross and Kissinger, Aleks and Perdrix, Simon and {van de Wetering}, John},
@@ -187,6 +194,44 @@
   doi                 = {10.1109/DAC18072.2020.9218555},
 }
 
+@article{hong_tensor_2022,
+  title               = {A Tensor Network based Decision Diagram for Representation of Quantum Circuits},
+  author              = {Hong, Xin and Zhou, Xiangzhen and Li, Sanjiang and Feng, Yuan and Ying, Mingsheng},
+  doi                 = {10.1145/3514355},
+  url                 = {https://github.com/Veriqc/TDD},
+  journaltitle        = {{ACM} Transactions on Design Automation of Electronic Systems},
+  date                = {2022-11-30},
+}
+
+@article{Ittah2024,
+  title               = {Catalyst: a Python JIT compiler for auto-differentiable hybrid quantum programs},
+  author              = {David Ittah and Ali Asadi and Erick Ochoa Lopez and Sergei Mironov and Samuel Banning and Romain Moyard and Mai Jacob Peng and Josh Izaac},
+  year                = {2024},
+  journal             = {Journal of Open Source Software},
+  publisher           = {The Open Journal},
+  volume              = {9},
+  number              = {99},
+  pages               = {6720},
+  doi                 = {10.21105/joss.06720},
+  url                 = {https://doi.org/10.21105/joss.06720},
+}
+
+@inproceedings{kissingerPyZXLargeScale2020,
+  title               = {PyZX: Large Scale Automated Diagrammatic Reasoning},
+  author              = {Kissinger, Aleks and van de Wetering, John},
+  year                = {2020},
+  booktitle           = {Int'l Conf. Quantum Physics and Logic},
+  doi                 = {10.4204/EPTCS.318.14},
+}
+
+@article{niemann_qmdds_2016,
+  title               = {{QMDDs}: Efficient quantum function representation and manipulation},
+  author              = {Niemann, Philipp and Wille, Robert and Miller, David Michael and Thornton, Mitchell A. and Drechsler, Rolf},
+  doi                 = {10.1109/TCAD.2015.2459034},
+  journaltitle        = {Transactions on Computer-Aided Design of Integrated Circuits and Systems},
+  date                = {2016},
+}
+
 @inproceedings{pehamDepthoptimalSynthesisClifford2023,
   title               = {Depth-optimal synthesis of {{Clifford}} circuits with {{SAT}} solvers},
   author              = {Peham, Tom and Brandl, Nina and Kueng, Richard and Wille, Robert and Burgholzer, Lukas},
@@ -277,6 +322,24 @@
   doi                 = {10.1145/3566097.3567929},
 }
 
+@article{sivarajah_tket_2021,
+  title               = {t$|$ket$\rangle$: a retargetable compiler for {NISQ} devices},
+  shorttitle          = {t$|$ket$\rangle$},
+  author              = {Sivarajah, Seyon and Dilkes, Silas and Cowtan, Alexander and Simmons, Will and Edgington, Alec and Duncan, Ross},
+  doi                 = {10.1088/2058-9565/ab8e92},
[…]
