+++ docs/refs.bib
@@ -1,10 +1,10 @@
 @incollection{wille2023decisionDiagramsForQuantumComputing,
-	author        = {R. Wille and S. Hillmich and L. Burgholzer},
-	title         = {{Decision Diagrams for Quantum Computing}},
-	year          = {2023},
-	publisher     = {Springer},
-	booktitle     = {Design Automation of Quantum Computers},
-	editor        = {R. O. Topaloglu},
-	doi           = {10.1007/978-3-031-15699-1_1},
-	url           = {https://www.cda.cit.tum.de/files/eda/2023_springer_decision_diagrams_for_quantum_computing.pdf},
+  title               = {Decision Diagrams for Quantum Computing},
+  author              = {R. Wille and S. Hillmich and L. Burgholzer},
+  year                = {2023},
+  booktitle           = {Design Automation of Quantum Computers},
+  publisher           = {Springer},
+  doi                 = {10.1007/978-3-031-15699-1_1},
+  url                 = {https://www.cda.cit.tum.de/files/eda/2023_springer_decision_diagrams_for_quantum_computing.pdf},
+  editor              = {R. O. Topaloglu},
 }
