### 2025-01-03 5ca5391
🩹 avoid unicode characters in docs
Signed-off-by: burgholzer <burgholzer@me.com>
