+++ docs/installation.md
@@ -213,7 +213,7 @@ concurrency:
 
 jobs:
   update-mqt-core:
-    name: ⬆️ Update MQT Core
+    name: Update MQT Core
     uses: cda-tum/mqt-workflows/.github/workflows/reusable-mqt-core-update.yml@v1.5
     with:
       update-to-head: ${{ github.event.inputs.update-to-head == 'true' }}
