### 2025-01-03 #7478 @burgholzer
Happy new year everyone! 🎉

I am coming back here with an update on addressing the review feedback. Finally found some time to continue working on this.

> @edyounis
> > Installation: Does installation proceed as outlined in the documentation?
>
> Installation went well, but it would be nice in the documentation to have a "Run this command to ensure installation worked..." I just did `python -c 'import mqt'` and saw no error. I don't think this is important or required, but is always nice to see.

Improved upon the installation guide in the documentation and added a respective command for ensuring that everything worked in https://github.com/cda-tum/mqt-core/pull/789 ✅

> @edyounis
> > State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> Regarding this requirement, the paper may benefit from a related works section.
>
> @josh146
> Regarding the paper, I thought it was succinct and well written, with a clear statement of need. Based on the review checklist, it is perhaps missing a section for
>
> >State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> @1ucian0
> I second that.

I added a related work section to the JOSS paper in https://github.com/cda-tum/mqt-core/commit/262fd6dbab907cee3f72452b4e916ef4c8c42fdd ✅
Hopefully this addresses the points raised by all three reviewers.
Please let me know if you feel that the section is missing something.

> @josh146
> Separately, I was wondering about the inclusion of Tom Peham, who also looks to have contributed significantly, but not recently -- I assume that these code additions may not survive in the current version of MQT submitted to JOSS?

Tom indeed contributed significantly in the past. He is the creator of the ZX-calculus package that is part of MQT Core. I'd personally have no objections to including him in the list of authors. However, I find it challenging to define who should be an author on this submission and who should not. I'd have a tendency to keep the author list as it is, as these are the people who are still committed to further advancing MQT Core. Happy to adjust that though, if the reviewers have a different feeling. 🤷🏼‍♂️

> @josh146
> Something I struggled to understand was the notation/differentiation between the algorithmic qubit mapping, and the output measurement mapping. Maybe something that could be clarified a bit in the documentation?

I rephrased that part in the documentation as part of https://github.com/cda-tum/mqt-core/pull/789. ✅
I hope this is clearer now.

> @edyounis
> > Functionality: Have the functional claims of the software been confirmed?
>
> You mentioned a ZX-calculus module in the paper, but I could not find the relevant documentation. Can you point me in the right direction?
>
> @josh146
> I also struggled to parse/fully understand the 'Benchmarking the DD Package' part of the documentation. This is partly because I was missing the context of what you use the DD package for (as decision diagrams weren't introduced in the quickstart), as well as the motivation for benchmarking (comparing changes in MQT-core, comparing changes in MQT-core workflows, etc.).
>
> Separately, I looked though the mqt.core.dd docstrings, but couldn't find an example to showcase the decision diagram features. But perhaps I am missing something :)
>
> I have similar questions about the ZX functionality.
>
> In addition, a small workflow example demonstrating the ZX and DD usage would be beneficial.

I am still working towards this one. 🔴
Everything else should be adequately addressed.
