### 2025-01-04 13d73cf
⬆️🪝 Update pre-commit hook pre-commit/mirrors-mypy to v1.14.1 (#791)
This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
|
[pre-commit/mirrors-mypy](https://redirect.github.com/pre-commit/mirrors-mypy)
| repository | patch | `v1.14.0` -> `v1.14.1` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
there, instead [create a Discussion in the Renovate
repository](https://redirect.github.com/renovatebot/renovate/discussions/new)
if you have any questions.

---

##
[…]
