+++ .pre-commit-config.yaml
@@ -67,7 +67,7 @@ repos:
 
   # Static type checking using mypy
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.14.0
+    rev: v1.14.1
     hooks:
       - id: mypy
         files: ^(src/mqt|test/python|noxfile.py)
