### 2025-01-04 19471b3
⬆️🪝 Update pre-commit hook pre-commit/mirrors-clang-format to v19.1.6 (#793)
This PR contains the following updates:

| Package | Type | Update | Change |
|---|---|---|---|
|
[pre-commit/mirrors-clang-format](https://redirect.github.com/pre-commit/mirrors-clang-format)
| repository | patch | `v19.1.5` -> `v19.1.6` |

Note: The `pre-commit` manager in Renovate is not supported by the
`pre-commit` maintainers or community. Please do not report any problems
there, instead [create a Discussion in the Renovate
repository](https://redirect.github.com/renovatebot/renovate/discussions/new)
if you have
[…]
