+++ .pre-commit-config.yaml
@@ -92,7 +92,7 @@ repos:
 
   # Clang-format the C++ part of the code base automatically
   - repo: https://github.com/pre-commit/mirrors-clang-format
-    rev: v19.1.5
+    rev: v19.1.6
     hooks:
       - id: clang-format
         types_or: [c++, c, cuda]
