### 2025-01-04 7d92989
⬆️🪝 Update pre-commit hook astral-sh/ruff-pre-commit to v0.8.6
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
