### 2025-01-06 fdb61f0
👷🔧 do not build emulated wheels for mqt-core
This is mostly to avoid PyPI space limitations as we typically support all non-EOL Python versions (5) and a total of three emulated platforms.
This adds 15 wheels to each release on PyPI, which amounts to over 42% of all wheels (15/35).
It also considerably adds to the overall build time.

Signed-off-by: burgholzer <burgholzer@me.com>
