### 2025-01-06 99f5680
⬆️🪝 update pre-commit hooks
updates:
- [github.com/crate-ci/typos: v1.29.4 → dictgen-v0.3.1](https://github.com/crate-ci/typos/compare/v1.29.4...dictgen-v0.3.1)
