### 2025-01-08 cd94730
🚸 slightly tweak how optional qiskit dependency is handled
Signed-off-by: burgholzer <burgholzer@me.com>
