### 2025-01-08 2404c85
🚸 slightly tweak how optional qiskit dependency is handled (#797)
## Description

This PR slightly adjusts how the optional qiskit dependency is handled
in the `load` method.
At that point in the function, the user is passing in a Qiskit
`QuantumCircuit` anyway, so Qiskit will naturally be available.

An observation I had while working on this:
We currently do not enforce a particular version constraint on qiskit at
runtime.
Thus, someone might use a version of qiskit that is too old at runtime,
and will probably run into errors.
We could probably do a better job of suggesting people to instal
[…]
