### 2025-01-08 a9b7e70
♻️ refactor algorithms to use factory functions instead of inheritance
Signed-off-by: burgholzer <burgholzer@me.com>
