+++ test/algorithms/test_random_clifford.cpp
@@ -18,14 +18,14 @@
 #include <memory>
 #include <sstream>
 
-class RandomClifford : public testing::TestWithParam<std::size_t> {
+class RandomClifford : public testing::TestWithParam<qc::Qubit> {
 protected:
   void TearDown() override {}
   void SetUp() override {}
 };
 
 INSTANTIATE_TEST_SUITE_P(
-    RandomClifford, RandomClifford, testing::Range<std::size_t>(1U, 9U),
+    RandomClifford, RandomClifford, testing::Range<qc::Qubit>(1U, 9U),
     [](const testing::TestParamInfo<RandomClifford::ParamType>& inf) {
       // Generate names for test cases
       const auto nqubits = inf.param;
@@ -38,7 +38,7 @@ TEST_P(RandomClifford, simulate) {
   const auto nq = GetParam();
 
   auto dd = std::make_unique<dd::Package<>>(nq);
-  auto qc = qc::RandomCliffordCircuit(nq, nq * nq, 12345);
+  auto qc = qc::createRandomCliffordCircuit(nq, nq * nq, 12345);
   auto in = dd->makeZeroState(nq);
   ASSERT_NO_THROW({ dd::simulate(&qc, in, *dd); });
   qc.printStatistics(std::cout);
@@ -48,7 +48,7 @@ TEST_P(RandomClifford, buildFunctionality) {
   const auto nq = GetParam();
 
   auto dd = std::make_unique<dd::Package<>>(nq);
-  auto qc = qc::RandomCliffordCircuit(nq, nq * nq, 12345);
+  auto qc = qc::createRandomCliffordCircuit(nq, nq * nq, 12345);
   ASSERT_NO_THROW({ dd::buildFunctionality(&qc, *dd); });
   qc.printStatistics(std::cout);
 }
