+++ test/algorithms/test_grover.cpp
@@ -26,7 +26,7 @@
 #include <tuple>
 
 class Grover
-    : public testing::TestWithParam<std::tuple<std::size_t, std::size_t>> {
+    : public testing::TestWithParam<std::tuple<qc::Qubit, std::size_t>> {
 protected:
   void TearDown() override {
     dd->decRef(sim);
@@ -39,19 +39,26 @@ protected:
   void SetUp() override {
     std::tie(nqubits, seed) = GetParam();
     dd = std::make_unique<dd::Package<>>(nqubits + 1);
-    qc = std::make_unique<qc::Grover>(nqubits, seed);
-    qc->printStatistics(std::cout);
+    qc = qc::createGrover(nqubits, seed);
+    qc.printStatistics(std::cout);
+
+    // parse expected result from circuit name
+    const auto& name = qc.getName();
+    expected = name.substr(name.find_last_of('_') + 1);
+    targetValue = qc::BitString(expected);
   }
 
-  std::size_t nqubits = 0;
+  qc::Qubit nqubits = 0;
   std::size_t seed = 0;
   std::unique_ptr<dd::Package<>> dd;
-  std::unique_ptr<qc::Grover> qc;
+  qc::QuantumComputation qc;
   qc::VectorDD sim{};
   qc::MatrixDD func{};
+  std::string expected{};
+  qc::BitString targetValue{};
 };
 
-constexpr std::size_t GROVER_MAX_QUBITS = 15;
+constexpr qc::Qubit GROVER_MAX_QUBITS = 15;
 constexpr std::size_t GROVER_NUM_SEEDS = 5;
 constexpr dd::fp GROVER_ACCURACY = 1e-2;
 constexpr dd::fp GROVER_GOAL_PROBABILITY = 0.9;
@@ -59,7 +66,7 @@ constexpr dd::fp GROVER_GOAL_PROBABILITY = 0.9;
 INSTANTIATE_TEST_SUITE_P(
     Grover, Grover,
     testing::Combine(
-        testing::Range(static_cast<std::size_t>(2), GROVER_MAX_QUBITS + 1, 3),
+        testing::Range(static_cast<qc::Qubit>(2), GROVER_MAX_QUBITS + 1, 3),
         testing::Range(static_cast<std::size_t>(0), GROVER_NUM_SEEDS)),
     [](const testing::TestParamInfo<Grover::ParamType>& inf) {
       const auto nqubits = std::get<0>(inf.param);
@@ -76,19 +83,20 @@ INSTANTIATE_TEST_SUITE_P(
     });
 
 TEST_P(Grover, Functionality) {
-  auto x = '1' + qc->expected;
+  auto x = '1' + expected;
   std::reverse(x.begin(), x.end());
   std::replace(x.begin(), x.end(), '1', '2');
 
-  qc::QuantumComputation groverIteration(qc->getNqubits());
-  qc->oracle(groverIteration);
-  qc->diffusion(groverIteration);
+  qc::QuantumComputation groverIteration(qc.getNqubits());
+  qc::appendGroverOracle(groverIteration, targetValue);
+  qc::appendGroverDiffusion(groverIteration);
 
   const auto iteration = buildFunctionality(&groverIteration, *dd);
 
   auto e = iteration;
   dd->incRef(e);
-  for (std::size_t i = 0U; i < qc->iterations - 1U; ++i) {
+  const auto iterations = qc::computeNumberOfIterations(nqubits);
+  for (std::size_t i = 0U; i < iterations - 1U; ++i) {
     auto f = dd->multiply(iteration, e);
     dd->incRef(f);
     dd->decRef(e);
@@ -96,8 +104,8 @@ TEST_P(Grover, Functionality) {
     dd->garbageCollect();
   }
 
[…]
