+++ src/algorithms/QPE.cpp
@@ -16,155 +16,188 @@
 #include <cmath>
 #include <cstddef>
 #include <cstdint>
-#include <ostream>
+#include <limits>
 #include <random>
+#include <sstream>
 
 namespace qc {
-QPE::QPE(const std::size_t nq, const bool exact, const bool iter)
-    : precision(nq), iterative(iter) {
-  if (exact) {
-    // if an exact solution is wanted, generate a random n-bit number and
-    // convert it to an appropriate phase
-    const std::uint64_t max = 1ULL << nq;
-    auto distribution =
-        std::uniform_int_distribution<std::uint64_t>(0, max - 1);
-    std::uint64_t theta = 0;
-    while (theta == 0) {
-      theta = distribution(mt);
-    }
-    lambda = 0.;
-    for (std::size_t i = 0; i < nq; ++i) {
-      if ((theta & (1ULL << (nq - i - 1))) != 0) {
-        lambda += 1. / static_cast<double>(1ULL << i);
-      }
-    }
-  } else {
-    // if an inexact solution is wanted, generate a random n+1-bit number (that
-    // has its last bit set) and convert it to an appropriate phase
-    const std::uint64_t max = 1ULL << (nq + 1);
-    auto distribution =
-        std::uniform_int_distribution<std::uint64_t>(0, max - 1);
-    std::uint64_t theta = 0;
-    while ((theta & 1) == 0) {
-      theta = distribution(mt);
-    }
-    lambda = 0.;
-    for (std::size_t i = 0; i <= nq; ++i) {
-      if ((theta & (1ULL << (nq - i))) != 0) {
-        lambda += 1. / static_cast<double>(1ULL << i);
-      }
+
+// generate a random n-bit number and convert it to an appropriate phase
+[[nodiscard]] auto createExactPhase(const Qubit nq, std::mt19937_64& mt) -> fp {
+  const std::uint64_t max = 1ULL << nq;
+  auto distribution = std::uniform_int_distribution<std::uint64_t>(0, max - 1);
+  std::uint64_t theta = 0;
+  while (theta == 0) {
+    theta = distribution(mt);
+  }
+  fp lambda = 0.;
+  for (std::size_t i = 0; i < nq; ++i) {
+    if ((theta & (1ULL << (nq - i - 1))) != 0) {
+      lambda += 1. / static_cast<double>(1ULL << i);
     }
   }
-  createCircuit();
+  return lambda;
 }
 
-QPE::QPE(const fp l, const std::size_t prec, const bool iter)
-    : lambda(l), precision(prec), iterative(iter) {
-  createCircuit();
-}
-
-std::ostream& QPE::printStatistics(std::ostream& os) const {
-  os << "QPE Statistics:\n";
-  os << "\tn: " << nqubits + 1 << "\n";
-  os << "\tm: " << getNindividualOps() << "\n";
-  os << "\tlambda: " << lambda << "π"
-     << "\n";
-  os << "\tprecision: " << precision << "\n";
-  os << "\titerative: " << iterative << "\n";
-  os << "--------------"
-     << "\n";
-  return os;
-}
-
-void QPE::createCircuit() {
-  addQubitRegister(1, "psi");
-
-  if (iterative) {
-    addQubitRegister(1, "q");
-  } else {
-    addQubitRegister(precision, "q");
+// generate a random n+1-bit number (that has its last bit set) and convert it
+// to an appropriate phase
+[[nodiscard]] auto createInexactPhase(const Qubit nq, std::mt19937_64& mt)
+    -> fp {
[…]
