### 2025-01-08 1aa3166
🎨 code quality improvements and cleanup
- opting for references where null values are not expected or accepted.
- moving code from headers to source files.
- fixing new clang-tidy warnings
- adding a little more noexcept were appropriate

Signed-off-by: burgholzer <burgholzer@me.com>
