+++ src/ir/QuantumComputation.cpp
@@ -41,6 +41,171 @@
 
 namespace qc {
 
+namespace {
+template <class RegisterType>
+void printSortedRegisters(const RegisterMap<RegisterType>& regmap,
+                          const std::string& identifier, std::ostream& of,
+                          const bool openQASM3) {
+  // sort regs by start index
+  std::map<decltype(RegisterType::first), std::pair<std::string, RegisterType>>
+      sortedRegs{};
+  for (const auto& reg : regmap) {
+    sortedRegs.insert({reg.second.first, reg});
+  }
+
+  for (const auto& reg : sortedRegs) {
+    if (openQASM3) {
+      of << identifier << "[" << reg.second.second.second << "] "
+         << reg.second.first << ";" << std::endl;
+    } else {
+      of << identifier << " " << reg.second.first << "["
+         << reg.second.second.second << "];" << std::endl;
+    }
+  }
+}
+
+template <class RegisterType>
+void consolidateRegister(RegisterMap<RegisterType>& regs) {
+  bool finished = regs.empty();
+  while (!finished) {
+    for (const auto& qreg : regs) {
+      finished = true;
+      auto regname = qreg.first;
+      // check if lower part of register
+      if (regname.length() > 2 &&
+          regname.compare(regname.size() - 2, 2, "_l") == 0) {
+        auto lowidx = qreg.second.first;
+        auto lownum = qreg.second.second;
+        // search for higher part of register
+        auto highname = regname.substr(0, regname.size() - 1) + 'h';
+        auto it = regs.find(highname);
+        if (it != regs.end()) {
+          auto highidx = it->second.first;
+          auto highnum = it->second.second;
+          // fusion of registers possible
+          if (lowidx + lownum == highidx) {
+            finished = false;
+            auto targetname = regname.substr(0, regname.size() - 2);
+            auto targetidx = lowidx;
+            auto targetnum = lownum + highnum;
+            regs.insert({targetname, {targetidx, targetnum}});
+            regs.erase(regname);
+            regs.erase(highname);
+          }
+        }
+        break;
+      }
+    }
+  }
+}
+
+template <class RegisterType>
+void createRegisterArray(const RegisterMap<RegisterType>& regs,
+                         RegisterNames& regnames) {
+  regnames.clear();
+  std::stringstream ss;
+  // sort regs by start index
+  std::map<decltype(RegisterType::first), std::pair<std::string, RegisterType>>
+      sortedRegs{};
+  for (const auto& reg : regs) {
+    sortedRegs.insert({reg.second.first, reg});
+  }
+
+  for (const auto& reg : sortedRegs) {
+    for (decltype(RegisterType::second) i = 0; i < reg.second.second.second;
+         ++i) {
+      ss << reg.second.first << "[" << i << "]";
+      regnames.push_back(std::make_pair(reg.second.first, ss.str()));
+      ss.str(std::string());
+    }
+  }
+}
+
+/**
+ * @brief Removes a certain qubit in a register from the register map
+ * @details If this was the last qubit in the register, the register is
+ * deleted. Removals at the beginning or the end of a register just modify the
+ * existing register. Removals in the middle of a register split the register
+ * into two new registers. The new registers are named by appending "_l" and
+ * "_h" to the original register name.
+ * @param regs A collection of all the registers
+ * @param reg The name of the register containing the qubit to be removed
[…]
