+++ include/mqt-core/ir/QuantumComputation.hpp
@@ -14,40 +14,30 @@
 #include "operations/CompoundOperation.hpp"
 #include "operations/Control.hpp"
 #include "operations/Expression.hpp"
-#include "operations/NonUnitaryOperation.hpp"
 #include "operations/OpType.hpp"
-#include "operations/StandardOperation.hpp"
-#include "operations/SymbolicOperation.hpp"
 
-#include <algorithm>
-#include <array>
 #include <cstddef>
 #include <cstdint>
 #include <iostream>
 #include <map>
 #include <memory>
-#include <numeric>
 #include <optional>
 #include <random>
-#include <sstream>
 #include <string>
 #include <unordered_set>
 #include <utility>
-#include <variant>
 #include <vector>
 
 namespace qc {
-class CircuitOptimizer;
-
 class QuantumComputation {
 public:
-  using iterator = typename std::vector<std::unique_ptr<Operation>>::iterator;
+  using iterator = std::vector<std::unique_ptr<Operation>>::iterator;
   using const_iterator =
-      typename std::vector<std::unique_ptr<Operation>>::const_iterator;
+      std::vector<std::unique_ptr<Operation>>::const_iterator;
   using reverse_iterator =
-      typename std::vector<std::unique_ptr<Operation>>::reverse_iterator;
+      std::vector<std::unique_ptr<Operation>>::reverse_iterator;
   using const_reverse_iterator =
-      typename std::vector<std::unique_ptr<Operation>>::const_reverse_iterator;
+      std::vector<std::unique_ptr<Operation>>::const_reverse_iterator;
 
 protected:
   std::vector<std::unique_ptr<Operation>> ops;
@@ -69,240 +59,20 @@ protected:
 
   std::unordered_set<sym::Variable> occurringVariables;
 
-  void importOpenQASM3(std::istream& is);
-  void importReal(std::istream& is);
-  int readRealHeader(std::istream& is);
-  void readRealGateDescriptions(std::istream& is, int line);
-  void importTFC(std::istream& is);
-  int readTFCHeader(std::istream& is, std::map<std::string, Qubit>& varMap);
-  void readTFCGateDescriptions(std::istream& is, int line,
-                               std::map<std::string, Qubit>& varMap);
-  void importQC(std::istream& is);
-  int readQCHeader(std::istream& is, std::map<std::string, Qubit>& varMap);
-  void readQCGateDescriptions(std::istream& is, int line,
-                              std::map<std::string, Qubit>& varMap);
-
-  template <class RegisterType>
-  static void printSortedRegisters(const RegisterMap<RegisterType>& regmap,
-                                   const std::string& identifier,
-                                   std::ostream& of, bool openQASM3 = false) {
-    // sort regs by start index
-    std::map<decltype(RegisterType::first),
-             std::pair<std::string, RegisterType>>
-        sortedRegs{};
-    for (const auto& reg : regmap) {
-      sortedRegs.insert({reg.second.first, reg});
-    }
-
-    for (const auto& reg : sortedRegs) {
-      if (openQASM3) {
-        of << identifier << "[" << reg.second.second.second << "] "
-           << reg.second.first << ";" << std::endl;
-      } else {
-        of << identifier << " " << reg.second.first << "["
-           << reg.second.second.second << "];" << std::endl;
-      }
-    }
-  }
-  template <class RegisterType>
-  static void consolidateRegister(RegisterMap<RegisterType>& regs) {
-    bool finished = regs.empty();
-    while (!finished) {
-      for (const auto& qreg : regs) {
-        finished = true;
-        auto regname = qreg.first;
[…]
