### 2025-01-08 2d43ea9
♻️ replace pointers to QuantumComputation with references
Signed-off-by: burgholzer <burgholzer@me.com>
