+++ include/mqt-core/dd/FunctionalityConstruction.hpp
@@ -25,14 +25,14 @@ namespace dd {
 using namespace qc;
 
 template <class Config>
-MatrixDD buildFunctionality(const QuantumComputation* qc, Package<Config>& dd);
+MatrixDD buildFunctionality(const QuantumComputation& qc, Package<Config>& dd);
 
 template <class Config>
-MatrixDD buildFunctionalityRecursive(const QuantumComputation* qc,
+MatrixDD buildFunctionalityRecursive(const QuantumComputation& qc,
                                      Package<Config>& dd);
 
 template <class Config>
-bool buildFunctionalityRecursive(const QuantumComputation* qc,
+bool buildFunctionalityRecursive(const QuantumComputation& qc,
                                  std::size_t depth, std::size_t opIdx,
                                  std::stack<MatrixDD>& s,
                                  Permutation& permutation, Package<Config>& dd);
