### 2025-01-09 515023f
♻️ refactor to use references instead of pointers for operations
Replaced raw pointer usages with references across the codebase for improved clarity and type safety. This adjustment simplifies operations and ensures consistency, particularly when applying, measuring, or resetting operations in simulation and functionality logic. Corresponding tests have been updated to reflect this change.

Signed-off-by: burgholzer <burgholzer@me.com>
