+++ include/mqt-core/dd/Operations.hpp
@@ -37,13 +37,13 @@ namespace dd {
 // single-target Operations
 template <class Config>
 qc::MatrixDD
-getStandardOperationDD(const qc::StandardOperation* op, Package<Config>& dd,
+getStandardOperationDD(const qc::StandardOperation& op, Package<Config>& dd,
                        const qc::Controls& controls, const qc::Qubit target,
                        const bool inverse) {
   GateMatrix gm;
 
-  const auto type = op->getType();
-  const auto& parameter = op->getParameter();
+  const auto type = op.getType();
+  const auto& parameter = op.getParameter();
 
   switch (type) {
   case qc::I:
@@ -107,7 +107,7 @@ getStandardOperationDD(const qc::StandardOperation* op, Package<Config>& dd,
     break;
   default:
     std::ostringstream oss{};
-    oss << "DD for gate" << op->getName() << " not available!";
+    oss << "DD for gate" << op.getName() << " not available!";
     throw qc::QFRException(oss.str());
   }
   return dd.makeGateDD(gm, controls, target);
@@ -116,11 +116,11 @@ getStandardOperationDD(const qc::StandardOperation* op, Package<Config>& dd,
 // two-target Operations
 template <class Config>
 qc::MatrixDD
-getStandardOperationDD(const qc::StandardOperation* op, Package<Config>& dd,
+getStandardOperationDD(const qc::StandardOperation& op, Package<Config>& dd,
                        const qc::Controls& controls, qc::Qubit target0,
                        qc::Qubit target1, const bool inverse) {
-  const auto type = op->getType();
-  const auto& parameter = op->getParameter();
+  const auto type = op.getType();
+  const auto& parameter = op.getParameter();
 
   if (type == qc::DCX && inverse) {
     // DCX is not self-inverse, but the inverse is just swapping the targets
@@ -172,7 +172,7 @@ getStandardOperationDD(const qc::StandardOperation* op, Package<Config>& dd,
     break;
   default:
     std::ostringstream oss{};
-    oss << "DD for gate " << op->getName() << " not available!";
+    oss << "DD for gate " << op.getName() << " not available!";
     throw qc::QFRException(oss.str());
   }
 
@@ -186,17 +186,17 @@ getStandardOperationDD(const qc::StandardOperation* op, Package<Config>& dd,
 // An empty permutation marks the identity permutation.
 
 template <class Config>
-qc::MatrixDD getDD(const qc::Operation* op, Package<Config>& dd,
+qc::MatrixDD getDD(const qc::Operation& op, Package<Config>& dd,
                    const qc::Permutation& permutation = {},
                    const bool inverse = false) {
-  const auto type = op->getType();
+  const auto type = op.getType();
 
   if (type == qc::Barrier) {
     return dd.makeIdent();
   }
 
   if (type == qc::GPhase) {
-    auto phase = op->getParameter()[0U];
+    auto phase = op.getParameter()[0U];
     if (inverse) {
       phase = -phase;
     }
@@ -205,10 +205,10 @@ qc::MatrixDD getDD(const qc::Operation* op, Package<Config>& dd,
     return id;
   }
 
-  if (op->isStandardOperation()) {
-    const auto* standardOp = dynamic_cast<const qc::StandardOperation*>(op);
-    const auto& targets = permutation.apply(op->getTargets());
-    const auto& controls = permutation.apply(op->getControls());
+  if (op.isStandardOperation()) {
+    const auto& standardOp = dynamic_cast<const qc::StandardOperation&>(op);
+    const auto& targets = permutation.apply(standardOp.getTargets());
[…]

[…]
