### 2025-01-09 9ba2dbe
⚡ enable LTO by default when not deploying
Signed-off-by: burgholzer <burgholzer@me.com>
