### 2025-01-10 61abcad
♻️ make `ancillary` and `garbage` protected functions
these members had getters since a while back. there is no need to keep them as public members.

Signed-off-by: burgholzer <burgholzer@me.com>
