### 2025-01-11 0895ba2
⬆️🪝 Update pre-commit hook astral-sh/ruff-pre-commit to v0.9.1
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
