### 2025-01-11 9a0a6e8
⬆️🪝 Update pre-commit hook python-jsonschema/check-jsonschema to v0.31.0
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
