+++ include/mqt-core/dd/FunctionalityConstruction.hpp
@@ -9,17 +9,12 @@
 
 #pragma once
 
-#include "dd/Operations.hpp"
 #include "dd/Package_fwd.hpp"
 #include "ir/Permutation.hpp"
 #include "ir/QuantumComputation.hpp"
-#include "ir/operations/OpType.hpp"
 
 #include <cstddef>
-#include <memory>
-#include <ostream>
 #include <stack>
-#include <vector>
 
 namespace dd {
 using namespace qc;
