### 2025-01-12 451d064
✅ add tests for register management
Signed-off-by: burgholzer <burgholzer@me.com>
