+++ test/ir/test_qfr_functionality.cpp
@@ -1173,3 +1173,164 @@ TEST_F(QFRFunctionality, NoRegisterOnEmptyCircuit) {
   EXPECT_NO_THROW(qc.addQubitRegister(1U, "q"));
   EXPECT_EQ(qc.getQregs().size(), 2U);
 }
+
+TEST_F(QFRFunctionality, AddQubitAtFrontOfRegister) {
+  QuantumComputation qc{};
+  qc.addQubitRegister(2, "q");
+  const auto& qregs = qc.getQregs();
+  EXPECT_EQ(qregs.size(), 1U);
+  const auto& [name, reg] = *qregs.begin();
+  const auto& [start, size] = reg;
+  EXPECT_EQ(name, "q");
+  EXPECT_EQ(start, 0U);
+  EXPECT_EQ(size, 2U);
+
+  // first remove the qubit
+  const auto& [physicalIndex, outputIndex] = qc.removeQubit(0);
+  EXPECT_EQ(physicalIndex, 0U);
+  EXPECT_EQ(outputIndex, 0U);
+
+  const auto& qregsAfter = qc.getQregs();
+  EXPECT_EQ(qregsAfter.size(), 1U);
+  const auto& [nameAfter, regAfter] = *qregsAfter.begin();
+  const auto& [startAfter, sizeAfter] = regAfter;
+  EXPECT_EQ(nameAfter, "q");
+  EXPECT_EQ(startAfter, 1U);
+  EXPECT_EQ(sizeAfter, 1U);
+
+  // add the qubit back at the front
+  qc.addQubit(0, physicalIndex, outputIndex);
+  const auto& qregsAfterAdd = qc.getQregs();
+  EXPECT_EQ(qregsAfterAdd.size(), 1U);
+  const auto& [nameAfterAdd, regAfterAdd] = *qregsAfterAdd.begin();
+  const auto& [startAfterAdd, sizeAfterAdd] = regAfterAdd;
+  EXPECT_EQ(nameAfterAdd, "q");
+  EXPECT_EQ(startAfterAdd, 0U);
+  EXPECT_EQ(sizeAfterAdd, 2U);
+}
+
+TEST_F(QFRFunctionality, AddQubitAtEndOfRegister) {
+  QuantumComputation qc{};
+  qc.addQubitRegister(2, "q");
+  const auto& qregs = qc.getQregs();
+  EXPECT_EQ(qregs.size(), 1U);
+  const auto& [name, reg] = *qregs.begin();
+  const auto& [start, size] = reg;
+  EXPECT_EQ(name, "q");
+  EXPECT_EQ(start, 0U);
+  EXPECT_EQ(size, 2U);
+
+  // first remove the qubit
+  const auto& [physicalIndex, outputIndex] = qc.removeQubit(1);
+  EXPECT_EQ(physicalIndex, 1U);
+  EXPECT_EQ(outputIndex, 1U);
+
+  const auto& qregsAfter = qc.getQregs();
+  EXPECT_EQ(qregsAfter.size(), 1U);
+  const auto& [nameAfter, regAfter] = *qregsAfter.begin();
+  const auto& [startAfter, sizeAfter] = regAfter;
+  EXPECT_EQ(nameAfter, "q");
+  EXPECT_EQ(startAfter, 0U);
+  EXPECT_EQ(sizeAfter, 1U);
+
+  // add the qubit back at the end
+  qc.addQubit(1, physicalIndex, outputIndex);
+  const auto& qregsAfterAdd = qc.getQregs();
+  EXPECT_EQ(qregsAfterAdd.size(), 1U);
+  const auto& [nameAfterAdd, regAfterAdd] = *qregsAfterAdd.begin();
[…]
