### 2025-01-12 c381adb
🚸✅ add equality operator for `QuantumComputation` and respective tests
Signed-off-by: burgholzer <burgholzer@me.com>
