### 2025-01-13 ee328e7
Split StatePreparation into hpp, cpp file
