+++ src/algorithms/StatePreparation.cpp
@@ -1,0 +1,242 @@
+#include "algorithms/StatePreparation.hpp"
+
+static const double EPS = 1e-10;
+
+namespace qc {
+using Matrix = std::vector<std::vector<double>>;
+
+StatePreparation::StatePreparation(
+    const std::vector<std::complex<double>>& amplitudes) {
+
+}
+
+template <typename T> bool StatePreparation::isNormalized(std::vector<T> vec) {
+  return std::abs(1 - twoNorm(vec)) < EPS;
+}
+
+template <typename T> double StatePreparation::twoNorm(std::vector<T> vec) {
+  double norm = 0;
+  for (auto elem : vec) {
+    norm += std::norm(elem);
+  }
+  return sqrt(norm);
+}
+
+Matrix StatePreparation::kroneckerProduct(Matrix matrixA, Matrix matrixB) {
+  size_t const rowA = matrixA.size();
+  size_t const rowB = matrixB.size();
+  size_t const colA = matrixA[0].size();
+  size_t const colB = matrixB[0].size();
+  // initialize size
+  Matrix newMatrix{(rowA * rowB), std::vector<double>(colA * colB, 0)};
+  // code taken from RosettaCode slightly adapted
+  for (size_t i = 0; i < rowA; ++i) {
+    // k loops till rowB
+    for (size_t j = 0; j < colA; ++j) {
+      // j loops till colA
+      for (size_t k = 0; k < rowB; ++k) {
+        // l loops till colB
+        for (size_t l = 0; l < colB; ++l) {
+          // Each element of matrix A is
+          // multiplied by whole Matrix B
+          // resp and stored as Matrix C
+          newMatrix[i * rowB + k][j * colB + l] = matrixA[i][j] * matrixB[k][l];
+        }
+      }
+    }
+  }
+  return newMatrix;
+}
+
+Matrix StatePreparation::createIdentity(size_t size) {
+  Matrix identity{
+      std::vector<std::vector<double>>(size, std::vector<double>(size, 0))};
+  for (size_t i = 0; i < size; ++i) {
+    identity[i][i] = 1;
+  }
+  return identity;
+}
+
+std::vector<double>
+StatePreparation::matrixVectorProd(const Matrix& matrix,
+                                   std::vector<double> vector) {
+  std::vector<double> result;
+  for (const auto& matrixVec : matrix) {
+    double sum{0};
+    for (size_t i = 0; i < matrixVec.size(); ++i) {
+      sum += matrixVec[i] * vector[i];
+    }
+    result.push_back(sum);
+  }
+  return result;
+}
+
+// creates circuit that takes desired vector to zero
+qc::QuantumComputation
+StatePreparation::gatesToUncompute(std::vector<std::complex<double>> amplitudes,
+                                   size_t numQubits) {
+  qc::QuantumComputation disentangler{numQubits};
+  for (size_t i = 0; i < numQubits; ++i) {
+    // rotations to disentangle LSB
+    auto [remainingParams, thetas, phis] = rotationsToDisentangle(amplitudes);
+    amplitudes = remainingParams;
+    // perform required rotations
+    bool addLastCnot = true;
+    double const phisNorm = twoNorm(phis);
+    double const thetasNorm = twoNorm(thetas);
+    if (phisNorm != 0 && thetasNorm != 0) {
+      addLastCnot = false;
+    }
+    if (phisNorm != 0) {
+      // call multiplex with RZGate
+      qc::QuantumComputation rzMultiplexer =
+          multiplex(qc::OpType{qc::RZ}, phis, addLastCnot);
+      // append rzMultiplexer to disentangler, but it should only attach on
+      // qubits i-numQubits, thus "i" is added to the local qubit indices
+      for (auto& op : rzMultiplexer) {
+        for (auto& target : op->getTargets()) {
[…]
