### 2025-01-13 c359d2e
✨ standalone dialect
Signed-off-by: burgholzer <burgholzer@me.com>
