+++ mlir/src/Standalone/StandaloneDialect.cpp
@@ -1,0 +1,25 @@
+//===- StandaloneDialect.cpp - Standalone dialect ---------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "Standalone/StandaloneDialect.h"
+
+#include "Standalone/StandaloneOps.h"
+
+using namespace mlir;
+using namespace mlir::standalone;
+
+//===----------------------------------------------------------------------===//
+// Standalone dialect.
+//===----------------------------------------------------------------------===//
+
+void StandaloneDialect::initialize() {
+  addOperations<
+#define GET_OP_LIST
+#include "Standalone/StandaloneOps.cpp.inc"
+      >();
+}
