+++ mlir/include/dummy.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <ir/QuantumComputation.hpp>
-
-namespace mqt {
-auto dummyCircuit() -> qc::QuantumComputation;
-} // namespace mqt
