+++ mlir/src/Standalone/CMakeLists.txt
@@ -1,0 +1,12 @@
+add_mlir_dialect_library(
+  MLIRStandalone
+  StandaloneDialect.cpp
+  StandaloneOps.cpp
+  ADDITIONAL_HEADER_DIRS
+  ${PROJECT_SOURCE_DIR}/include/Standalone
+  DEPENDS
+  MLIRStandaloneOpsIncGen
+  LINK_LIBS
+  PUBLIC
+  MLIRIR
+  MLIRInferTypeOpInterface)
