+++ mlir/src/standalone-opt/CMakeLists.txt
@@ -1,0 +1,7 @@
+get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
+get_property(conversion_libs GLOBAL PROPERTY MLIR_CONVERSION_LIBS)
+set(LIBS ${dialect_libs} ${conversion_libs} MLIROptLib MLIRStandalone)
+add_llvm_executable(standalone-opt standalone-opt.cpp)
+
+llvm_update_compile_flags(standalone-opt)
+target_link_libraries(standalone-opt PRIVATE ${LIBS})
