### 2025-01-13 90eafa4
⚗️ copy and modify catalyst quantum dialect
Signed-off-by: burgholzer <burgholzer@me.com>
