+++ mlir/include/Standalone/StandaloneDialect.h
@@ -1,15 +1,0 @@
-//===- StandaloneDialect.h - Standalone dialect -----------------*- C++ -*-===//
-//
-// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
-#ifndef STANDALONE_STANDALONEDIALECT_H
-#define STANDALONE_STANDALONEDIALECT_H
-
-#include "Standalone/StandaloneOpsDialect.h.inc"
-#include "mlir/IR/Dialect.h"
-
-#endif // STANDALONE_STANDALONEDIALECT_H

[…]
