+++ mlir/test/lit.cfg.py
@@ -1,56 +1,65 @@
-# -*- Python -*-
-
+# pylint: disable=undefined-variable
 import os
-import platform
-import re
-import subprocess
-import tempfile
 
 import lit.formats
-import lit.util
-
 from lit.llvm import llvm_config
-from lit.llvm.subst import ToolSubst
-from lit.llvm.subst import FindTool
 
-# Configuration file for the 'lit' test runner.
+config.name = "Compiler test suite"
+config.test_format = lit.formats.ShTest(True)
 
-# name: The name of this test suite.
-config.name = 'QUANTUM'
+# Define the file extensions to treat as test files.
+config.suffixes = [".mlir", ".py"]
+config.excludes = ["lit.cfg.py"]
 
-config.test_format = lit.formats.ShTest(not llvm_config.use_lit_shell)
-
-# suffixes: A list of file extensions to treat as test files.
-config.suffixes = ['.mlir']
-
-# test_source_root: The root path where tests are located.
+# Define the root path of where to look for tests.
 config.test_source_root = os.path.dirname(__file__)
 
-# test_exec_root: The root path where tests should be run.
-config.test_exec_root = os.path.join(config.quantum_obj_root, 'test')
+# Define where to execute tests (and produce the output).
+config.test_exec_root = getattr(config, "quantum_test_dir", ".lit")
 
-config.substitutions.append(('%PATH%', config.environment['PATH']))
+# Define substitutions used at the top of lit test files, e.g. %PYTHON.
+python_executable = getattr(config, "python_executable", "python3.10")
 
-llvm_config.with_system_environment(
-    ['HOME', 'INCLUDE', 'LIB', 'TMP', 'TEMP'])
+if "Address" in getattr(config, "llvm_use_sanitizer", ""):
+    # With sanitized builds, Python tests require some preloading magic to run.
+    if "Linux" in config.host_os:
+        python_executable = f"LD_PRELOAD=$({config.host_cxx} -print-file-name=libclang_rt.asan-{config.host_arch}.so) {python_executable}"
+    elif "Darwin" in config.host_os:
+        # It's important that the python executable discovered by CMake is the true interpreter
+        # binary. On macOS, it appears that there can be multiple "python" executables which are
+        # just wrappers around the actual interpreter. In that case, the DYLD_INSERT_LIBRARIES trick
+        # will not work, as it will not be forwarded to the final process.
+        # See https://stackoverflow.com/a/47853433 for more information.
+        python_executable = f"DYLD_INSERT_LIBRARIES=$({config.host_cxx} -print-file-name=libclang_rt.asan_osx_dynamic.dylib) {python_executable}"
+    else:
+        assert False, "Testing with sanitized builds requires Linux or MacOS"
 
-llvm_config.use_default_substitutions()
+config.substitutions.append(("%PYTHON", python_executable))
 
-# excludes: A list of directories to exclude from the testsuite. The 'Inputs'
-# subdirectories contain auxiliary inputs for various tests in their parent
-# directories.
-config.excludes = ['Inputs', 'Examples', 'CMakeLists.txt', 'README.txt', 'LICENSE.txt']
+# Define PATH to include the various tools needed for our tests.
+try:
+    # From within a build target we have access to cmake variables configured in lit.site.cfg.py.in.
+    llvm_config.with_environment("PATH", config.llvm_tools_dir, append_path=True)  # FileCheck
+    llvm_config.with_environment("PATH", config.quantum_bin_dir, append_path=True)  # quantum-opt
+except AttributeError as e:
+    # The system PATH is available by default.
+    pass
 
-# test_exec_root: The root path where tests should be run.
[…]
