### 2025-01-13 69cc222
✅ simplify testing setup
Signed-off-by: burgholzer <burgholzer@me.com>
