+++ mlir/test/lit.site.cfg.py.in
@@ -1,12 +1,9 @@
 @LIT_SITE_CFG_IN_HEADER@
 
 config.llvm_tools_dir = lit_config.substitute("@LLVM_TOOLS_DIR@")
-config.lit_tools_dir = "@LLVM_LIT_TOOLS_DIR@"
-config.mlir_obj_dir = "@MLIR_BINARY_DIR@"
-config.python_executable = "@Python3_EXECUTABLE@"
-config.enable_bindings_python = "@MLIR_ENABLE_BINDINGS_PYTHON@"
-config.standalone_src_root = "@CMAKE_SOURCE_DIR@"
-config.standalone_obj_root = "@CMAKE_BINARY_DIR@"
+config.quantum_build_dir = "@CMAKE_BINARY_DIR@"
+config.quantum_test_dir = "@CMAKE_BINARY_DIR@" + "/test"
+config.quantum_bin_dir = "@CMAKE_BINARY_DIR@" + "/tools/quantum-opt"
 
 import lit.llvm
 lit.llvm.initialize(lit_config, config)
