+++ mlir/include/mlir/Dialect/MQT/CMakeLists.txt
@@ -1,0 +1,12 @@
+add_mlir_dialect(QuantumOps quantum)
+add_mlir_interface(QuantumInterfaces)
+add_mlir_doc(QuantumDialect -gen-dialect-doc QuantumDialect Quantum/)
+add_mlir_doc(QuantumOps -gen-op-doc QuantumOps Quantum/)
+add_mlir_doc(QuantumInterfaces -gen-op-interface-docs QuantumInterfaces Quantum/)
+
+set(LLVM_TARGET_DEFINITIONS QuantumOps.td)
+mlir_tablegen(QuantumEnums.h.inc -gen-enum-decls)
+mlir_tablegen(QuantumEnums.cpp.inc -gen-enum-defs)
+mlir_tablegen(QuantumAttributes.h.inc -gen-attrdef-decls -attrdefs-dialect=quantum)
+mlir_tablegen(QuantumAttributes.cpp.inc -gen-attrdef-defs -attrdefs-dialect=quantum)
+add_public_tablegen_target(MLIRQuantumEnumsIncGen)
