+++ mlir/test/Dialect/MQT/lit.site.cfg.py.in
@@ -1,0 +1,12 @@
+@LIT_SITE_CFG_IN_HEADER@
+
+config.llvm_tools_dir = lit_config.substitute("@LLVM_TOOLS_DIR@")
+config.quantum_build_dir = "@CMAKE_BINARY_DIR@"
+config.quantum_test_dir = "@CMAKE_BINARY_DIR@" + "/mlir/test"
+config.quantum_bin_dir = "@CMAKE_BINARY_DIR@" + "/mlir/tools/quantum-opt"
+
+import lit.llvm
+lit.llvm.initialize(lit_config, config)
+
+# Let the main config do the real work.
+lit_config.load_config(config, "@CMAKE_SOURCE_DIR@/mlir/test/Dialect/MQT/lit.cfg.py")
