+++ mlir/src/CatalystQuantum/QuantumOps.cpp
@@ -1,392 +1,0 @@
-// Copyright 2022-2023 Xanadu Quantum Technologies Inc.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-
-//     http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "CatalystQuantum/QuantumOps.h"
-
-#include "CatalystQuantum/QuantumDialect.h"
-#include "mlir/Dialect/Arith/IR/Arith.h"
-#include "mlir/IR/Builders.h"
-#include "mlir/IR/OpImplementation.h"
-
-#include "llvm/ADT/StringSet.h"
-#include "llvm/ADT/TypeSwitch.h"
-
-#include <optional>
-
-using namespace mlir;
-using namespace catalyst::quantum;
-
-//===----------------------------------------------------------------------===//
-// Quantum op definitions.
-//===----------------------------------------------------------------------===//
-
-#include "CatalystQuantum/QuantumEnums.cpp.inc"
-#define GET_OP_CLASSES
-#include "CatalystQuantum/QuantumOps.cpp.inc"
-
-//===----------------------------------------------------------------------===//
-// Quantum op canonicalizers.
-//===----------------------------------------------------------------------===//
-static const mlir::StringSet<> hermitianOps = {"Hadamard", "PauliX", "PauliY",
-                                               "PauliZ",   "CNOT",   "CY",
-                                               "CZ",       "SWAP",   "Toffoli"};
-static const mlir::StringSet<> rotationsOps = {
-    "RX",  "RY",  "RZ",  "PhaseShift",
-    "CRX", "CRY", "CRZ", "ControlledPhaseShift"};
-
-LogicalResult StaticCustomOp::canonicalize(StaticCustomOp op,
-                                           mlir::PatternRewriter& rewriter) {
-  if (!op.getAdjoint()) {
-    return failure();
-  }
-  auto name = op.getGateName();
-
-  if (hermitianOps.contains(name)) {
-    rewriter.modifyOpInPlace(op, [&op]() { op.setAdjoint(false); });
-    return success();
-  }
-
-  if (rotationsOps.contains(name)) {
-    auto params = op.getStaticParams();
-    SmallVector<double> paramsNeg;
-    for (auto param : params) {
-      auto paramNeg = -1 * param;
-      paramsNeg.push_back(paramNeg);
-    }
-
-    rewriter.replaceOpWithNewOp<StaticCustomOp>(
-        op, op.getOutQubits().getTypes(), op.getOutCtrlQubits().getTypes(),
-        rewriter.getDenseF64ArrayAttr(paramsNeg), op.getInQubits(), name,
-        nullptr, op.getInCtrlQubits(), op.getInCtrlValues());
-
-    return success();
-  }
-
-  return failure();
-}
-
-LogicalResult CustomOp::canonicalize(CustomOp op,
-                                     mlir::PatternRewriter& rewriter) {
-  if (!op.getAdjoint()) {
-    return failure();
-  }
-  auto name = op.getGateName();
-
-  if (hermitianOps.contains(name)) {
-    rewriter.modifyOpInPlace(op, [&op]() { op.setAdjoint(false); });
-    return success();
-  }
-
-  if (rotationsOps.contains(name)) {
[…]
