+++ mlir/mlir/tools/quantum-opt/quantum-opt.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "CatalystQuantum/QuantumDialect.h"
+#include "mlir/Dialect/MQT/QuantumDialect.h"
 #include "mlir/Dialect/Func/Extensions/AllExtensions.h"
 #include "mlir/IR/DialectRegistry.h"
 #include "mlir/InitAllDialects.h"
