+++ mlir/src/CatalystQuantum/CMakeLists.txt
@@ -1,11 +1,0 @@
-add_mlir_library(
-  MLIRQuantum
-  QuantumDialect.cpp
-  QuantumInterfaces.cpp
-  QuantumOps.cpp
-  ADDITIONAL_HEADER_DIRS
-  ${MQT_MLIR_INCLUDE_BUILD_DIR}/CatalystQuantum
-  DEPENDS
-  MLIRQuantumInterfacesIncGen
-  MLIRQuantumOpsIncGen
-  MLIRQuantumEnumsIncGen)
