### 2025-01-14 2101718
fix: :bug: Remove lib folder from gitignore so that lib dir can be added
