+++ mlir/include/mlir/Dialect/MQT/QuantumOps.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "mlir/Dialect/MQT/QuantumInterfaces.h"
+#include "mlir/Dialect/MQT/MQTInterfaces.h"
 #include "mlir/Dialect/Bufferization/IR/AllocationOpInterface.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/Dialect.h"
@@ -30,7 +30,7 @@
 #include <optional>
 
 //===----------------------------------------------------------------------===//
-// Quantum trait declarations.
+// MQT trait declarations.
 //===----------------------------------------------------------------------===//
 
 namespace mlir {
@@ -50,12 +50,12 @@ class QuantumMemory : public mlir::SideEffects::Resource::Base<QuantumMemory> {
 };
 
 //===----------------------------------------------------------------------===//
-// Quantum ops declarations.
+// MQT ops declarations.
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Dialect/MQT/QuantumDialect.h"
-#include "mlir/Dialect/MQT/QuantumEnums.h.inc"
+#include "mlir/Dialect/MQT/MQTDialect.h"
+#include "mlir/Dialect/MQT/MQTEnums.h.inc"
 #define GET_ATTRDEF_CLASSES
-#include "mlir/Dialect/MQT/QuantumAttributes.h.inc"
+#include "mlir/Dialect/MQT/MQTAttributes.h.inc"
 #define GET_OP_CLASSES
-#include "mlir/Dialect/MQT/QuantumOps.h.inc"
+#include "mlir/Dialect/MQT/MQTOps.h.inc"
