+++ mlir/mlir/include/mlir/Dialect/MQT/CMakeLists.txt
@@ -1,12 +1,12 @@
-add_mlir_dialect(QuantumOps quantum)
-add_mlir_interface(QuantumInterfaces)
-add_mlir_doc(QuantumDialect -gen-dialect-doc QuantumDialect Quantum/)
-add_mlir_doc(QuantumOps -gen-op-doc QuantumOps Quantum/)
-add_mlir_doc(QuantumInterfaces -gen-op-interface-docs QuantumInterfaces Quantum/)
+add_mlir_dialect(MQTOps quantum)
+add_mlir_interface(MQTInterfaces)
+add_mlir_doc(MQTDialect -gen-dialect-doc MQTDialect MQT/)
+add_mlir_doc(MQTOps -gen-op-doc MQTOps MQT/)
+add_mlir_doc(MQTInterfaces -gen-op-interface-docs MQTInterfaces MQT/)
 
-set(LLVM_TARGET_DEFINITIONS QuantumOps.td)
-mlir_tablegen(QuantumEnums.h.inc -gen-enum-decls)
-mlir_tablegen(QuantumEnums.cpp.inc -gen-enum-defs)
-mlir_tablegen(QuantumAttributes.h.inc -gen-attrdef-decls -attrdefs-dialect=quantum)
-mlir_tablegen(QuantumAttributes.cpp.inc -gen-attrdef-defs -attrdefs-dialect=quantum)
-add_public_tablegen_target(MLIRQuantumEnumsIncGen)
+set(LLVM_TARGET_DEFINITIONS MQTOps.td)
+mlir_tablegen(MQTEnums.h.inc -gen-enum-decls)
+mlir_tablegen(MQTEnums.cpp.inc -gen-enum-defs)
+mlir_tablegen(MQTAttributes.h.inc -gen-attrdef-decls -attrdefs-dialect=quantum)
+mlir_tablegen(MQTAttributes.cpp.inc -gen-attrdef-defs -attrdefs-dialect=quantum)
+add_public_tablegen_target(MLIRMQTEnumsIncGen)
