+++ mlir/mlir/include/mlir/Dialect/MQT/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_mlir_dialect(MQTOps quantum)
+add_mlir_dialect(MQTOps mqt)
 add_mlir_interface(MQTInterfaces)
 add_mlir_doc(MQTDialect -gen-dialect-doc MQTDialect MQT/)
 add_mlir_doc(MQTOps -gen-op-doc MQTOps MQT/)
@@ -7,6 +7,6 @@ add_mlir_doc(MQTInterfaces -gen-op-interface-docs MQTInterfaces MQT/)
 set(LLVM_TARGET_DEFINITIONS MQTOps.td)
 mlir_tablegen(MQTEnums.h.inc -gen-enum-decls)
 mlir_tablegen(MQTEnums.cpp.inc -gen-enum-defs)
-mlir_tablegen(MQTAttributes.h.inc -gen-attrdef-decls -attrdefs-dialect=quantum)
-mlir_tablegen(MQTAttributes.cpp.inc -gen-attrdef-defs -attrdefs-dialect=quantum)
+mlir_tablegen(MQTAttributes.h.inc -gen-attrdef-decls -attrdefs-dialect=mqt)
+mlir_tablegen(MQTAttributes.cpp.inc -gen-attrdef-defs -attrdefs-dialect=mqt)
 add_public_tablegen_target(MLIRMQTEnumsIncGen)
