### 2025-01-14 84b5041
Refactored StatePreparation files to match the other algorithms
