+++ src/algorithms/StatePreparation.cpp
@@ -6,22 +6,52 @@
  *
  * Licensed under the MIT License
  */
-
 #include "algorithms/StatePreparation.hpp"
+
+#include "CircuitOptimizer.hpp"
+#include "ir/operations/StandardOperation.hpp"
+
+#include <cmath>
+#include <complex>
+#include <utility>
 
 static const double EPS = 1e-10;
 
 namespace qc {
 using Matrix = std::vector<std::vector<double>>;
 
-StatePreparation::StatePreparation(
-    const std::vector<std::complex<double>>& amplitudes) {}
+auto createStatePreparationCircuit(
+    const std::vector<std::complex<double>>& amplitudes) -> QuantumComputation {
 
-template <typename T> bool StatePreparation::isNormalized(std::vector<T> vec) {
+  if (!isNormalized(amplitudes)) {
+    throw std::invalid_argument{
+        "Using State Preparation with Amplitudes that are not normalized"};
+  }
+
+  // get number of qubits needed
+  double const numQubits = std::log2(amplitudes.size());
+
+  if (numQubits == 0 || std::floor(numQubits) != numQubits) {
+    throw std::invalid_argument{
+        "Using State Preparation with vector size that is not a power of 2"};
+  }
+
+  QuantumComputation toZeroCircuit =
+      gatesToUncompute(amplitudes, static_cast<size_t>(numQubits));
+
+  // invert circuit
+  CircuitOptimizer::flattenOperations(toZeroCircuit);
+  toZeroCircuit.invert();
+
+  return toZeroCircuit;
+}
+
+template <typename T>
+[[noexcept]] auto isNormalized(std::vector<T> vec) -> bool {
   return std::abs(1 - twoNorm(vec)) < EPS;
 }
 
-template <typename T> double StatePreparation::twoNorm(std::vector<T> vec) {
+template <typename T>[[noexcept]] auto twoNorm(std::vector<T> vec) -> double {
   double norm = 0;
   for (auto elem : vec) {
     norm += std::norm(elem);
@@ -29,7 +59,7 @@ template <typename T> double StatePreparation::twoNorm(std::vector<T> vec) {
   return sqrt(norm);
 }
 
-Matrix StatePreparation::kroneckerProduct(Matrix matrixA, Matrix matrixB) {
+[[noexcept]] auto kroneckerProduct(Matrix matrixA, Matrix matrixB) -> Matrix {
   size_t const rowA = matrixA.size();
   size_t const rowB = matrixB.size();
   size_t const colA = matrixA[0].size();
@@ -55,7 +85,7 @@ Matrix StatePreparation::kroneckerProduct(Matrix matrixA, Matrix matrixB) {
   return newMatrix;
 }
 
-Matrix StatePreparation::createIdentity(size_t size) {
+[[noexcept]] auto createIdentity(size_t size) -> Matrix {
   Matrix identity{
       std::vector<std::vector<double>>(size, std::vector<double>(size, 0))};
   for (size_t i = 0; i < size; ++i) {
@@ -64,9 +94,9 @@ Matrix StatePreparation::createIdentity(size_t size) {
   return identity;
 }
 
-std::vector<double>
-StatePreparation::matrixVectorProd(const Matrix& matrix,
-                                   std::vector<double> vector) {
+[[noexcept]] auto
+matrixVectorProd(const Matrix& matrix,
+                 std::vector<double> vector) -> std::vector<double> {
   std::vector<double> result;
   for (const auto& matrixVec : matrix) {
     double sum{0};
@@ -79,10 +109,9 @@ StatePreparation::matrixVectorProd(const Matrix& matrix,
 }
 
 // creates circuit that takes desired vector to zero
-qc::QuantumComputation
[…]
