+++ include/mqt-core/StatePreparation.hpp
@@ -1,299 +1,0 @@
-/*
- * Copyright (c) 2025 Chair for Design Automation, TUM
- * All rights reserved.
- *
- * SPDX-License-Identifier: MIT
- *
- * Licensed under the MIT License
- */
-
-#pragma once
-
-#include "CircuitOptimizer.hpp"
-#include "QuantumComputation.hpp"
-#include "ir/operations/StandardOperation.hpp"
-
-#include <cmath>
-#include <complex>
-#include <utility>
-
-/**
- * Class to prepare a generic Quantum State from a list of normalized complex
- amplitudes
- * Adapted implementation of Qiskit State Preparation:
- *
- https://github.com/Qiskit/qiskit/blob/main/qiskit/circuit/library/data_preparation/state_preparation.py#
- * based on paper:
- *      Shende, Bullock, Markov. Synthesis of Quantum Logic Circuits (2004)
-        [`https://ieeexplore.ieee.org/document/1629135`]
- * */
-static const double EPS = 1e-10;
-
-class StatePreparation {
-  using Matrix = std::vector<std::vector<double>>;
-
-public:
-  /**
-   * @throws invalid_argument when amplitudes are not normalized or length not
-   * power of 2
-   * @param list of complex amplitudes to initialize to
-   * @return MQT Circuit that initializes a state
-   * */
-  static qc::QuantumComputation
-  prepareState(const std::vector<std::complex<double>>& amplitudes) {
-
-    if (!isNormalized(amplitudes)) {
-      throw std::invalid_argument{
-          "Using State Preparation with Amplitudes that are not normalized"};
-    }
-
-    // get number of qubits needed
-    double const numQubits = std::log2(amplitudes.size());
-
-    if (numQubits == 0 || std::floor(numQubits) != numQubits) {
-      throw std::invalid_argument{
-          "Using State Preparation with vector size that is not a power of 2"};
-    }
-
-    qc::QuantumComputation toZeroCircuit =
-        gatesToUncompute(amplitudes, static_cast<size_t>(numQubits));
-
-    // invert circuit
-    qc::CircuitOptimizer::flattenOperations(toZeroCircuit);
-    toZeroCircuit.invert();
-
-    return toZeroCircuit;
-  }
-
-private:
-  // check whether vector is normalized
-  template <typename T> static bool isNormalized(std::vector<T> vec) {
-    return std::abs(1 - twoNorm(vec)) < EPS;
-  }
-
-  template <typename T> static double twoNorm(std::vector<T> vec) {
-    double norm = 0;
-    for (auto elem : vec) {
-      norm += std::norm(elem);
-    }
-    return sqrt(norm);
-  }
-
-  // kronecker product of two matrices with form vector<vector> where inner
-  // vectors represent rows
-  static Matrix kroneckerProduct(Matrix matrixA, Matrix matrixB) {
-    size_t const rowA = matrixA.size();
-    size_t const rowB = matrixB.size();
-    size_t const colA = matrixA[0].size();
-    size_t const colB = matrixB[0].size();
-    // initialize size
-    Matrix newMatrix{(rowA * rowB), std::vector<double>(colA * colB, 0)};
-    // code taken from RosettaCode slightly adapted
-    for (size_t i = 0; i < rowA; ++i) {
-      // k loops till rowB
-      for (size_t j = 0; j < colA; ++j) {
-        // j loops till colA
-        for (size_t k = 0; k < rowB; ++k) {
-          // l loops till colB
-          for (size_t l = 0; l < colB; ++l) {
-            // Each element of matrix A is
-            // multiplied by whole Matrix B
-            // resp and stored as Matrix C
[…]
