### 2025-01-14 4918724
Compiler error and linting
