+++ src/algorithms/StatePreparation.cpp
@@ -10,11 +10,14 @@
 #include "algorithms/StatePreparation.hpp"
 
 #include "CircuitOptimizer.hpp"
+#include "ir/operations/Operation.hpp"
 #include "ir/operations/StandardOperation.hpp"
+#include "ir/operations/OpType.hpp"
 
 #include <cmath>
 #include <complex>
-#include <utility>
+#include <tuple>
+#include <stddef.h>
 
 static const double EPS = 1e-10;
 
@@ -22,7 +25,7 @@ namespace qc {
 using Matrix = std::vector<std::vector<double>>;
 
 auto createStatePreparationCircuit(
-    const std::vector<std::complex<double>>& amplitudes) -> QuantumComputation {
+    std::vector<std::complex<double>>& amplitudes) -> QuantumComputation {
 
   if (!isNormalized(amplitudes)) {
     throw std::invalid_argument{
@@ -95,9 +98,9 @@ template <typename T>[[noexcept]] auto twoNorm(std::vector<T> vec) -> double {
   return identity;
 }
 
-[[noexcept]] auto matrixVectorProd(const Matrix& matrix,
-                                   std::vector<double> vector)
-    -> std::vector<double> {
+[[noexcept]] auto
+matrixVectorProd(const Matrix& matrix,
+                 std::vector<double> vector) -> std::vector<double> {
   std::vector<double> result;
   for (const auto& matrixVec : matrix) {
     double sum{0};
