### 2025-01-14 1a72f17
Fixed all build issues and linting problems
