+++ src/algorithms/StatePreparation.cpp
@@ -9,53 +9,34 @@
 
 #include "algorithms/StatePreparation.hpp"
 
-#include "CircuitOptimizer.hpp"
+#include "Definitions.hpp"
+#include "circuit_optimizer/CircuitOptimizer.hpp"
+#include "ir/QuantumComputation.hpp"
+#include "ir/operations/OpType.hpp"
 #include "ir/operations/Operation.hpp"
 #include "ir/operations/StandardOperation.hpp"
-#include "ir/operations/OpType.hpp"
 
+#include <algorithm>
 #include <cmath>
 #include <complex>
+#include <cstddef>
+#include <cstdint>
+#include <iterator>
+#include <numeric>
 #include <tuple>
-#include <stddef.h>
+#include <vector>
 
 static const double EPS = 1e-10;
 
 namespace qc {
 using Matrix = std::vector<std::vector<double>>;
 
-auto createStatePreparationCircuit(
-    std::vector<std::complex<double>>& amplitudes) -> QuantumComputation {
-
-  if (!isNormalized(amplitudes)) {
-    throw std::invalid_argument{
-        "Using State Preparation with Amplitudes that are not normalized"};
-  }
-
-  // get number of qubits needed
-  double const numQubits = std::log2(amplitudes.size());
-
-  if (numQubits == 0 || std::floor(numQubits) != numQubits) {
-    throw std::invalid_argument{
-        "Using State Preparation with vector size that is not a power of 2"};
-  }
-
-  QuantumComputation toZeroCircuit =
-      gatesToUncompute(amplitudes, static_cast<size_t>(numQubits));
-
-  // invert circuit
-  CircuitOptimizer::flattenOperations(toZeroCircuit);
-  toZeroCircuit.invert();
-
-  return toZeroCircuit;
-}
-
 template <typename T>
-[[noexcept]] auto isNormalized(std::vector<T> vec) -> bool {
+[[nodiscard]] auto isNormalized(std::vector<T> vec) -> bool {
   return std::abs(1 - twoNorm(vec)) < EPS;
 }
 
-template <typename T>[[noexcept]] auto twoNorm(std::vector<T> vec) -> double {
+template <typename T> [[nodiscard]] auto twoNorm(std::vector<T> vec) -> double {
   double norm = 0;
   for (auto elem : vec) {
     norm += std::norm(elem);
@@ -63,7 +44,7 @@ template <typename T>[[noexcept]] auto twoNorm(std::vector<T> vec) -> double {
   return sqrt(norm);
 }
 
-[[noexcept]] auto kroneckerProduct(Matrix matrixA, Matrix matrixB) -> Matrix {
+[[nodiscard]] auto kroneckerProduct(Matrix matrixA, Matrix matrixB) -> Matrix {
   size_t const rowA = matrixA.size();
   size_t const rowB = matrixB.size();
   size_t const colA = matrixA[0].size();
@@ -89,7 +70,7 @@ template <typename T>[[noexcept]] auto twoNorm(std::vector<T> vec) -> double {
   return newMatrix;
 }
 
-[[noexcept]] auto createIdentity(size_t size) -> Matrix {
+[[nodiscard]] auto createIdentity(size_t size) -> Matrix {
   Matrix identity{
       std::vector<std::vector<double>>(size, std::vector<double>(size, 0))};
   for (size_t i = 0; i < size; ++i) {
@@ -98,7 +79,7 @@ template <typename T>[[noexcept]] auto twoNorm(std::vector<T> vec) -> double {
   return identity;
 }
 
-[[noexcept]] auto
+[[nodiscard]] auto
 matrixVectorProd(const Matrix& matrix,
                  std::vector<double> vector) -> std::vector<double> {
[…]
