### 2025-01-14 fc06910
Made zero check more robust
