+++ mlir/lib/Dialect/MQT/Transforms/ThePass.cpp
@@ -1,0 +1,29 @@
+#include "mlir/Dialect/MQT/Transforms/Passes.h"
+
+#include <mlir/IR/BuiltinOps.h>
+#include <mlir/IR/MLIRContext.h>
+#include <mlir/IR/PatternMatch.h>
+#include <mlir/Pass/Pass.h>
+#include <mlir/Support/LLVM.h>
+#include <mlir/Transforms/GreedyPatternRewriteDriver.h>
+#include <utility>
+
+namespace mlir::mqt {
+
+struct ThePass : PassWrapper<ThePass, OperationPass<ModuleOp>> {
+  void runOnOperation() override {
+    // Get the current operation being operated on.
+    const ModuleOp op = getOperation();
+    MLIRContext* ctx = &getContext();
+
+    // Define the set of patterns to use.
+    RewritePatternSet thePatterns(ctx);
+    populateThePassPatterns(thePatterns);
+
+    // Apply patterns in an iterative and greedy manner.
+    if (failed(applyPatternsAndFoldGreedily(op, std::move(thePatterns)))) {
+      signalPassFailure();
+    }
+  }
+};
+} // namespace mlir::mqt
