+++ mlir/mlir/include/mlir/Dialect/MQT/Transforms/Passes.h
@@ -6,7 +6,7 @@ namespace mlir {
 
 class RewritePatternSet;
 
-// namespace mqt {
+namespace mqt {
 
 #define GEN_PASS_DECL
 #include "mlir/Dialect/MQT/Transforms/Passes.h.inc"
@@ -21,5 +21,5 @@ void populateThePassPatterns(RewritePatternSet& patterns);
 #define GEN_PASS_REGISTRATION
 #include "mlir/Dialect/MQT/Transforms/Passes.h.inc"
 
-// } // namespace mqt
+} // namespace mqt
 } // namespace mlir
