### 2025-01-15 6b3a521
fix: :bug: Resolve issue that deleted program with ThePass
The greedy rewriter sees all instructions in the code as "dead" and deletes them. By returning the final result as an output for the program, this behaviour can be prevented. Furthermore, handled Hadamards must be marked so that future ThePattern calls will fail, otherwise the whole rewrite process fails
