+++ mlir/mlir/lib/Dialect/MQT/Transforms/ThePattern.cpp
@@ -5,27 +5,37 @@
 #include <mlir/IR/MLIRContext.h>
 #include <mlir/IR/PatternMatch.h>
 #include <mlir/Support/LogicalResult.h>
+#include <set>
 
 namespace mlir::mqt {
 /// Multi-step rewrite using "match" and "rewrite". This allows for separating
 /// the concerns of matching and rewriting.
 struct ThePattern final : OpRewritePattern<CustomOp> {
-  explicit ThePattern(MLIRContext* context) : OpRewritePattern(context) {}
+  DenseSet<Operation*>& handled;
+
+  explicit ThePattern(MLIRContext* context,
+                      DenseSet<Operation*>& handledOperations)
+      : OpRewritePattern(context), handled(handledOperations) {}
 
   LogicalResult match(CustomOp op) const override {
+    if (handled.contains(op)) {
+      return failure(); // Skip already handled operations
+    }
     if (op.getGateName() == "Hadamard") {
       return success();
     }
     return failure();
   }
 
-  void rewrite(CustomOp /*op*/, PatternRewriter& /*rewriter*/) const override {
+  void rewrite(CustomOp op, PatternRewriter& /*rewriter*/) const override {
+    handled.insert(op);
     std::cout << "ATTENTION: Hadarmard detected!\n";
   }
 };
 
-void populateThePassPatterns(RewritePatternSet& patterns) {
-  patterns.add<ThePattern>(patterns.getContext());
+void populateThePassPatterns(RewritePatternSet& patterns,
+                             DenseSet<Operation*>& handledOperations) {
+  patterns.add<ThePattern>(patterns.getContext(), handledOperations);
 }
 
 } // namespace mlir::mqt
