### 2025-01-16 829de15
🔒 update to latest `uv` version that properly supports dynamically versioned packages
Signed-off-by: burgholzer <burgholzer@me.com>
