### 2025-01-16 f8a1eca
working rewrite
