### 2025-01-16 2edd072
✨ Setup MQT optimization dialect
