+++ mlir/lib/Dialect/MQTO/IR/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_mlir_library(MLIRMQTO MQTO_Ops.cpp DEPENDS MLIRMQTO_OpsIncGen)
