### 2025-01-16 a30887b
feat: :sparkles: ToQuantumComputation filter now correctly generates QC for simple single-qubit gates
