+++ mlir/lib/Dialect/MQT/Transforms/CMakeLists.txt
@@ -1,5 +1,6 @@
 get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
-set(LIBRARIES ${dialect_libs} MLIRMQT)
+set(LIBRARIES ${dialect_libs} MLIRMQT MQT::CoreIR)
+add_compile_options(-fexceptions)
 
 file(GLOB TRANSFORMS_SOURCES *.cpp)
 
