### 2025-01-16 c41302d
fixed equality check on floating point with bit manipulation
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>
Signed-off-by: M-J-Hochreiter <77507116+M-J-Hochreiter@users.noreply.github.com>
