### 2025-01-16 8a7193c
add hadamard rewrite placeholder
