### 2025-01-16 d66bd2b
update deprecated pattern call
