+++ mlir/mlir/tools/quantum-opt/quantum-opt.cpp
@@ -15,6 +15,7 @@
 #include "mlir/Dialect/Func/Extensions/AllExtensions.h"
 #include "mlir/Dialect/MQT/IR/MQTDialect.h"
 #include "mlir/Dialect/MQT/Transforms/Passes.h"
+#include "mlir/Dialect/MQTO/IR/MQTO_Ops.h"
 #include "mlir/IR/DialectRegistry.h"
 #include "mlir/InitAllDialects.h"
 #include "mlir/InitAllPasses.h"
@@ -28,6 +29,7 @@ int main(int argc, char** argv) {
   mlir::registerAllDialects(registry);
   mlir::func::registerAllExtensions(registry);
   registry.insert<mlir::mqt::MQTDialect>();
+  registry.insert<mlir::mqto::MQTODialect>();
 
   return mlir::asMainReturnCode(
       mlir::MlirOptMain(argc, argv, "Quantum optimizer driver\n", registry));
