### 2025-01-16 b91836f
empty check with proper function
