### 2025-01-16 99647df
style: :art: Resolve pre-commit issues
