### 2025-01-16 2e064cb
style: :art: Resolve pre-commit issues
