+++ mlir/lib/Dialect/MQT/Transforms/ToQuantumComputationPattern.cpp
@@ -1,67 +1,72 @@
-#include "mlir/Dialect/MQT/IR/MQTOps.h"
+#include "ir/QuantumComputation.hpp"
 #include "mlir/Dialect/MQT/IR/MQTDialect.h"
+#include "mlir/Dialect/MQT/IR/MQTOps.h"
 #include "mlir/Dialect/MQT/Transforms/Passes.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/ValueRange.h"
 #include "mlir/Support/LLVM.h"
 
-#include "ir/QuantumComputation.hpp"
-
 #include <iostream>
+#include <mlir/Dialect/Arith/IR/Arith.h>
 #include <mlir/IR/MLIRContext.h>
 #include <mlir/IR/PatternMatch.h>
 #include <mlir/Support/LogicalResult.h>
-#include <mlir/Dialect/Arith/IR/Arith.h>
 #include <set>
 
 namespace mlir::mqt {
-/// Analysis pattern that filters out all quantum operations from a given program and creates a quantum computation from them.
+/// Analysis pattern that filters out all quantum operations from a given
+/// program and creates a quantum computation from them.
 struct ToQuantumComputationPattern final : OpRewritePattern<AllocOp> {
   std::set<Operation*>& handledOperations;
 
-  explicit ToQuantumComputationPattern(MLIRContext* context, std::set<Operation*>& handled)
+  explicit ToQuantumComputationPattern(MLIRContext* context,
+                                       std::set<Operation*>& handled)
       : OpRewritePattern(context), handledOperations(handled) {}
 
   LogicalResult match(AllocOp op) const override {
     if (handledOperations.find(op) == handledOperations.end()) {
       return success();
     }
-    
+
     return failure();
   }
 
-  void handleCustomOp(CustomOp op, qc::QuantumComputation& circuit, std::vector<Value>& currentQubitVariables) const {
+  void handleCustomOp(CustomOp op, qc::QuantumComputation& circuit,
+                      std::vector<Value>& currentQubitVariables) const {
     const auto ins = op.getInQubits();
     const auto outs = op.getOutQubits();
 
     std::vector<size_t> ins_indices(ins.size());
-    std::transform(ins.begin(), ins.end(), ins_indices.begin(), [&currentQubitVariables](Value val) {
-        const auto found = std::find(currentQubitVariables.begin(), currentQubitVariables.end(), val);
-        if(currentQubitVariables.end() == found) {
-          llvm::outs() << "ERROR: Qubit not found!\n";
-          return -1UL;
-        }
-        return (size_t)std::distance(currentQubitVariables.begin(), found);
-    });
+    std::transform(
+        ins.begin(), ins.end(), ins_indices.begin(),
+        [&currentQubitVariables](Value val) {
+          const auto found = std::find(currentQubitVariables.begin(),
+                                       currentQubitVariables.end(), val);
+          if (currentQubitVariables.end() == found) {
+            llvm::outs() << "ERROR: Qubit not found!\n";
+            return -1UL;
+          }
+          return (size_t)std::distance(currentQubitVariables.begin(), found);
+        });
 
-    for(size_t i = 0; i < ins.size(); i++) {
+    for (size_t i = 0; i < ins.size(); i++) {
       currentQubitVariables[ins_indices[i]] = outs[i];
     }
 
     const auto name = op.getGateName();
-    if(name == "Hadamard") {
+    if (name == "Hadamard") {
       circuit.h(ins_indices[0]);
     }
-    if(name == "PauliX") {
+    if (name == "PauliX") {
       circuit.x(ins_indices[0]);
     }
-    if(name == "PauliY") {
+    if (name == "PauliY") {
       circuit.y(ins_indices[0]);
     }
[…]
