+++ mlir/mlir/include/mlir/Dialect/MQT/Transforms/Passes.h
@@ -14,8 +14,8 @@ namespace mqt {
 #include "mlir/Dialect/MQT/Transforms/Passes.h.inc"
 
 void populateThePassPatterns(RewritePatternSet& patterns);
-void populateThePassWithHadamardRewritePattern(RewritePatternSet& patterns);
-void populateThePassWithCXRewritePattern(RewritePatternSet& patterns);
+void populatePassWithSingleQubitGateRewritePattern(RewritePatternSet& patterns);
+void populatePassWithMultiQubitGateRewritePattern(RewritePatternSet& patterns);
 void populateToQuantumComputationPatterns(RewritePatternSet& patterns,
                              std::set<Operation*>& handledOperations);
 
