### 2025-01-16 41df611
🎨 Improve verifier
