### 2025-01-16 bf0b0b9
🐛 Fix type parsing
