### 2025-01-16 743281a
🐛 Fix type problems with signless integer operands
