### 2025-01-16 1ca56fe
📝 Add documentation for Operations and Dialect
