+++ mlir/mlir/include/mlir/Dialect/MQTO/IR/CMakeLists.txt
@@ -1,2 +1,2 @@
 add_mlir_dialect(MQTO_Ops mqto)
-add_mlir_doc(MQTO_Ops MQTO_Dialect Dialects/ -gen-dialect-doc)
+add_mlir_doc(MQTO_Ops MLIRMQTO_Dialect Dialects/ -gen-dialect-doc)
