### 2025-01-16 470ed8e
✨ Setup enums for op type
