### 2025-01-16 cedd2fb
✨👷️ add extensive C++ CI setting triggered by PR label
Signed-off-by: burgholzer <burgholzer@me.com>
