+++ .github/workflows/ci.yml
@@ -20,7 +20,7 @@ jobs:
     name: 🇨‌ Test
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cpp-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@v1.5
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@extensive-ci
     with:
       cmake-args-macos: -DMQT_CORE_WITH_GMP=ON
 
@@ -36,7 +36,7 @@ jobs:
     name: 🐍 Test
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-python-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.5
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@extensive-ci
 
   code-ql:
     name: 📝 CodeQL
