### 2025-01-17 0309e31
Change namespace to mqt::ir::
