+++ mlir/include/mlir/Dialect/MQTOpt/IR/MQTOptDialect.h
@@ -9,26 +9,26 @@
 // Dialect declarations
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Dialect/MQTO/IR/MQTO_OpsDialect.h.inc"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptOpsDialect.h.inc"
 
 //===----------------------------------------------------------------------===//
 // Type declarations
 //===----------------------------------------------------------------------===//
 
 #define GET_TYPEDEF_CLASSES
-#include "mlir/Dialect/MQTO/IR/MQTO_OpsTypes.h.inc"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptOpsTypes.h.inc"
 
 //===----------------------------------------------------------------------===//
 // Enum declarations
 //===----------------------------------------------------------------------===//
 
-#include "mlir/Dialect/MQTO/IR/MQTO_OpsEnums.h.inc"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptOpsEnums.h.inc"
 #define GET_ATTRDEF_CLASSES
-#include "mlir/Dialect/MQTO/IR/MQTO_OpsAttributes.h.inc"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptOpsAttributes.h.inc"
 
 //===----------------------------------------------------------------------===//
 // Operation declarations
 //===----------------------------------------------------------------------===//
 
 #define GET_OP_CLASSES
-#include "mlir/Dialect/MQTO/IR/MQTO_Ops.h.inc"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptOps.h.inc"
