+++ mlir/tools/quantum-opt/quantum-opt.cpp
@@ -1,7 +1,7 @@
 #include "mlir/Dialect/Func/Extensions/AllExtensions.h"
 #include "mlir/Dialect/MQT/IR/MQTDialect.h"
 #include "mlir/Dialect/MQT/Transforms/Passes.h"
-#include "mlir/Dialect/MQTO/IR/MQTO_Dialect.h"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h"
 #include "mlir/IR/DialectRegistry.h"
 #include "mlir/InitAllDialects.h"
 #include "mlir/InitAllPasses.h"
@@ -15,7 +15,7 @@ int main(int argc, char** argv) {
   mlir::registerAllDialects(registry);
   mlir::func::registerAllExtensions(registry);
   registry.insert<mlir::mqt::MQTDialect>();
-  registry.insert<mqtmlir::mqto::MQTODialect>();
+  registry.insert<mqt::ir::opt::MQTOptDialect>();
 
   return mlir::asMainReturnCode(
       mlir::MlirOptMain(argc, argv, "Quantum optimizer driver\n", registry));
