+++ mlir/lib/Dialect/MQTO/IR/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_mlir_dialect_library(MLIRMQTO MQTO_Ops.cpp DEPENDS MLIRMQTO_OpsIncGen MLIRMQTO_EnumsIncGen)
