### 2025-01-17 506a7bb
Change namespace to mqt::ir::
