### 2025-01-17 13b4a0c
🎨 Op type compiling but ivalid input
