### 2025-01-17 79a62f4
fixed new flatten operation function, added first test structure
