+++ test/algorithms/test_statepreparation.cpp
@@ -1,3 +1,12 @@
+/*
+ * Copyright (c) 2025 Chair for Design Automation, TUM
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
 #include "Definitions.hpp"
 #include "algorithms/StatePreparation.hpp"
 #include "dd/Package.hpp"
