+++ mlir/mlir/lib/Dialect/MQTOpt/IR/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_mlir_dialect_library(MLIRMQTOpt MQTOptOps.cpp DEPENDS MLIRMQTOptOpsIncGen MLIRMQTOptEnumsIncGen)
+add_mlir_dialect_library(MLIRMQTOpt MQTOptOps.cpp DEPENDS MLIRMQTOptOpsIncGen)
