### 2025-01-17 95d1832
removed test suffix
they cannot really be represented with only alphanumericals
