+++ mlir/lib/Dialect/MQT/Transforms/ToQuantumComputation.cpp
@@ -1,3 +1,4 @@
+#include "ir/QuantumComputation.hpp"
 #include "mlir/Dialect/MQT/Transforms/Passes.h"
 #include "mlir/IR/Operation.h"
 
@@ -17,6 +18,7 @@ struct ToQuantumComputation final
     : impl::ToQuantumComputationBase<ToQuantumComputation> {
 
   std::set<Operation*> handledOperations;
+  qc::QuantumComputation circuit;
 
   void runOnOperation() override {
     // Get the current operation being operated on.
@@ -25,7 +27,9 @@ struct ToQuantumComputation final
 
     // Define the set of patterns to use.
     RewritePatternSet patterns(ctx);
-    populateToQuantumComputationPatterns(patterns, handledOperations);
+    populateToQuantumComputationPatterns(patterns, handledOperations, circuit);
+    populateFromQuantumComputationPatterns(patterns, handledOperations,
+                                           circuit);
 
     // Apply patterns in an iterative and greedy manner.
     if (failed(applyPatternsAndFoldGreedily(op, std::move(patterns)))) {
