### 2025-01-17 95ebee0
refactor: :recycle: Slight refactor of Roundtrip patterns
