+++ mlir/lib/Dialect/MQT/Transforms/FromQuantumComputationPattern.cpp
@@ -17,107 +17,104 @@ namespace mlir::mqt {
 /// Analysis pattern that filters out all quantum operations from a given
 /// program and creates a quantum computation from them.
 struct FromQuantumComputationPattern final : OpRewritePattern<AllocOp> {
-  std::set<Operation*>& handledOperations;
   qc::QuantumComputation& circuit;
 
   explicit FromQuantumComputationPattern(MLIRContext* context,
-                                       std::set<Operation*>& handled, qc::QuantumComputation& qc)
-      : OpRewritePattern(context), handledOperations(handled), circuit(qc) {}
+                                         qc::QuantumComputation& qc)
+      : OpRewritePattern(context), circuit(qc) {}
 
   LogicalResult match(AllocOp op) const override {
     return op->hasAttr("to_replace") ? success() : failure();
   }
 
-  ExtractOp createRegisterAccess(AllocOp& reg, size_t index, PatternRewriter& rewriter) const {
-    return rewriter.create<ExtractOp>(reg.getLoc(), 
-        QubitType::get(rewriter.getContext()), 
-        reg.getResult(),
-        nullptr,
-        rewriter.getI64IntegerAttr(index));
+  ExtractOp createRegisterAccess(AllocOp& reg, size_t index,
+                                 PatternRewriter& rewriter) const {
+    return rewriter.create<ExtractOp>(
+        reg.getLoc(), QubitType::get(rewriter.getContext()), reg.getResult(),
+        nullptr, rewriter.getI64IntegerAttr(index));
   }
 
-  CustomOp createCustomOp(std::string name, mlir::Location loc, mlir::ValueRange inputs, mlir::ValueRange controlQubits, mlir::ValueRange controlValues, PatternRewriter& rewriter) const {
-    return rewriter.create<CustomOp>(loc, 
-        name, 
-        inputs,
-        controlQubits,
-        controlValues);
+  CustomOp createCustomOp(std::string name, mlir::Location loc,
+                          mlir::ValueRange inputs,
+                          mlir::ValueRange controlQubits,
+                          mlir::ValueRange controlValues,
+                          PatternRewriter& rewriter) const {
+    return rewriter.create<CustomOp>(loc, name, inputs, controlQubits,
+                                     controlValues);
   }
 
   void rewrite(AllocOp op, PatternRewriter& rewriter) const override {
-    handledOperations.insert(op);
-
     std::size_t numQubits = circuit.getNqubits();
 
     static std::map<qc::OpType, std::string> opNames = {
         {qc::OpType::H, "Hadamard"},
         {qc::OpType::X, "PauliX"},
         {qc::OpType::Y, "PauliY"},
-        {qc::OpType::Z, "PauliZ"}
-    };
+        {qc::OpType::Z, "PauliZ"}};
 
-
-    auto newAlloc = rewriter.create<AllocOp>(op.getLoc(), QuregType::get(rewriter.getContext()), nullptr, rewriter.getIntegerAttr(rewriter.getI64Type(), numQubits));
+    auto newAlloc = rewriter.create<AllocOp>(
+        op.getLoc(), QuregType::get(rewriter.getContext()), nullptr,
+        rewriter.getIntegerAttr(rewriter.getI64Type(), numQubits));
     newAlloc->setAttr("mqt_core", rewriter.getUnitAttr());
 
     std::vector<Value> currentQubitVariables(numQubits);
-    for(size_t i = 0; i < numQubits; i++) {
-        currentQubitVariables[i] = createRegisterAccess(newAlloc, i, rewriter).getResult();
+    for (size_t i = 0; i < numQubits; i++) {
+      currentQubitVariables[i] =
+          createRegisterAccess(newAlloc, i, rewriter).getResult();
     }
 
-    for(const auto& o : circuit) {
-        std::string opName = "";
[…]
