### 2025-01-17 e827de2
refactor: :recycle: Slight refactor of Roundtrip patterns
