+++ mlir/mlir/lib/Dialect/MQT/Transforms/SingleQubitGateRewritePattern.cpp
@@ -1,16 +1,15 @@
-#include "mlir/Dialect/MQT/IR/MQTOps.h"
 #include "mlir/Dialect/MQT/IR/MQTDialect.h"
+#include "mlir/Dialect/MQT/IR/MQTOps.h"
 #include "mlir/Dialect/MQT/Transforms/Passes.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/ValueRange.h"
 #include "mlir/Support/LLVM.h"
 
 #include <iostream>
+#include <mlir/Dialect/Arith/IR/Arith.h>
 #include <mlir/IR/MLIRContext.h>
 #include <mlir/IR/PatternMatch.h>
 #include <mlir/Support/LogicalResult.h>
-#include <mlir/Dialect/Arith/IR/Arith.h>
-
 
 namespace mlir::mqt {
 
@@ -31,15 +30,17 @@ struct SingleQubitGateRewritePattern : public OpRewritePattern<CustomOp> {
     auto qubit = op.getInQubits()[0];
 
     // Replace with a new MQT operation.
-    auto mqtoPlaceholder = rewriter.create<CustomOp>(op.getLoc(), "MQTOSingleQubitGatePlaceholder", mlir::ValueRange{qubit});
-    
+    auto mqtoPlaceholder = rewriter.create<CustomOp>(
+        op.getLoc(), "MQTOSingleQubitGatePlaceholder", mlir::ValueRange{qubit});
+
     // Replace the original operation with the new MQT operation.
     rewriter.replaceOp(op, mqtoPlaceholder);
   }
 };
 
 /// Populate patterns for the transformation pass.
-void populatePassWithSingleQubitGateRewritePattern(RewritePatternSet& patterns) {
+void populatePassWithSingleQubitGateRewritePattern(
+    RewritePatternSet& patterns) {
   patterns.add<SingleQubitGateRewritePattern>(patterns.getContext());
 }
 
