### 2025-01-17 bdb4442
add first working pass with mqtopt
