### 2025-01-17 7bc347e
add first working pass with mqtopt
