+++ mlir/lib/Dialect/MQT/Transforms/ThePass.cpp
@@ -1,4 +1,5 @@
 #include "mlir/Dialect/MQT/Transforms/Passes.h"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h"
 #include "mlir/IR/Operation.h"
 
 #include <mlir/IR/MLIRContext.h>
@@ -14,6 +15,11 @@ namespace mlir::mqt {
 
 struct ThePass final : impl::ThePassBase<ThePass> {
 
+    void getDependentDialects(DialectRegistry& registry) const override {
+    // Register the required dialects.
+    registry.insert<::mqt::ir::opt::MQTOptDialect>();
+  }
+
   void runOnOperation() override {
     // Get the current operation being operated on.
     Operation* op = getOperation();
@@ -21,9 +27,11 @@ struct ThePass final : impl::ThePassBase<ThePass> {
 
     // Define the set of patterns to use.
     RewritePatternSet thePatterns(ctx);
-    populateThePassPatterns(thePatterns);
-    populatePassWithSingleQubitGateRewritePattern(thePatterns);
-    populatePassWithMultiQubitGateRewritePattern(thePatterns);
+    ctx->getOrLoadDialect<::mqt::ir::opt::MQTOptDialect>();
+    //populateThePassPatterns(thePatterns);
+    //populatePassWithSingleQubitGateRewritePattern(thePatterns);
+    //populatePassWithMultiQubitGateRewritePattern(thePatterns);
+    populatePassWithAllocRewritePattern(thePatterns);
 
     // Apply patterns in an iterative and greedy manner.
     if (failed(applyPatternsAndFoldGreedily(op, std::move(thePatterns)))) {
