### 2025-01-17 8bd075c
Revert "Merge remote-tracking branch 'origin/standalone' into quantum_to_mqto"
This reverts commit 2af2274d2fc66bae591b7529521d886c2938d20d, reversing
changes made to c62934ef4a109c6a404337a965ece26ab64e9102.
