+++ mlir/include/mlir/Dialect/MQT/Transforms/Passes.h
@@ -14,11 +14,13 @@ namespace mqt {
 #define GEN_PASS_DECL
 #include "mlir/Dialect/MQT/Transforms/Passes.h.inc"
 
-//void populateThePassPatterns(RewritePatternSet& patterns);
-//void populatePassWithSingleQubitGateRewritePattern(RewritePatternSet& patterns);
-//void populatePassWithMultiQubitGateRewritePattern(RewritePatternSet& patterns);
+void populateThePassPatterns(RewritePatternSet& patterns);
 void populateToQuantumComputationPatterns(RewritePatternSet& patterns,
-                             std::set<Operation*>& handledOperations);
+                                          qc::QuantumComputation& circuit);
+void populateFromQuantumComputationPatterns(RewritePatternSet& patterns,
+                                            qc::QuantumComputation& circuit);
+void populatePassWithSingleQubitGateRewritePattern(RewritePatternSet& patterns);
+void populatePassWithMultiQubitGateRewritePattern(RewritePatternSet& patterns);
 
 //===----------------------------------------------------------------------===//
 // Registration
