### 2025-01-17 afe23c1
Revert "Merge remote-tracking branch 'origin/standalone' into quantum_to_mqto"
This reverts commit cd5518dbe612d74aac91418abbf379485141efd9, reversing
changes made to 89fded204e2e7f1b8038dc7163400d05eb74622e.
