+++ mlir/mlir/lib/Dialect/MQT/Transforms/SingleQubitGateRewritePattern.cpp
@@ -1,7 +1,6 @@
 #include "mlir/Dialect/MQT/IR/MQTDialect.h"
 #include "mlir/Dialect/MQT/IR/MQTOps.h"
 #include "mlir/Dialect/MQT/Transforms/Passes.h"
-#include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h"
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/ValueRange.h"
 #include "mlir/Support/LLVM.h"
@@ -14,28 +13,26 @@
 
 namespace mlir::mqt {
 
-/// Pattern to transform `quantum.custom` single-qubit gate to MQT equivalent.
+/// Pattern to transform `quantum.custom` multi-qubit gate to MQT equivalent.
 struct SingleQubitGateRewritePattern : public OpRewritePattern<CustomOp> {
   explicit SingleQubitGateRewritePattern(MLIRContext* context)
       : OpRewritePattern(context) {}
 
   LogicalResult match(CustomOp op) const override {
-    // NOTE: This pattern is currently only checking for the X gate.
-    return op.getGateName() == "PauliX" ? success() : failure();
+    // NOTE: This pattern is currently only checking for the Hadamard gate.
+    return op.getGateName() == "Hadamard" ? success() : failure();
   }
 
   void rewrite(CustomOp op, PatternRewriter& rewriter) const override {
     std::cout << "APPLYING: SingleQubitGateRewritePattern\n";
 
-    // Extract the input qubit from the original operation.
-    auto inQubit = op.getInQubits()[0];
-    ValueRange ctrlQubits; // No control qubits in this example.
+    // Extract the input qubit for the operation.
+    auto qubit = op.getInQubits()[0];
 
     // Replace with a new MQT operation.
-    auto mqtoPlaceholder = rewriter.create<::mqt::ir::opt::XOp>(
-      op.getLoc(), TypeRange{inQubit.getType()}, inQubit, ctrlQubits
-      );
-    
+    auto mqtoPlaceholder = rewriter.create<CustomOp>(
+        op.getLoc(), "MQTOSingleQubitGatePlaceholder", mlir::ValueRange{qubit});
+
     // Replace the original operation with the new MQT operation.
     rewriter.replaceOp(op, mqtoPlaceholder);
   }
