+++ mlir/include/mlir/Dialect/MQTOpt/IR/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_mlir_dialect(MQTOptOps mqtopt)
-# add_mlir_interface(MQTOptOps)
+add_mlir_interface(MQTOptInterfaces)
 add_mlir_doc(MQTOptOps MLIRMQTOptDialect Dialects/ -gen-dialect-doc)
+add_mlir_doc(MQTOptInterfaces MLIRMQTOptInterfaces Dialects/ -gen-op-interface-docs)
