### 2025-01-18 1128c3e
⬆️🪝 Update pre-commit hook pre-commit/mirrors-clang-format to v19.1.7
Signed-off-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
