### 2025-01-18 5bc6692
⚗️ experiment with an extensive CI setting triggered by PR labels
Signed-off-by: burgholzer <burgholzer@me.com>
