+++ .github/workflows/ci.yml
@@ -22,7 +22,99 @@ jobs:
     if: fromJSON(needs.change-detection.outputs.run-cpp-tests)
     uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@extensive-ci
     with:
+      # Runs to enable
+      enable-ubuntu2404-gcc-release: true
+      enable-ubuntu2404-arm-gcc-release: true
+      enable-macos13-clang-release: true
+      enable-macos14-clang-release: true
+      enable-windows2022-msvc-release: true
+      # Runs to disable
+      enable-ubuntu2404-gcc-debug: false
+      enable-ubuntu2404-clang-release: false
+      enable-ubuntu2404-clang-debug: false
+      enable-ubuntu2204-gcc-release: false
+      enable-ubuntu2204-gcc-debug: false
+      enable-ubuntu2204-clang-release: false
+      enable-ubuntu2204-clang-debug: false
+      enable-ubuntu2404-arm-gcc-debug: false
+      enable-ubuntu2404-arm-clang-release: false
+      enable-ubuntu2404-arm-clang-debug: false
+      enable-ubuntu2204-arm-gcc-release: false
+      enable-ubuntu2204-arm-gcc-debug: false
+      enable-ubuntu2204-arm-clang-release: false
+      enable-ubuntu2204-arm-clang-debug: false
+      enable-macos13-clang-debug: false
+      enable-macos13-gcc-release: false
+      enable-macos13-gcc-debug: false
+      enable-macos14-clang-debug: false
+      enable-macos14-gcc-release: false
+      enable-macos14-gcc-debug: false
+      enable-macos15-clang-release: false
+      enable-macos15-clang-debug: false
+      enable-macos15-gcc-release: false
+      enable-macos15-gcc-debug: false
+      enable-windows2022-msvc-debug: false
+      enable-windows2022-clang-release: false
+      enable-windows2022-clang-debug: false
+      enable-windows2025-msvc-release: false
+      enable-windows2025-msvc-debug: false
+      enable-windows2025-clang-release: false
+      enable-windows2025-clang-debug: false
+      # General settings
       cmake-args-macos: -DMQT_CORE_WITH_GMP=ON
+
+  # run extensive cpp tests whenever this is on a PR and the PR has the `extensive-ci` label
+  cpp-tests-extensive:
+    name: 🇨‌ Test (Extensive)
+    needs: change-detection
+    if: fromJSON(needs.change-detection.outputs.run-cpp-tests) && github.event_name == 'pull_request' && contains(github.event.pull_request.labels.*.name, 'extensive-ci')
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@extensive-ci
+    with:
+      # Runs to enable
+      enable-ubuntu2404-gcc-release: true
+      enable-ubuntu2404-gcc-debug: true
+      enable-ubuntu2404-clang-release: true
+      enable-ubuntu2404-clang-debug: true
+      enable-ubuntu2204-gcc-release: true
+      enable-ubuntu2204-gcc-debug: true
+      enable-ubuntu2204-clang-release: true
+      enable-ubuntu2204-clang-debug: true
+      enable-ubuntu2404-arm-gcc-release: true
+      enable-ubuntu2404-arm-gcc-debug: true
+      enable-ubuntu2404-arm-clang-release: true
+      enable-ubuntu2404-arm-clang-debug: true
+      enable-ubuntu2204-arm-gcc-release: true
+      enable-ubuntu2204-arm-gcc-debug: true
+      enable-ubuntu2204-arm-clang-release: true
+      enable-ubuntu2204-arm-clang-debug: true
[…]
