### 2025-01-18 d740492
⚗️ disable other tests for now
Signed-off-by: burgholzer <burgholzer@me.com>
