+++ mlir/include/mlir/Dialect/MQTOpt/Transforms/CMakeLists.txt
@@ -1,1 +1,4 @@
-# empty so far
+set(LLVM_TARGET_DEFINITIONS Passes.td)
+mlir_tablegen(Passes.h.inc -gen-pass-decls -name MQTOpt)
+add_public_tablegen_target(MLIRMQTOptTransformsIncGen)
+add_mlir_doc(MQTOptPasses MLIRMQTOptPasses Passes/ -gen-pass-doc)
