### 2025-01-20 3f200fe
🎨 Add common interface methods
