+++ mlir/mlir/lib/Dialect/MQTOpt/Transforms/ThePattern.cpp
@@ -1,10 +1,12 @@
 #include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h"
 #include "mlir/Dialect/MQTOpt/Transforms/Passes.h"
+#include "mlir/IR/Value.h"
 
 #include <mlir/IR/MLIRContext.h>
 #include <mlir/IR/PatternMatch.h>
 #include <mlir/Support/LLVM.h>
 #include <mlir/Support/LogicalResult.h>
+#include <vector>
 
 namespace mqt::ir::opt {
 /// Multistep rewrite using "match" and "rewrite". This allows for separating
