### 2025-01-20 5b93fd0
✨  added qasm string for passby operation.
