+++ mlir/mlir/include/mlir/Dialect/MQTOpt/Transforms/Passes.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include "ir/QuantumComputation.hpp"
+
+#include <mlir/Pass/Pass.h>
+#include <set>
+
+namespace mlir {
+
+    class RewritePatternSet;
+
+} // namespace mlir
+
+namespace mqt::ir::opt {
+
+#define GEN_PASS_DECL
+#include "mlir/Dialect/MQTOpt/Transforms/Passes.h.inc"
+
+void populateToQuantumComputationPatterns(mlir::RewritePatternSet& patterns,
+                                          qc::QuantumComputation& circuit);
+void populateFromQuantumComputationPatterns(mlir::RewritePatternSet& patterns,
+                                            qc::QuantumComputation& circuit);
+
+//===----------------------------------------------------------------------===//
+// Registration
+//===----------------------------------------------------------------------===//
+
+/// Generate the code for registering passes.
+#define GEN_PASS_REGISTRATION
+#include "mlir/Dialect/MQTOpt/Transforms/Passes.h.inc"
+
+} // namespace mqt::ir::opt
