+++ mlir/mlir/lib/Dialect/MQTOpt/Transforms/ToQuantumComputationPattern.cpp
@@ -1,0 +1,250 @@
+#include "ir/QuantumComputation.hpp"
+#include "ir/operations/StandardOperation.hpp"
+#include "ir/operations/OpType.hpp"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h"
+#include "mlir/Dialect/MQTOpt/Transforms/Passes.h"
+#include "mlir/IR/BuiltinTypes.h"
+#include "mlir/IR/ValueRange.h"
+#include "mlir/Support/LLVM.h"
+
+#include <iostream>
+#include <mlir/Dialect/Arith/IR/Arith.h>
+#include <mlir/IR/MLIRContext.h>
+#include <mlir/IR/PatternMatch.h>
+#include <mlir/Support/LogicalResult.h>
+#include <set>
+
+namespace mqt::ir::opt {
+/// Analysis pattern that filters out all quantum operations from a given
+/// program and creates a quantum computation from them.
+struct ToQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
+  qc::QuantumComputation& circuit;
+
+  explicit ToQuantumComputationPattern(mlir::MLIRContext* context,
+                                       qc::QuantumComputation& qc)
+      : OpRewritePattern(context), circuit(qc) {}
+
+  mlir::LogicalResult match(AllocOp op) const override {
+    return (op->hasAttr("to_replace") || op->hasAttr("mqt_core")) ? mlir::failure()
+                                                                  : mlir::success();
+  }
+
+  size_t findQubitIndex(mlir::Value input, std::vector<mlir::Value>& currentQubitVariables) const {
+    size_t arrayIndex = 0;
+    if (auto opResult = mlir::dyn_cast<mlir::OpResult>(input)) {
+      arrayIndex = opResult.getResultNumber();
+    } else {
+      throw std::runtime_error("Operand is not an operation result. This should never happen!");
+    }
+    for(size_t i = 0; i < currentQubitVariables.size(); i++) {
+      size_t qubitArrayIndex = 0;
+      if (auto opResult = mlir::dyn_cast<mlir::OpResult>(currentQubitVariables[i])) {
+        qubitArrayIndex = opResult.getResultNumber();
+      } else {
+        throw std::runtime_error("Qubit is not an operation result. This should never happen!");
+      }
+
+      if(currentQubitVariables[i] == input && arrayIndex == qubitArrayIndex) {
+        return i;
+      }
+    }
+
+    throw std::runtime_error("Qubit was not found in list of previously defined qubits");
+  }
+
+  void handleMeasureOp(MeasureOp op, std::vector<mlir::Value>& currentQubitVariables) const {
+    const auto ins = op.getInQubits();
+    const auto outs = op.getOutQubits();
+
+    std::vector<size_t> insIndices(ins.size());
+    std::transform(
+        ins.begin(), ins.end(), insIndices.begin(),
+        [&currentQubitVariables, this](mlir::Value val) {
+          return findQubitIndex(val, currentQubitVariables);
+        });
+
+    for (size_t i = 0; i < insIndices.size(); i++) {
+      currentQubitVariables[insIndices[i]] = outs[i];
+      circuit.measure(insIndices[i], insIndices[i]);
+    }
+  }
+
+  void handleXOp(XOp op,
+                      std::vector<mlir::Value>& currentQubitVariables) const {
+    const auto in = op.getInQubit();
+    const auto ctrlIns = op.getCtrlQubits();
+    const auto outs = op.getOutQubits();
+
+
+    std::vector<size_t> insIndices(1 + ctrlIns.size());
+
+    const auto found = std::find(currentQubitVariables.begin(),
+                                       currentQubitVariables.end(), in);
[…]
