+++ mlir/lib/Dialect/MQTOpt/Transforms/FromQuantumComputationPattern.cpp
@@ -1,0 +1,124 @@
+#include "ir/QuantumComputation.hpp"
+#include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h"
+#include "mlir/Dialect/MQTOpt/Transforms/Passes.h"
+#include "mlir/IR/BuiltinTypes.h"
+#include "mlir/IR/ValueRange.h"
+#include "mlir/Support/LLVM.h"
+
+#include <iostream>
+#include <mlir/Dialect/Arith/IR/Arith.h>
+#include <mlir/IR/MLIRContext.h>
+#include <mlir/IR/PatternMatch.h>
+#include <mlir/Support/LogicalResult.h>
+#include <set>
+
+namespace mqt::ir::opt {
+/// Analysis pattern that filters out all quantum operations from a given
+/// program and creates a quantum computation from them.
+struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
+  qc::QuantumComputation& circuit;
+
+  explicit FromQuantumComputationPattern(mlir::MLIRContext* context,
+                                         qc::QuantumComputation& qc)
+      : OpRewritePattern(context), circuit(qc) {}
+
+  mlir::LogicalResult match(AllocOp op) const override {
+    return op->hasAttr("to_replace") ? mlir::success() : mlir::failure();
+  }
+
+  ExtractOp createRegisterAccess(AllocOp& reg, size_t index,
+                                 mlir::PatternRewriter& rewriter) const {
+    return rewriter.create<ExtractOp>(
+        reg.getLoc(), mlir::TypeRange{QubitRegisterType::get(rewriter.getContext()), QubitType::get(rewriter.getContext())}, reg.getResult(),
+        nullptr, rewriter.getI64IntegerAttr(index));
+  }
+
+  XOp createXOp(mlir::Location loc,
+                          mlir::Value inQubit,
+                          mlir::ValueRange controlQubits,
+                          mlir::PatternRewriter& rewriter) const {
+    llvm::SmallVector<mlir::Type, 4> resultTypes(4, inQubit.getType());
+    return rewriter.create<XOp>(loc, resultTypes, inQubit, controlQubits);
+  }
+
+  MeasureOp createMeasureOp(mlir::Location loc,
+                          mlir::Value targetQubit,
+                          mlir::PatternRewriter& rewriter) const {
+    return rewriter.create<MeasureOp>(loc, mlir::TypeRange{QubitType::get(rewriter.getContext()), rewriter.getI1Type()}, targetQubit);
+  }
+
+  void updateReturnOperation(mlir::Operation* op, mlir::Operation* returnOperation, std::vector<mlir::Value>& measurementValues, mlir::PatternRewriter& rewriter) const {
+    const auto cloned = rewriter.clone(*returnOperation);
+    for(size_t i = 0; i < measurementValues.size(); i++) {
+      cloned->setOperand(i + 1, measurementValues[i]);
+    }
+    rewriter.replaceOp(returnOperation, cloned);
+  }
+
+  void rewrite(AllocOp op, mlir::PatternRewriter& rewriter) const override {
+    std::size_t numQubits = circuit.getNqubits();
+    std::vector<mlir::Value> measurementValues(numQubits);
+
+    auto newAlloc = rewriter.create<AllocOp>(
+        op.getLoc(), QubitRegisterType::get(rewriter.getContext()), nullptr,
+        rewriter.getIntegerAttr(rewriter.getI64Type(), numQubits));
+    newAlloc->setAttr("mqt_core", rewriter.getUnitAttr());
+
+    std::vector<mlir::Value> currentQubitVariables(numQubits);
+    for (size_t i = 0; i < numQubits; i++) {
+      currentQubitVariables[i] =
[…]
