+++ mlir/lib/Dialect/MQTOpt/Transforms/CMakeLists.txt
@@ -1,1 +1,8 @@
-# empty so far
+get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
+set(LIBRARIES ${dialect_libs} MLIRMQT MQT::CoreIR)
+add_compile_options(-fexceptions)
+
+file(GLOB TRANSFORMS_SOURCES *.cpp)
+
+add_mlir_library(MLIRMQTOptTransforms ${TRANSFORMS_SOURCES} LINK_LIBS ${LIBRARIES} DEPENDS
+                 MLIRMQTOptTransformsIncGen)
