### 2025-01-21 e04ab7a
🚚 define dedicated `BitString` types for BV and Grover
Signed-off-by: burgholzer <burgholzer@me.com>
