+++ include/mqt-core/algorithms/Grover.hpp
@@ -12,17 +12,20 @@
 #include "Definitions.hpp"
 #include "ir/QuantumComputation.hpp"
 
+#include <bitset>
 #include <cstddef>
 
 namespace qc {
+
+using GroverBitString = std::bitset<128>;
 auto appendGroverInitialization(QuantumComputation& qc) -> void;
-auto appendGroverOracle(QuantumComputation& qc, const BitString& targetValue)
-    -> void;
+auto appendGroverOracle(QuantumComputation& qc,
+                        const GroverBitString& targetValue) -> void;
 auto appendGroverDiffusion(QuantumComputation& qc) -> void;
 
 [[nodiscard]] auto computeNumberOfIterations(Qubit nq) -> std::size_t;
 
-[[nodiscard]] auto createGrover(Qubit nq, const BitString& targetValue)
+[[nodiscard]] auto createGrover(Qubit nq, const GroverBitString& targetValue)
     -> QuantumComputation;
 [[nodiscard]] auto createGrover(Qubit nq, std::size_t seed = 0)
     -> QuantumComputation;
