### 2025-01-21 5e7186a
🚚 move `OUTPUT_INDENT_SIZE` to `Operation`
Signed-off-by: burgholzer <burgholzer@me.com>
