### 2025-01-21 ca6a9e2
🚚 move `DAG` related typdefs to circuit optimizer
Signed-off-by: burgholzer <burgholzer@me.com>
