+++ include/mqt-core/Definitions.hpp
@@ -11,7 +11,6 @@
 
 #include <cstddef>
 #include <cstdint>
-#include <deque>
 #include <functional>
 #include <map>
 #include <memory>
@@ -62,18 +61,8 @@ static constexpr fp TAU = static_cast<fp>(
 static constexpr fp E = static_cast<fp>(
     2.718281828459045235360287471352662497757247093699959574967L);
 
-// forward declaration
-class Operation;
-
 // supported file formats
 enum class Format : uint8_t { Real, OpenQASM2, OpenQASM3, TFC, QC, Tensor };
-
-using DAG = std::vector<std::deque<std::unique_ptr<Operation>*>>;
-using DAGIterator = std::deque<std::unique_ptr<Operation>*>::iterator;
-using DAGReverseIterator =
-    std::deque<std::unique_ptr<Operation>*>::reverse_iterator;
-using DAGIterators = std::vector<DAGIterator>;
-using DAGReverseIterators = std::vector<DAGReverseIterator>;
 
 /**
  * @brief 64bit mixing hash (from MurmurHash3)
