+++ test/ir/test_qfr_functionality.cpp
@@ -172,18 +172,18 @@ TEST_F(QFRFunctionality, StripIdleAndDump) {
                                "cx q[0],q[4];\n";
 
   ss << testfile;
-  auto qc = qc::QuantumComputation();
-  qc.import(ss, qc::Format::OpenQASM2);
+  auto qc = QuantumComputation();
+  qc.import(ss, Format::OpenQASM2);
   qc.print(std::cout);
   qc.stripIdleQubits();
   qc.print(std::cout);
   std::stringstream goal{};
   qc.print(goal);
   std::stringstream test{};
-  qc.dump(test, qc::Format::OpenQASM2);
+  qc.dump(test, Format::OpenQASM2);
   std::cout << test.str() << "\n";
   qc.reset();
-  qc.import(test, qc::Format::OpenQASM2);
+  qc.import(test, Format::OpenQASM2);
   qc.print(std::cout);
   qc.stripIdleQubits();
   qc.print(std::cout);
@@ -332,15 +332,15 @@ TEST_F(QFRFunctionality, wrongRegisterSizes) {
 }
 
 TEST_F(QFRFunctionality, OperationEquality) {
-  const auto x = StandardOperation(0, qc::X);
-  const auto z = StandardOperation(0, qc::Z);
+  const auto x = StandardOperation(0, X);
+  const auto z = StandardOperation(0, Z);
   EXPECT_TRUE(x.equals(x));
   EXPECT_EQ(x, x);
   EXPECT_FALSE(x.equals(z));
   EXPECT_NE(x, z);
 
-  const auto x0 = StandardOperation(0, qc::X);
-  const auto x1 = StandardOperation(1, qc::X);
+  const auto x0 = StandardOperation(0, X);
+  const auto x1 = StandardOperation(1, X);
   EXPECT_FALSE(x0.equals(x1));
   EXPECT_NE(x0, x1);
   Permutation perm0{};
@@ -349,15 +349,15 @@ TEST_F(QFRFunctionality, OperationEquality) {
   EXPECT_TRUE(x0.equals(x1, perm0, {}));
   EXPECT_TRUE(x0.equals(x1, {}, perm0));
 
-  const auto cx01 = StandardOperation(0, 1, qc::X);
-  const auto cx10 = StandardOperation(1, 0, qc::X);
+  const auto cx01 = StandardOperation(0, 1, X);
+  const auto cx10 = StandardOperation(1, 0, X);
   EXPECT_FALSE(cx01.equals(cx10));
   EXPECT_NE(cx01, cx10);
   EXPECT_FALSE(x0.equals(cx01));
   EXPECT_NE(x0, cx01);
 
-  const auto p = StandardOperation(0, qc::P, {2.0});
-  const auto pm = StandardOperation(0, qc::P, {-2.0});
+  const auto p = StandardOperation(0, P, {2.0});
+  const auto pm = StandardOperation(0, P, {-2.0});
   EXPECT_FALSE(p.equals(pm));
   EXPECT_NE(p, pm);
 
@@ -407,14 +407,14 @@ TEST_F(QFRFunctionality, OperationEquality) {
   EXPECT_NE(classic0, classic3);
 
   auto compound0 = CompoundOperation();
-  compound0.emplace_back<StandardOperation>(0, qc::X);
+  compound0.emplace_back<StandardOperation>(0, X);
 
[…]
