+++ src/circuit_optimizer/CircuitOptimizer.cpp
@@ -48,7 +48,7 @@ void CircuitOptimizer::removeIdentities(QuantumComputation& qc) {
 }
 
 void CircuitOptimizer::removeOperation(
-    qc::QuantumComputation& qc, const std::unordered_set<OpType>& opTypes,
+    QuantumComputation& qc, const std::unordered_set<OpType>& opTypes,
     const size_t opSize) {
   // opSize = 0 means that the operation can have any number of qubits
   auto it = qc.begin();
@@ -112,9 +112,9 @@ void CircuitOptimizer::swapReconstruction(QuantumComputation& qc) {
     auto* opT = dag.at(target).back();
 
     // previous operation is not a CNOT
-    if ((*opC)->getType() != qc::X || (*opC)->getNcontrols() != 1 ||
+    if ((*opC)->getType() != X || (*opC)->getNcontrols() != 1 ||
         (*opC)->getControls().begin()->type != Control::Type::Pos ||
-        (*opT)->getType() != qc::X || (*opT)->getNcontrols() != 1 ||
+        (*opT)->getType() != X || (*opT)->getNcontrols() != 1 ||
         (*opT)->getControls().begin()->type != Control::Type::Pos) {
       addToDag(dag, &it);
       continue;
@@ -174,11 +174,9 @@ CircuitOptimizer::DAG CircuitOptimizer::constructDAG(QuantumComputation& qc) {
 }
 
 void CircuitOptimizer::singleQubitGateFusion(QuantumComputation& qc) {
-  static const std::map<qc::OpType, qc::OpType> INVERSE_MAP = {
-      {qc::I, qc::I},     {qc::X, qc::X},     {qc::Y, qc::Y},
-      {qc::Z, qc::Z},     {qc::H, qc::H},     {qc::S, qc::Sdg},
-      {qc::Sdg, qc::S},   {qc::T, qc::Tdg},   {qc::Tdg, qc::T},
-      {qc::SX, qc::SXdg}, {qc::SXdg, qc::SX}, {qc::Barrier, qc::Barrier}};
+  static const std::map<OpType, OpType> INVERSE_MAP = {
+      {I, I},   {X, X},   {Y, Y},   {Z, Z},     {H, H},     {S, Sdg},
+      {Sdg, S}, {T, Tdg}, {Tdg, T}, {SX, SXdg}, {SXdg, SX}, {Barrier, Barrier}};
 
   auto dag = DAG(qc.getHighestPhysicalQubitIndex() + 1);
 
@@ -238,7 +236,7 @@ void CircuitOptimizer::singleQubitGateFusion(QuantumComputation& qc) {
       if (inverseIt != INVERSE_MAP.end() &&
           it->getType() == inverseIt->second) {
         compop->pop_back();
-        it->setGate(qc::I);
+        it->setGate(I);
       } else {
         compop->emplace_back<StandardOperation>(
             it->getTargets().at(0), it->getType(), it->getParameter());
@@ -253,8 +251,8 @@ void CircuitOptimizer::singleQubitGateFusion(QuantumComputation& qc) {
     // check if current operation is the inverse of the previous operation
     auto inverseIt = INVERSE_MAP.find((*op)->getType());
     if (inverseIt != INVERSE_MAP.end() && it->getType() == inverseIt->second) {
-      (*op)->setGate(qc::I);
-      it->setGate(qc::I);
+      (*op)->setGate(I);
+      it->setGate(I);
     } else {
       auto compop = std::make_unique<CompoundOperation>();
       compop->emplace_back<StandardOperation>(
@@ -319,7 +317,7 @@ void removeDiagonalGatesBeforeMeasureRecursive(
     } else if (op->isCompoundOperation()) {
[…]
