### 2025-01-21 395e65e
♻️ refactor register handling
Signed-off-by: burgholzer <burgholzer@me.com>
