+++ src/python/ir/register_ir.cpp
@@ -15,6 +15,7 @@ namespace py = pybind11;
 using namespace pybind11::literals;
 
 // forward declarations
+void registerRegisters(py::module& m);
 void registerPermutation(py::module& m);
 void registerOperations(py::module& m);
 void registerSymbolic(py::module& m);
@@ -22,6 +23,7 @@ void registerQuantumComputation(py::module& m);
 
 PYBIND11_MODULE(ir, m, py::mod_gil_not_used()) {
   registerPermutation(m);
+  registerRegisters(m);
 
   py::module symbolic = m.def_submodule("symbolic");
   registerSymbolic(symbolic);
