+++ include/mqt-core/ir/operations/Operation.hpp
@@ -13,6 +13,7 @@
 #include "Definitions.hpp"
 #include "OpType.hpp"
 #include "ir/Permutation.hpp"
+#include "ir/Register.hpp"
 
 #include <algorithm>
 #include <array>
@@ -36,11 +37,12 @@ protected:
 
   static constexpr size_t OUTPUT_INDENT_SIZE = 2;
 
-  static bool isWholeQubitRegister(const RegisterNames& reg, std::size_t start,
-                                   std::size_t end) {
-    return !reg.empty() && reg[start].first == reg[end].first &&
-           (start == 0 || reg[start].first != reg[start - 1].first) &&
-           (end == reg.size() - 1 || reg[end].first != reg[end + 1].first);
+  static bool isWholeQubitRegister(const QubitIndexToRegisterMap& regMap,
+                                   const Qubit start, const Qubit end) {
+    const auto& startReg = regMap.at(start).first;
+    const auto& endReg = regMap.at(end).first;
+    return startReg == endReg && startReg.getStartIndex() == start &&
+           endReg.getEndIndex() == end;
   }
 
 public:
@@ -180,16 +182,17 @@ public:
                               std::size_t prefixWidth,
                               std::size_t nqubits) const;
 
-  void dumpOpenQASM2(std::ostream& of, const RegisterNames& qreg,
-                     const RegisterNames& creg) const {
-    dumpOpenQASM(of, qreg, creg, 0, false);
+  void dumpOpenQASM2(std::ostream& of, const QubitIndexToRegisterMap& qubitMap,
+                     const BitIndexToRegisterMap& bitMap) const {
+    dumpOpenQASM(of, qubitMap, bitMap, 0, false);
   }
-  void dumpOpenQASM3(std::ostream& of, const RegisterNames& qreg,
-                     const RegisterNames& creg) const {
-    dumpOpenQASM(of, qreg, creg, 0, true);
+  void dumpOpenQASM3(std::ostream& of, const QubitIndexToRegisterMap& qubitMap,
+                     const BitIndexToRegisterMap& bitMap) const {
+    dumpOpenQASM(of, qubitMap, bitMap, 0, true);
   }
-  virtual void dumpOpenQASM(std::ostream& of, const RegisterNames& qreg,
-                            const RegisterNames& creg, size_t indent,
+  virtual void dumpOpenQASM(std::ostream& of,
+                            const QubitIndexToRegisterMap& qubitMap,
+                            const BitIndexToRegisterMap& bitMap, size_t indent,
                             bool openQASM3) const = 0;
 
   /// Checks whether operation commutes with other operation on a given qubit.
