### 2025-01-21 6d88b97
♻️ refactor classic-controlled operation to either take a full register or a single bit
Signed-off-by: burgholzer <burgholzer@me.com>
