+++ test/circuit_optimizer/test_eliminate_resets.cpp
@@ -84,7 +84,7 @@ TEST(EliminateResets, eliminateResetsClassicControlled) {
   qc.h(0);
   qc.measure(0, 0U);
   qc.reset(0);
-  qc.classicControlled(qc::X, 0, {0, 1U}, 1U);
+  qc.classicControlled(qc::X, 0, 0, 1U);
   std::cout << qc << "\n";
 
   EXPECT_TRUE(qc.isDynamic());
@@ -180,7 +180,7 @@ TEST(EliminateResets, eliminateResetsCompoundOperation) {
   comp.cx(1, 0);
   comp.reset(0);
   comp.measure(0, 0);
-  comp.classicControlled(qc::X, 0, {0, 1U}, 1U);
+  comp.classicControlled(qc::X, 0, 0, 1U);
   qc.emplace_back(comp.asOperation());
 
   std::cout << qc << "\n";
