+++ test/dd/test_dd_noise_functionality.cpp
@@ -321,8 +321,7 @@ TEST_F(DDNoiseFunctionalityTest, testingUsedQubits) {
   EXPECT_TRUE(compoundOp.getUsedQubits().count(1));
 
   auto classicalControlledOp = qc::ClassicControlledOperation(
-      std::make_unique<qc::StandardOperation>(0, qc::X), std::pair{0, nqubits},
-      1U);
+      std::make_unique<qc::StandardOperation>(0, qc::X), {0, nqubits}, 1U);
   EXPECT_EQ(classicalControlledOp.getUsedQubits().size(), 1);
   EXPECT_TRUE(classicalControlledOp.getUsedQubits().count(0) == 1U);
 }
