### 2025-01-21 912365e
♻️ always reduce IO permutations in idle qubit stripping
Signed-off-by: burgholzer <burgholzer@me.com>
