+++ src/python/ir/register_operations.cpp
@@ -12,14 +12,14 @@
 namespace mqt {
 
 // forward declarations
-void registerOptype(py::module& m);
-void registerControl(py::module& m);
-void registerOperation(py::module& m);
-void registerStandardOperation(py::module& m);
-void registerCompoundOperation(py::module& m);
-void registerNonUnitaryOperation(py::module& m);
-void registerSymbolicOperation(py::module& m);
-void registerClassicControlledOperation(py::module& m);
+void registerOptype(const py::module& m);
+void registerControl(const py::module& m);
+void registerOperation(const py::module& m);
+void registerStandardOperation(const py::module& m);
+void registerCompoundOperation(const py::module& m);
+void registerNonUnitaryOperation(const py::module& m);
+void registerSymbolicOperation(const py::module& m);
+void registerClassicControlledOperation(const py::module& m);
 
 void registerOperations(py::module& m) {
   registerOptype(m);
