### 2025-01-21 dca4aae
🐛 make sure layout indices are consecutive when dumping
Signed-off-by: burgholzer <burgholzer@me.com>
