### 2025-01-21 400e90a
refactor: :recycle: Refactor FromQuantumComputation pattern
