+++ mlir/lib/Dialect/MQTOpt/Transforms/FromQuantumComputationPattern.cpp
@@ -13,8 +13,8 @@
 #include <set>
 
 namespace mqt::ir::opt {
-/// Analysis pattern that filters out all quantum operations from a given
-/// program and creates a quantum computation from them.
+/// Analysis pattern that creates MLIR instructions from a given qc::QuantumComputation.
+/// These instructions replace an existing `AllocOp` that has the `to_replace` attribute set.
 struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
   qc::QuantumComputation& circuit;
 
@@ -26,6 +26,15 @@ struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
     return op->hasAttr("to_replace") ? mlir::success() : mlir::failure();
   }
 
+  /**
+   * @brief Creates an ExtractOp that extracts the qubit at the given index from the given register.
+   *
+   * @param reg The register to extract the qubit from.
+   * @param index The index of the qubit to extract.
+   * @param rewriter The pattern rewriter to use.
+   *
+   * @return The created ExtractOp.
+   */
   ExtractOp createRegisterAccess(AllocOp& reg, size_t index,
                                  mlir::PatternRewriter& rewriter) const {
     return rewriter.create<ExtractOp>(
@@ -33,21 +42,66 @@ struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
         nullptr, rewriter.getI64IntegerAttr(index));
   }
 
-  XOp createXOp(mlir::Location loc,
+  /**
+   * @brief Creates a unitary operation on a given qubit with any number of positive or negative controls.
+   *
+   * @param loc The location of the operation.
+   * @param inQubit The qubit to apply the unitary operation to.
+   * @param controlQubitsPositive The positive control qubits.
+   * @param controlQubitsNegative The negative control qubits.
+   * @param rewriter The pattern rewriter to use.
+   *
+   * @return The created UnitaryOp.
+   */
+  UnitaryInterface createUnitaryOp(mlir::Location loc,
+                          qc::OpType type,
                           mlir::Value inQubit,
-                          mlir::ValueRange controlQubits,
+                          mlir::ValueRange controlQubitsPositive,
+                          mlir::ValueRange controlQubitsNegative,
                           mlir::PatternRewriter& rewriter) const {
-    llvm::SmallVector<mlir::Type, 4> resultTypes(4, inQubit.getType());
-    return rewriter.create<XOp>(loc, resultTypes, inQubit, controlQubits);
+    std::vector<mlir::Type> resultTypes(1 + controlQubitsPositive.size() + controlQubitsNegative.size(), inQubit.getType());
+    mlir::TypeRange outTypes(resultTypes);
+
+    switch(type) {
+      case qc::OpType::X:
+        return rewriter.create<XOp>(
+          loc,
+          outTypes,
+          mlir::ValueRange{},
+          mlir::ValueRange{inQubit},
+          controlQubitsPositive,
+          controlQubitsNegative);
+      default:
+        throw std::runtime_error("Unsupported operation type");
+    }
   }
 
+  /**
+   * @brief Creates a MeasureOp on a given qubit.
+   *
+   * @param loc The location of the operation.
+   * @param targetQubit The qubit to measure.
+   * @param rewriter The pattern rewriter to use.
+   *
+   * @return The created MeasureOp.
+   */
   MeasureOp createMeasureOp(mlir::Location loc,
                           mlir::Value targetQubit,
                           mlir::PatternRewriter& rewriter) const {
     return rewriter.create<MeasureOp>(loc, mlir::TypeRange{QubitType::get(rewriter.getContext()), rewriter.getI1Type()}, targetQubit);
   }
 
[…]
