### 2025-01-21 154e40b
refactor: :recycle: Refactor FromQuantumComputation pattern
