+++ mlir/mlir/include/mlir/Dialect/MQTOpt/Transforms/Passes.h
@@ -16,6 +16,7 @@ namespace mqt::ir::opt {
 #define GEN_PASS_DECL
 #include "mlir/Dialect/MQTOpt/Transforms/Passes.h.inc"
 
+void populateThePassPatterns(mlir::RewritePatternSet& patterns);
 void populateToQuantumComputationPatterns(mlir::RewritePatternSet& patterns,
                                           qc::QuantumComputation& circuit);
 void populateFromQuantumComputationPatterns(mlir::RewritePatternSet& patterns,
