### 2025-01-21 b7a4c2b
refactor: :recycle: Refactor and document ToQuantumComputationPattern
