+++ mlir/mlir/lib/Dialect/MQTOpt/Transforms/ToQuantumComputationPattern.cpp
@@ -29,6 +29,17 @@ struct ToQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
                                                                   : mlir::success();
   }
 
+  /**
+   * @brief Finds the index of a qubit in the list of previously defined qubit variables.
+   *
+   * In particular, this function checks if two value definitions are the same, and, in case
+   * of array-style variables, also checks if the result index is the same.
+   *
+   * @param input The qubit to find.
+   * @param currentQubitVariables The list of previously defined qubit variables.
+   *
+   * @return The index of the qubit in the list of previously defined qubit variables.
+   */
   size_t findQubitIndex(mlir::Value input, std::vector<mlir::Value>& currentQubitVariables) const {
     size_t arrayIndex = 0;
     if (auto opResult = mlir::dyn_cast<mlir::OpResult>(input)) {
@@ -52,6 +63,12 @@ struct ToQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
     throw std::runtime_error("Qubit was not found in list of previously defined qubits");
   }
 
+  /**
+   * @brief Converts a measurement to an operation on the `qc::QuantumComputation` and updates the `currentQubitVariables`.
+   *
+   * @param op The operation to convert.
+   * @param currentQubitVariables The list of previously defined qubit variables.
+   */
   void handleMeasureOp(MeasureOp op, std::vector<mlir::Value>& currentQubitVariables) const {
     const auto ins = op.getInQubits();
     const auto outs = op.getOutQubits();
@@ -69,39 +86,49 @@ struct ToQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
     }
   }
 
-  void handleXOp(XOp op,
+  /**
+   * @brief Converts a unitary operation to an operation on the `qc::QuantumComputation` and updates the `currentQubitVariables`.
+   *
+   * @param op The operation to convert.
+   * @param currentQubitVariables The list of previously defined qubit variables.
+   */
+  void handleUnitaryOp(UnitaryInterface op,
                       std::vector<mlir::Value>& currentQubitVariables) const {
     const auto in = op.getInQubits()[0];
     const auto ctrlIns = op.getCtrlQubits();
     const auto outs = op.getOutQubits();
 
-
-    std::vector<size_t> insIndices(1 + ctrlIns.size());
-
-    const auto found = std::find(currentQubitVariables.begin(),
-                                       currentQubitVariables.end(), in);
-    if (currentQubitVariables.end() == found) {
-      throw std::runtime_error("Qubit not found!");
-    }
-    insIndices[0] = findQubitIndex(in, currentQubitVariables);
-
-
+    // Get the qubit index of every control qubit.
+    std::vector<size_t> ctrlInsIndices(ctrlIns.size());
     std::transform(
-        ctrlIns.begin(), ctrlIns.end(), (insIndices.begin() + 1),
+        ctrlIns.begin(), ctrlIns.end(), ctrlInsIndices.begin(),
         [&currentQubitVariables, this](mlir::Value val) {
           return findQubitIndex(val, currentQubitVariables);
         });
 
-    for (size_t i = 0; i < insIndices.size(); i++) {
-      currentQubitVariables[insIndices[i]] = outs[i];
-    }
-    const auto index0 = insIndices[0];
+    // Get the qubit index of the target qubit.
+    size_t targetIndex = findQubitIndex(in, currentQubitVariables);
 
-    insIndices.erase(insIndices.begin());
-    auto operation = qc::StandardOperation(qc::Controls{insIndices.cbegin(), insIndices.cend()}, index0, qc::OpType::X);
[…]
