### 2025-01-21 c739b54
refactor: :recycle: Change from operation name checks to `llvm::isa`
