### 2025-01-21 7175c66
refactor: :recycle: Change from operation name checks to `llvm::isa`
