+++ mlir/include/mlir/Dialect/MQT/IR/MQTOps.h
@@ -1,61 +1,0 @@
-// Copyright 2022-2023 Xanadu Quantum Technologies Inc.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-
-//     http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#pragma once
-
-#include "mlir/Dialect/Bufferization/IR/AllocationOpInterface.h"
-#include "mlir/Dialect/MQT/IR/MQTInterfaces.h"
-#include "mlir/IR/BuiltinTypes.h"
-#include "mlir/IR/Dialect.h"
-#include "mlir/IR/OpDefinition.h"
-#include "mlir/IR/PatternMatch.h"
-#include "mlir/Interfaces/ControlFlowInterfaces.h"
-#include "mlir/Interfaces/SideEffectInterfaces.h"
-#include "mlir/Support/LogicalResult.h"
-
-#include "llvm/ADT/StringRef.h"
-#include "llvm/Support/ErrorHandling.h"
-
-#include <optional>
-
-//===----------------------------------------------------------------------===//
-// MQT trait declarations.
-//===----------------------------------------------------------------------===//
-
-namespace mlir {
-namespace OpTrait {
-
-template <typename ConcreteType>
-class UnitaryTrait : public TraitBase<ConcreteType, UnitaryTrait> {};
-
-template <typename ConcreteType>
-class HermitianTrait : public TraitBase<ConcreteType, HermitianTrait> {};
-
-} // namespace OpTrait
-} // namespace mlir
-
-class QuantumMemory : public mlir::SideEffects::Resource::Base<QuantumMemory> {
-  llvm::StringRef getName() final { return "QuantumMemory"; }
-};
-
-//===----------------------------------------------------------------------===//
-// MQT ops declarations.
-//===----------------------------------------------------------------------===//
-
-#include "mlir/Dialect/MQT/IR/MQTDialect.h"
-#include "mlir/Dialect/MQT/IR/MQTEnums.h.inc"
-#define GET_ATTRDEF_CLASSES
-#include "mlir/Dialect/MQT/IR/MQTAttributes.h.inc"
-#define GET_OP_CLASSES
-#include "mlir/Dialect/MQT/IR/MQTOps.h.inc"
