+++ mlir/include/mlir/Dialect/CMakeLists.txt
@@ -1,2 +1,1 @@
-add_subdirectory(MQT)
 add_subdirectory(MQTOpt)
