+++ mlir/lib/Dialect/MQT/IR/CMakeLists.txt
@@ -1,11 +1,0 @@
-add_mlir_library(
-  MLIRMQT
-  MQTDialect.cpp
-  MQTInterfaces.cpp
-  MQTOps.cpp
-  ADDITIONAL_HEADER_DIRS
-  ${MQT_MLIR_INCLUDE_BUILD_DIR}/CatalystQuantum
-  DEPENDS
-  MLIRMQTInterfacesIncGen
-  MLIRMQTOpsIncGen
-  MLIRMQTEnumsIncGen)
