+++ mlir/tools/quantum-opt/CMakeLists.txt
@@ -6,8 +6,6 @@ set(LIBS
     ${conversion_libs}
     ${extension_libs}
     MLIROptLib
-    MLIRMQT
-    MLIRMQTTransforms
     MLIRMQTOpt
     MLIRMQTOptTransforms)
 
