### 2025-01-21 a760e24
style: :art: Fix pre-commit issues
