+++ mlir/lib/Dialect/MQTOpt/Transforms/ToQuantumComputationPattern.cpp
@@ -1,6 +1,6 @@
 #include "ir/QuantumComputation.hpp"
-#include "ir/operations/StandardOperation.hpp"
 #include "ir/operations/OpType.hpp"
+#include "ir/operations/StandardOperation.hpp"
 #include "mlir/Dialect/MQTOpt/IR/MQTOptDialect.h"
 #include "mlir/Dialect/MQTOpt/Transforms/Passes.h"
 #include "mlir/IR/BuiltinTypes.h"
@@ -25,60 +25,72 @@ struct ToQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
       : OpRewritePattern(context), circuit(qc) {}
 
   mlir::LogicalResult match(AllocOp op) const override {
-    return (op->hasAttr("to_replace") || op->hasAttr("mqt_core")) ? mlir::failure()
-                                                                  : mlir::success();
+    return (op->hasAttr("to_replace") || op->hasAttr("mqt_core"))
+               ? mlir::failure()
+               : mlir::success();
   }
 
   /**
-   * @brief Finds the index of a qubit in the list of previously defined qubit variables.
+   * @brief Finds the index of a qubit in the list of previously defined qubit
+   * variables.
    *
-   * In particular, this function checks if two value definitions are the same, and, in case
-   * of array-style variables, also checks if the result index is the same.
+   * In particular, this function checks if two value definitions are the same,
+   * and, in case of array-style variables, also checks if the result index is
+   * the same.
    *
    * @param input The qubit to find.
-   * @param currentQubitVariables The list of previously defined qubit variables.
+   * @param currentQubitVariables The list of previously defined qubit
+   * variables.
    *
-   * @return The index of the qubit in the list of previously defined qubit variables.
+   * @return The index of the qubit in the list of previously defined qubit
+   * variables.
    */
-  size_t findQubitIndex(mlir::Value input, std::vector<mlir::Value>& currentQubitVariables) const {
+  size_t findQubitIndex(mlir::Value input,
+                        std::vector<mlir::Value>& currentQubitVariables) const {
     size_t arrayIndex = 0;
     if (auto opResult = mlir::dyn_cast<mlir::OpResult>(input)) {
       arrayIndex = opResult.getResultNumber();
     } else {
-      throw std::runtime_error("Operand is not an operation result. This should never happen!");
+      throw std::runtime_error(
+          "Operand is not an operation result. This should never happen!");
     }
-    for(size_t i = 0; i < currentQubitVariables.size(); i++) {
+    for (size_t i = 0; i < currentQubitVariables.size(); i++) {
       size_t qubitArrayIndex = 0;
-      if (auto opResult = mlir::dyn_cast<mlir::OpResult>(currentQubitVariables[i])) {
+      if (auto opResult =
+              mlir::dyn_cast<mlir::OpResult>(currentQubitVariables[i])) {
         qubitArrayIndex = opResult.getResultNumber();
       } else {
-        throw std::runtime_error("Qubit is not an operation result. This should never happen!");
+        throw std::runtime_error(
+            "Qubit is not an operation result. This should never happen!");
       }
 
-      if(currentQubitVariables[i] == input && arrayIndex == qubitArrayIndex) {
+      if (currentQubitVariables[i] == input && arrayIndex == qubitArrayIndex) {
         return i;
       }
     }
 
-    throw std::runtime_error("Qubit was not found in list of previously defined qubits");
+    throw std::runtime_error(
+        "Qubit was not found in list of previously defined qubits");
   }
 
   /**
-   * @brief Converts a measurement to an operation on the `qc::QuantumComputation` and updates the `currentQubitVariables`.
[…]
