+++ mlir/lib/Dialect/MQTOpt/Transforms/FromQuantumComputationPattern.cpp
@@ -13,8 +13,9 @@
 #include <set>
 
 namespace mqt::ir::opt {
-/// Analysis pattern that creates MLIR instructions from a given qc::QuantumComputation.
-/// These instructions replace an existing `AllocOp` that has the `to_replace` attribute set.
+/// Analysis pattern that creates MLIR instructions from a given
+/// qc::QuantumComputation. These instructions replace an existing `AllocOp`
+/// that has the `to_replace` attribute set.
 struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
   qc::QuantumComputation& circuit;
 
@@ -27,7 +28,8 @@ struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
   }
 
   /**
-   * @brief Creates an ExtractOp that extracts the qubit at the given index from the given register.
+   * @brief Creates an ExtractOp that extracts the qubit at the given index from
+   * the given register.
    *
    * @param reg The register to extract the qubit from.
    * @param index The index of the qubit to extract.
@@ -38,12 +40,15 @@ struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
   ExtractOp createRegisterAccess(AllocOp& reg, size_t index,
                                  mlir::PatternRewriter& rewriter) const {
     return rewriter.create<ExtractOp>(
-        reg.getLoc(), mlir::TypeRange{QubitRegisterType::get(rewriter.getContext()), QubitType::get(rewriter.getContext())}, reg.getResult(),
-        nullptr, rewriter.getI64IntegerAttr(index));
+        reg.getLoc(),
+        mlir::TypeRange{QubitRegisterType::get(rewriter.getContext()),
+                        QubitType::get(rewriter.getContext())},
+        reg.getResult(), nullptr, rewriter.getI64IntegerAttr(index));
   }
 
   /**
-   * @brief Creates a unitary operation on a given qubit with any number of positive or negative controls.
+   * @brief Creates a unitary operation on a given qubit with any number of
+   * positive or negative controls.
    *
    * @param loc The location of the operation.
    * @param inQubit The qubit to apply the unitary operation to.
@@ -53,26 +58,23 @@ struct FromQuantumComputationPattern final : mlir::OpRewritePattern<AllocOp> {
    *
    * @return The created UnitaryOp.
    */
-  UnitaryInterface createUnitaryOp(mlir::Location loc,
-                          qc::OpType type,
-                          mlir::Value inQubit,
-                          mlir::ValueRange controlQubitsPositive,
-                          mlir::ValueRange controlQubitsNegative,
-                          mlir::PatternRewriter& rewriter) const {
-    std::vector<mlir::Type> resultTypes(1 + controlQubitsPositive.size() + controlQubitsNegative.size(), inQubit.getType());
+  UnitaryInterface createUnitaryOp(mlir::Location loc, qc::OpType type,
+                                   mlir::Value inQubit,
+                                   mlir::ValueRange controlQubitsPositive,
+                                   mlir::ValueRange controlQubitsNegative,
+                                   mlir::PatternRewriter& rewriter) const {
+    std::vector<mlir::Type> resultTypes(1 + controlQubitsPositive.size() +
+                                            controlQubitsNegative.size(),
+                                        inQubit.getType());
     mlir::TypeRange outTypes(resultTypes);
 
-    switch(type) {
-      case qc::OpType::X:
-        return rewriter.create<XOp>(
-          loc,
-          outTypes,
-          mlir::ValueRange{},
-          mlir::ValueRange{inQubit},
-          controlQubitsPositive,
-          controlQubitsNegative);
-      default:
-        throw std::runtime_error("Unsupported operation type");
[…]
