+++ mlir/mlir/lib/Dialect/MQTOpt/Transforms/MQTCoreRoundTrip.cpp
@@ -16,10 +16,10 @@ namespace mqt::ir::opt {
 
 struct MQTCoreRoundTrip final : impl::MQTCoreRoundTripBase<MQTCoreRoundTrip> {
 
-  MQTCoreRoundTrip() : circuit(0, 100) {}
+  MQTCoreRoundTrip()
+      : circuit(0, 100) {} // qc has fixed number of 100 classical bits.
 
-  qc::QuantumComputation
-      circuit; // TODO qc has fixed number of 100 classical bits.
+  qc::QuantumComputation circuit;
 
   void runOnOperation() override {
     // Get the current operation being operated on.
