### 2025-01-21 0564cf6
♻️ reorganize Python bindings
Signed-off-by: burgholzer <burgholzer@me.com>
