### 2025-01-21 cd45ac3
📝 adjust docs example
Signed-off-by: burgholzer <burgholzer@me.com>
