### 2025-01-21 2a4a41d
🚨 address clang-tidy warnings
Signed-off-by: burgholzer <burgholzer@me.com>
