### 2025-01-21 ef7b2ec
✅ add a couple more unittests
Signed-off-by: burgholzer <burgholzer@me.com>
