### 2025-01-21 89e2271
🐛 do not invalidate iterator in block collection
Signed-off-by: Lukas Burgholzer <burgholzer@me.com>
