### 2025-01-21 3ebd48a
🐛 use fuzzy comparison for floating point in const evaluation equality check
Signed-off-by: burgholzer <burgholzer@me.com>
