### 2025-01-21 cefedf2
🐛 do not invalidate iterator in block collection (#808)
## Description

This fixes a small bug observed as part of #803, which only surfaced on
Windows in Debug mode when compiling using Clang 🤯
Part of the block collection code in the circuit optimizer would
invalidate an iterator due to an update of the respective container
being iterated over within the loop.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropri
[…]
