### 2025-01-21 5eaebd7
✨👷️ extend extensive CI to Python tests
Signed-off-by: burgholzer <burgholzer@me.com>
