### 2025-01-21 f61e7ff
♻️ run Python linter separate from Python tests
this avoids running it twice when running extensive Python CI

Signed-off-by: burgholzer <burgholzer@me.com>
