+++ .github/workflows/ci.yml
@@ -161,6 +161,12 @@ jobs:
       enable-macos14: false
       enable-windows2022: false
 
+  python-linter:
+    name: 🐍 Lint
+    needs: change-detection
+    if: fromJSON(needs.change-detection.outputs.run-python-tests)
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-linter.yml@v1.5
+
   code-ql:
     name: 📝 CodeQL
     needs: change-detection
@@ -186,6 +192,7 @@ jobs:
       - cpp-linter
       - python-tests
       - python-tests-extensive
+      - python-linter
       - code-ql
       - cd
     runs-on: ubuntu-latest
