### 2025-01-22 2e1db67
👷⬆️ update MQT workflows to v1.6
Signed-off-by: burgholzer <burgholzer@me.com>
