### 2025-01-22 2ca6df8
👷🔥 do not upload to TestPyPI
Turns out these uploads hardly provided any value and just consumed CI runtime. At least we now know that this is possible and how to achieve it, if we ever need it again.

Signed-off-by: burgholzer <burgholzer@me.com>
