### 2025-01-22 df5befa
🩹 check also needs to depend on the Python linter job
Signed-off-by: burgholzer <burgholzer@me.com>
