### 2025-01-22 5668f71
🩹 check also needs to depend on the C++ coverage job
Signed-off-by: burgholzer <burgholzer@me.com>
