### 2025-01-22 5959931
👷⬆️ update MQT workflows to v1.6 (#803)
## Description

This PR adopts the latest version of the MQT workflows, which contains
the changes from https://github.com/cda-tum/mqt-workflows/pull/60.
By default, C++ CI is now only run in Release mode for all major
platforms. However, when a PR is tagged with `extensive-cpp-ci`, a large
set of test runs will be started on various operating system versions
using various compilers.
Similarly, adding a `extensive-python-ci` label will trigger additional
Python test runs on different OS versions.
Note that the labels need to be present _before_ CI starts
[…]
