### 2025-01-22 472f6b9
🧪 add failing test
Signed-off-by: burgholzer <burgholzer@me.com>
