### 2025-01-22 ca1e4c4
🐛 fix permutation handling in QASM dump (#810)
## Description

This PR fixes an oversight in #753 where the respective changes would
sometimes create invalid output permutation dumps.
This surfaced as part of https://github.com/cda-tum/mqt-qmap/pull/418.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request in
[…]
