### 2025-01-22 4176e32
👷⬆️ update MQT workflows to v1.7 (#812)
## Description

This small PR updates the MQT workflows to version 1.7, which includes
some fixes for CD jobs that were overlooked in v1.6

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed
[…]
