+++ .github/workflows/cd.yml
@@ -11,16 +11,11 @@ permissions:
 
 jobs:
   # Builds the sdist and wheels on all supported platforms and uploads the resulting
-  # wheels as GitHub artifacts `dev-cibw-*`, `test-cibw-*`, or `cibw-*`, depending on
-  # whether the workflow is triggered from a PR, a push to main, or a release, respectively.
+  # wheels as GitHub artifacts `dev-cibw-*` or `cibw-*`, depending on whether the
+  # workflow is triggered from a PR or a release, respectively.
   python-packaging:
     name: 🐍 Packaging
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-packaging.yml@v1.6
-    with:
-      # Do not include local version information on pushes to main to facilitate TestPyPI uploads.
-      no-local-version: ${{ github.ref == 'refs/heads/main' && github.event_name == 'push' }}
-      # Do not build emulated wheels for mqt-core to avoid issues with PyPI space limitations.
-      build-emulated-wheels: false
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-packaging.yml@v1.7
 
   # Downloads the previously generated artifacts and deploys to PyPI on published releases.
   deploy:
