+++ .github/workflows/ci.yml
@@ -14,13 +14,13 @@ concurrency:
 jobs:
   change-detection:
     name: 🔍 Change
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-change-detection.yml@v1.6
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-change-detection.yml@v1.7
 
   cpp-tests:
     name: 🇨‌ Test
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cpp-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@v1.6
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@v1.7
     with:
       # Runs to enable
       enable-ubuntu2404-gcc-release: true
@@ -68,7 +68,7 @@ jobs:
     name: 🇨‌ Test (Extensive)
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cpp-tests) && github.event_name == 'pull_request' && contains(github.event.pull_request.labels.*.name, 'extensive-cpp-ci')
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@v1.6
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-ci.yml@v1.7
     with:
       # Runs to enable
       enable-ubuntu2404-gcc-debug: true
@@ -115,13 +115,13 @@ jobs:
     name: 🇨‌ Coverage
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cpp-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-coverage.yml@v1.6
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-coverage.yml@v1.7
 
   cpp-linter:
     name: 🇨‌ Lint
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-cpp-linter)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-linter.yml@v1.6
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-cpp-linter.yml@v1.7
     with:
       cmake-args: -DBUILD_MQT_CORE_BENCHMARKS=ON
 
@@ -129,7 +129,7 @@ jobs:
     name: 🐍 Test
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-python-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.6
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.7
     with:
       # Runs to enable
       enable-ubuntu2404: true
@@ -148,7 +148,7 @@ jobs:
     name: 🐍 Test (Extensive)
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-python-tests) && github.event_name == 'pull_request' && contains(github.event.pull_request.labels.*.name, 'extensive-python-ci')
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.6
+    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-ci.yml@v1.7
     with:
       # Runs to enable
       enable-ubuntu2204: true
@@ -166,22 +166,19 @@ jobs:
     name: 🐍 Lint
     needs: change-detection
     if: fromJSON(needs.change-detection.outputs.run-python-tests)
-    uses: cda-tum/mqt-workflows/.github/workflows/reusable-python-linter.yml@v1.6
[…]
