### 2025-01-23 e0ea685
✨ Added Bridge Gate as Standard Operation
