+++ mlir/mlir/include/mlir/Dialect/MQTOpt/Transforms/Passes.h
@@ -3,7 +3,6 @@
 #include "ir/QuantumComputation.hpp"
 
 #include <mlir/Pass/Pass.h>
-#include <set>
 
 namespace mlir {
 
