### 2025-01-23 f8bf8f9
Potential fix for linter error in real parser. Added back required include in tests of real parser
