### 2025-01-23 36e925f
refactor: :building_construction: Refactor quantum-opt CMakeLists.txt, removing an unnecessarily listed dependency
