### 2025-01-23 37d87ae
refactor: :recycle: remove the arbitrary 100 classical bits for a dynamic classical bit register in the round-trip pass
