### 2025-01-24 cda-tum/mqt-core#813 @burgholzer
📝 Add documentation for ZX package
The ZX package currently has no real documentation. Neither in it's public API nor in the actual documentation hosted on RtD.
This is a tracking issue for adding such documentation.
At the bare minimum, this should document all public methods in C++ and add a dedicated documentation section explaining the main features and limitations.
