### 2025-01-24 cda-tum/mqt-core#814 @burgholzer
📝 Add documentation for DD package
The DD package currently has no real documentation besides the documentation of the evaluation functionality.
This is a tracking issue for adding such documentation.

At the bare minimum, this should document the main C++ used to interact with the package and add a dedicated documentation section explaining the main features and limitations.
To achieve this, it might be worth considering exposing the DD package to Python and documenting the corresponding Python stubs (similar to how this is done for the IR submodule of MQT Core).
