### 2025-01-24 d29b4cd
👥 add Tom (@pehamTom) as author to JOSS paper
Signed-off-by: burgholzer <burgholzer@me.com>
