### 2025-01-24 #7478 @burgholzer
Hey everyone 👋🏼

I have created tracking issues for the remaining open points on adding documentation for two of the subpackages of MQT Core here:
- https://github.com/cda-tum/mqt-core/issues/813
- https://github.com/cda-tum/mqt-core/issues/814

Tom (@pehamTom), the original author of the ZX-package in MQT Core, has volunteered to write the corresponding documentation. As such, I believe it is now more than justified to include Tom in the list of authors for this submission. I have updated the metadata of the paper submission in https://github.com/cda-tum/mqt-core/commit/3b09c84644d7f7edcf58e7e76b766134ede9d2b6 and rebased the paper branch on the latest main.
