### 2025-01-24 aa4d01d
:recycle: Fix descriptions of Standard Operations in MQTOpt dialect (#9)
This PR addresses the remaining changes from the previous PR #5.

---------

Signed-off-by: Lukas Burgholzer <burgholzer@me.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>
