### 2025-01-26 d9c8c22
:recycle: rename opToSingleGateMatrix to opToSingleQubitGateMatrix
