### 2025-01-26 76468f5
:white_check_mark: add tests to check for exceptions on invalid input when retrieving DDs from standardoperations
