### 2025-01-27 #7478 @josh146
@danielskatz I have gone over the changes and have updated my checklist.

> I added a related work section to the JOSS paper in cda-tum/mqt-core@262fd6d ✅ Hopefully this addresses the points raised by all three reviewers. Please let me know if you feel that the section is missing something.

The new related work section looks great, and satisfies the State of the field requirement from the reviewing checklist :+1:

> I rephrased that part in the documentation as part of cda-tum/mqt-core#789. ✅ I hope this is clearer now.

This is much clearer now!

Thanks @burgholzer, once https://github.com/cda-tum/mqt-core/issues/813 https://github.com/cda-tum/mqt-core/issues/814 are available this will complete my review!
