### 2025-01-27 84a32fa
first few signature changes, docstring adapted
