+++ include/mqt-core/algorithms/StatePreparation.hpp
@@ -16,22 +16,19 @@
 
 namespace qc {
 /**
- * Prepares a generic Quantum State from a list of normalized complex
- amplitudes
- * Adapted implementation of Qiskit State Preparation:
+ * @brief Prepares a generic quantum state from a list of normalized
+ *        complex amplitudes
  *
- https://github.com/Qiskit/qiskit/blob/main/qiskit/circuit/library/data_preparation/state_preparation.py#
- * based on paper:
- *      Shende, Bullock, Markov. Synthesis of Quantum Logic Circuits (2004)
-        [`https://ieeexplore.ieee.org/document/1629135`]
- * */
-
-/**
- * @throws invalid_argument when amplitudes are not normalized or length not
- * power of 2
- * @param list of complex amplitudes to initialize to
- * @return MQT Circuit that initializes a state
- * */
+ * Adapted implementation of IBM Qiskit's State Preparation:
+ * https://github.com/Qiskit/qiskit/blob/e9ccd3f374fd5424214361d47febacfa5919e1e3/qiskit/circuit/library/data_preparation/state_preparation.py
+ * based on the following paper:
+ *      V. V. Shende, S. S. Bullock and I. L. Markov, "Synthesis of quantum-logic circuits," in IEEE Transactions on Computer-Aided Design of Integrated Circuits and Systems, vol. 25, no. 6, pp. 1000-1010, June 2006, doi: 10.1109/TCAD.2005.855930.
+ *
+ * @param amplitudes state (vector) to prepare. Must be normalized and have a size that is a power of two
+ * @return quantum computation that prepares the state
+ * @throws invalid_argument @p amplitudes is not normalized or its length is not a
+ * power of two
+ **/
 [[nodiscard]] auto
 createStatePreparationCircuit(std::vector<std::complex<double>>& amplitudes)
     -> QuantumComputation;
