### 2025-01-27 f5f4104
Added docstrings to public methods of the ZXDiagram class.
