+++ include/mqt-core/zx/ZXDiagram.hpp
@@ -19,94 +19,292 @@
 #include <vector>
 
 namespace zx {
+
+/**
+ * @brief Class representing a ZX-diagram.
+ *
+ * @details Labelled undirected graph. Every node is either of type
+ * VertexType::Z, VertexType::X or VertexType::Boundary. The boundary vertices
+ * are the inputs and outputs of the diagram. The graph is stored as an
+ * adjacency list. The vertices are stored in a vector of optional VertexData.
+ * The optional is empty if the vertex has been deleted.
+ */
+
 class ZXDiagram {
 public:
   ZXDiagram() = default;
-  // create n_qubit identity_diagram
+  // Create an n_qubit identity diagram.
   explicit ZXDiagram(std::size_t nqubits);
 
+  /**
+   * @brief Add an edge to the diagram.
+   * @param from first vertex of the edge
+   * @param to second vertex of the edge
+   * @param type type of the edge (Simple or Hadamard)
+   */
   void addEdge(Vertex from, Vertex to, EdgeType type = EdgeType::Simple);
+
+  /**
+   * @brief Add a Hadamard edge to the diagram.
+   * @param from first vertex of the edge
+   * @param to second vertex of the edge
+   */
   void addHadamardEdge(const Vertex from, const Vertex to) {
     addEdge(from, to, EdgeType::Hadamard);
   };
+
+  /**
+   * @brief Add edge to diagram, simplifying the diagram if this results in
+   * parallel edges.
+   * @details In the ZX-calculus, parallel edges and self-loops can usually be
+   * simplified. Two parallel Simple edges between two opposite-colored
+   * vertices cancel out for exampl. This function adds an edge to the diagram
+   * and then simplifies the diagram if necessary.
+   * @param from first vertex of the edge
+   * @param to second vertex of the edge
+   * @param eType type of the edge (Simple or Hadamard)
+   */
   void addEdgeParallelAware(Vertex from, Vertex to,
                             EdgeType eType = EdgeType::Simple);
+
+  /**
+   * @brief Remove Edge from the diagram.
+   * @param from first vertex of the edge
+   * @param to second vertex of the edge
+   */
   void removeEdge(Vertex from, Vertex to);
 
+  /**
+   * @brief Add a vertex to the diagram.
+   * @param data data of the vertex
+   * @return the vertex that was added
+   */
   Vertex addVertex(const VertexData& data);
+
+  /**
+   * @brief Add a vertex to the diagram.
+   * @param qubit qubit index of the vertex
+   * @param col column index of the vertex
+   * @param phase phase of the vertex
+   * @param type type of the vertex
+   * @return the vertex that was added
+   */
   Vertex addVertex(Qubit qubit, Col col = 0,
                    const PiExpression& phase = PiExpression(),
                    VertexType type = VertexType::Z);
+
+  /**
+   * @brief Add a bare qubit to the diagram.
+   */
   void addQubit();
+
+  /**
+   * @brief Add multiple bare qubits to the diagram.
+   * @param n number of qubits to add
+   */
   void addQubits(zx::Qubit n);
+
+  /**
+   * @brief Remove a vertex from the diagram. Also removes all edges incident to
+   * the vertex.
+   * @param toRemove vertex to remove
+   */
   void removeVertex(Vertex toRemove);
 
+  /**
+   * @brief Get number of deleted vertices.
+   * @details Deleted vertices are not removed from the vertices vector, but
+   * are left as optional in the array.
+   * @return number of deleted vertices
+   */
   [[nodiscard]] std::size_t getNdeleted() const { return deleted.size(); }
+
+  /**
+   * @brief Get number of vertices.
[…]
