### 2025-01-27 4265f82
Added docstrings for Utils and Definitions.
