### 2025-01-27 2fce4f7
Document `FunctionalityConstruction`.
