### 2025-01-27 #7478 @danielskatz
Thanks @josh146
