### 2025-01-27 633861c
Added docs for Simplifications.
