+++ include/mqt-core/zx/Simplify.hpp
@@ -17,6 +17,17 @@
 
 namespace zx {
 
+/**
+ * @brief Simplify the diagram by applying the given rule to all vertices that
+ * match the given check.
+      * @tparam VertexCheckFun Type of the check function
+      * @tparam VertexRuleFun Type of the rule function
+      * @param diag The diagram to simplify
+      * @param check The check function that determines if a vertex should be
+      simplified
+      * @param rule The rule function that is applied to the vertex
+      * @return The number of simplifications that were applied
+      */
 template <class VertexCheckFun, class VertexRuleFun>
 std::size_t simplifyVertices(ZXDiagram& diag, VertexCheckFun check,
                              VertexRuleFun rule) {
@@ -37,6 +48,17 @@ std::size_t simplifyVertices(ZXDiagram& diag, VertexCheckFun check,
   return nSimplifications;
 }
 
+/**
+ * @brief Simplify the diagram by applying the given rule to all edges that
+ * match the given check.
+      * @tparam EdgeCheckFun Type of the check function
+      * @tparam EdgeRuleFun Type of the rule function
+      * @param diag The diagram to simplify
+      * @param check The check function that determines if an edge should be
+      simplified
+      * @param rule The rule function that is applied to the edge
+      * @return The number of simplifications that were applied
+      */
 template <class EdgeCheckFun, class EdgeRuleFun>
 std::size_t simplifyEdges(ZXDiagram& diag, EdgeCheckFun check,
                           EdgeRuleFun rule) {
@@ -60,23 +82,101 @@ std::size_t simplifyEdges(ZXDiagram& diag, EdgeCheckFun check,
 
 std::size_t gadgetSimp(ZXDiagram& diag);
 
+/**
+ * @brief Apply the identity rule to the Diagram until exhaustion.
+ * @details A spider with exactly two incident edges and a phase of 0 is
+ * removed. See https://arxiv.org/abs/2012.13966 page 27, rule (id).
+ * @param diag The diagram to simplify
+ * @return The number of simplifications that were applied
+ */
+
 std::size_t idSimp(ZXDiagram& diag);
 
+/**
+ * @brief Apply the spider rule to the Diagram until exhaustion.
+ * @details Spiders of the same type connected by a regular edge can be turned
+into one
+ * by adding both phases and fusing all incident edges.
+ * See https://arxiv.org/abs/2012.13966 page 27, rule (f).
+ * @param diag The diagram to simplify
+ * @return The number of simplifications that were applied
+ */
 std::size_t spiderSimp(ZXDiagram& diag);
 
+/**
+ * @brief Apply the local complementation rule to the Diagram until exhaustion.
+ * @details See https://arxiv.org/abs/2012.13966 equation (102).
+ * @param diag The diagram to simplify
+ * @return The number of simplifications that were applied
+ */
 std::size_t localCompSimp(ZXDiagram& diag);
 
-std::size_t pivotPauliSimp(ZXDiagram& diag);
-
+/**
+ * @brief Simplify the diagram by applying the pivot rule.
+ * @details See https://arxiv.org/abs/2012.13966, equation 103 for details.
+ * Similar to the Pauli pivot rule but can be applied to vertices connectedto
+ * the boundary.
+ * @param diag The diagram to simplify.
+ * @return The number of simplifications applied.
+ */
 std::size_t pivotSimp(ZXDiagram& diag);
 
+/**
+ * Simplify the diagram by applying the pivot rule in the case that the pivot
+ * spiders have phases of Pi. See https://arxiv.org/abs/2012.13966, equation 103
+ * for details.
+ * @param diag The diagram to simplify.
+ * @return The number of simplifications applied.
+ */
[…]
