### 2025-01-27 #7478 @danielskatz
@edyounis - thanks very much.

(As two small points, editorialbot commands need to be the first thing in a new comment, and if you try this, you will fund that only editors can issue this command)
