### 2025-01-30 #7478 @josh146
@danielskatz I've checked off my list! For this one,

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

it will be satisfied by the following issues being complete:

https://github.com/cda-tum/mqt-core/issues/813
https://github.com/cda-tum/mqt-core/issues/814
