### 2025-01-30 #7478 @danielskatz
@josh146 - normally, the checklist being complete means that the publication can go forward. So I wonder if you mean that you think these 2 issues should be closed before publication in this case, which normally would be handled by not checking off the last issue until then?
