### 2025-02-11 #7478 @josh146
Hey @danielskatz, sorry for the delayed response. To clarify, I think I would like to see the resulting documentation as per those two issues before recommending publication, as those two issues are designed to address documentation shortfalls as pointed out in my original review, as well as to include Tom Peham as an author on the submission.

@burgholzer do you have an update on when those two issues would have associated PRs (I see one is already open)?
