### 2025-02-11 #7478 @burgholzer
> @burgholzer do you have an update on when those two issues would have associated PRs (I see one is already open)?

One is already open and being actively worked on. The other one is still on me. Given how I am currently on vacation and how next week is already quite "full", I'd say a realistic estimate is that both PRs will be done by end of the month.
