### 2025-02-17 #7478 @burgholzer
> 👋 [@burgholzer](https://github.com/burgholzer) - are you ready to merge the 2 PRs? I think that once you do this, [@josh146](https://github.com/josh146) can complete his review, and then we can move this forward to the acceptance and publication steps, given that the other two reviews are complete.

Almost. I just created https://github.com/cda-tum/mqt-core/pull/831, which finally adds the API documentation to the DD package.
https://github.com/cda-tum/mqt-core/pull/817 adds the API documentation for the ZX-calculus package.
Both of these PRs still miss a quickstart guide for the respective datastructures, which we are about to create.
Other than that, there is only one more configuration error we are trying to track down regarding the C++ API docs build and the typical final cosmetics for the paper.
