### 2025-02-25 #7478 @burgholzer
> 👋 [@burgholzer](https://github.com/burgholzer) - do you have any update on this ☝ ?

I do. And, for once, a positive one 😌
The documentation for the DD Package library has been added in https://github.com/cda-tum/mqt-core/pull/831 (which I merged earlier today).
The documentation for the ZX Calculus library is in its final stretches over in https://github.com/cda-tum/mqt-core/pull/817.

If anyone (@josh146 😉) wants to check the current documentation build, that is available here: https://mqt--817.org.readthedocs.build/projects/core/en/817/
I'd expect this to be merged either today or tomorrow.

I also put some finishing touches on the paper itself. I'll invoke the editorialbot in a subsequent comment, so the latest version of the paper is available here as well.
