### 2025-02-26 9d8cfd5
📝 Add documentation for ZX-calculus package and C++ API docs setup (#817)
## Description

This PR introduces documentation for the C++ ZX-calculus library. 
It also adds the general infrastructure for C++ API documentation as
part of the MQT Core project based on Doxygen and the `breathe`
extension.

Fixes #813

This change addresses concerns about missing documentation that came up
in the revision openjournals/joss-reviews#7478.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only
[…]
