+++ noxfile.py
@@ -13,6 +13,7 @@ import argparse
 import os
 import shutil
 import sys
+from pathlib import Path
 from typing import TYPE_CHECKING
 
 import nox
@@ -111,6 +112,26 @@ def docs(session: nox.Session) -> None:
     serve = args.builder == "html" and session.interactive
     if serve:
         session.install("sphinx-autobuild")
+
+    # build the C++ API docs using doxygen
+    with session.chdir("docs"):
+        if shutil.which("doxygen") is None:
+            session.error("doxygen is required to build the C++ API docs")
+
+        Path("_build/doxygen").mkdir(parents=True, exist_ok=True)
+        session.run("doxygen", "Doxyfile", external=True)
+        Path("api/cpp").mkdir(parents=True, exist_ok=True)
+        session.run(
+            "breathe-apidoc",
+            "-o",
+            "api/cpp",
+            "-m",
+            "-f",
+            "-g",
+            "namespace",
+            "_build/doxygen/xml/",
+            external=True,
+        )
 
     env = {"UV_PROJECT_ENVIRONMENT": session.virtualenv.location}
     # install build and docs dependencies on top of the existing environment
