### 2025-02-26 #7478 @burgholzer
Alright. The deed is done. All open PRs are merged. The new extensive documentation is live at: https://mqt.readthedocs.io/projects/core 🙌🏼

With that, I believe all the open feedback is addressed. 🥳
