### 2025-02-26 #7478 @josh146
@burgholzer I just had a look, the new documentation fully resolves the remaining checklist item I had, so my checklist is complete and fully approved on my end!

In particular, I wanted to add that the decision diagram documentation is excellent, and I learnt a lot just by reading it.
