### 2025-03-15 #7478 @danielskatz
👋 @burgholzer - The paper generally looks good, except for some issues in the bib related to cases. Please merge https://github.com/cda-tum/mqt-core/pull/867 to fix these.

After this, could you:
