### 2025-03-24 #7478 @burgholzer
> 👋 [@burgholzer](https://github.com/burgholzer) - note that this is now waiting for you to take the last steps before acceptance and publication

Thanks for the ping here! I am on it. There are just a couple last changes I want to get in before tagging the next (major) version. Should only take a couple more days.
