### 2025-04-07 #7478 @burgholzer
> 👋 [@burgholzer](https://github.com/burgholzer) - Any progress to report?

Yes. Finally. Sorry for the long delays here.

- The GitHub release (`v3.0.0`) is here: https://github.com/munich-quantum-toolkit/core/releases/tag/v3.0.0
- The DOI from figshare is: https://doi.org/10.6084/m9.figshare.28740254.v1

Note that a couple of metadata have changed compared to the original submission (as listed in the issue description):
- The repository has moved from https://github.com/cda-tum/mqt-core to https://github.com/munich-quantum-toolkit/core
- The version is `v3.0.0`
- The JOSS paper has been merged into the `main` branch where it resides in the `paper` directory
- The respective `codemeta.json` has been updated to reflect the above.

Let me know if you need anything else at this point in time 😌
