### 2025-04-07 #7478 @burgholzer
> Congratulations to [@burgholzer](https://github.com/burgholzer) (Lukas Burgholzer) and co-authors on your publication!!
>
> And thanks to [@1ucian0](https://github.com/1ucian0), [@edyounis](https://github.com/edyounis), and [@josh146](https://github.com/josh146) for reviewing! JOSS depends on volunteers and we couldn't do this without you

Thanks to everyone involved 🙏🏼
I just added the JOSS reference and citation information to the repository and the new patch release, including everything, is live at https://github.com/munich-quantum-toolkit/core 🎉
